remove map tool variables countryCodes and allMaps to a new javascript file named map.js
