+++ map.js
@@ -1,0 +1,410 @@
+
+/***  Dinosaur Maps Tool data variables ***/
+// on hover, the event.target.id is a country code
+const countryCodes = {
+  AQ: 'Antarctica',
+  AR: 'Argentina',
+  AT: 'Austria',
+  AU: 'Australia',
+  BE: 'Belgium',
+  BR: 'Brazil',
+  CA: 'Canada',
+  CH: 'Switzerland',
+  CL: 'Chile',
+  CN: 'China',
+  DE: 'Germany',
+  DZ: 'Algeria',
+  EG: 'Egypt',
+  ES: 'Spain',
+  FR: 'France',
+  GB: 'United Kingdom',
+  IN: 'India',
+  JP: 'Japan',
+  KZ: 'Kazakhstan',
+  LS: 'Lesotho',
+  MA: 'Morocco',
+  MG: 'Madagascar',
+  MN: 'Mongolia',
+  MW: 'Malawi',
+  NE: 'Niger',
+  PT: 'Portugal',
+  RO: 'Romania',
+  RU: 'Russia',
+  TN: 'Tunisia',
+  TZ: 'Tanzania',
+  US: 'USA',
+  UY: 'Uruguay',
+  UZ: 'Uzbekistan',
+  ZA: 'South Africa',
+  ZW: 'Zimbabwe'
+}
+
+// object holding the svg text for each continent
+const allMaps = {
+  africa: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="continent" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 193.55 216.53">
+    <defs>
+      <style>
+        .cls-1 {
+          fill: #ccc;
+          stroke: #fff;
+          stroke-width: .5px;
+        }
+        .dinosaurs {
+          fill: #2a7035;
+          cursor: pointer;
+        }
+        .dinosaurs:hover {
+          fill: #194520;
+        }
+      </style>
+    </defs>
+    <g id="continent" data-name="Layer 1">
+      <g>
+        <path id="AO" class="cls-1" d="M95.7,129.04l.69,2.09.8,1.68.64.91,1.07,1.47,1.85-.23.93-.4,1.55.4.42-.7.7-1.64,1.74-.11.15-.49h1.43s-.24,1-.24,1l3.4-.02.05,1.77.57,1.09-.41,1.7.21,1.74.94,1.05-.15,3.37.69-.26,1.22.07,1.74-.42,1.28.17.3.88-.32,1.38.49,1.34-.42,1.07.24.99-5.84-.04-.13,9.16,1.89,2.38,1.83,1.82-5.15,1.19-6.79-.41-1.94-1.4-11.37.13-.42.21-1.67-1.32-1.82-.09-1.68.5-1.35.56-.26-1.83.39-2.55.97-2.65.15-1.24.91-2.59.67-1.17,1.61-1.87.9-1.27.29-2.11-.15-1.61-.84-1.01-.75-1.72-.69-1.69.15-.59.86-1.12-.85-2.72-.57-1.88-1.4-1.77.27-.54,1.16-.38.81.05.98-.34h8.27ZM84.79,128.5l-.71.3-.75-2.1,1.13-1.21.85-.47,1.05.96-1.02.59-.46.72-.09,1.21Z"/>
+        <path id="BF" class="cls-1" d="M42,85.66l-1.92-.73-1.32.11-.98.71-1.26-.6-.49-.93-1.26-.62-.19-1.64.77-1.21-.07-.96,2.23-2.36.41-1.96.77-.7,1.36.38,1.17-.58.38-.74,2.18-1.28.53-.9,2.62-1.2,1.55-.41.7.55h1.79s-.22,1.39-.22,1.39l.38,1.31,1.58,1.87.08,1.38,3.24.65-.07,1.95-.61.86-1.37.26-.57,1.24-.96.32-2.46-.06-1.3-.22-.9.46-1.24-.21-4.87.13-.07,1.61.39,2.13Z"/>
+        <path id="BI" class="cls-1" d="M132.19,125.19l-.18-3.37-.71-1.26,1.71.22.86-1.59,1.49.18.16,1.1.6.63.03.91-.69.58-1.1,1.46-1.01,1.01-1.16.13Z"/>
+        <path id="BJ" class="cls-1" d="M57.47,95.18l-2.32.33-.69-1.94.13-6.46-.57-.58-.1-1.39-.98-.99-.85-.83.36-1.5.96-.32.57-1.24,1.37-.26.61-.86.94-.83h1.01s2.14,1.63,2.14,1.63l-.11.95.63,1.68-.55,1.14.29.76-1.36,1.75-.86.87-.53,1.77.07,1.79-.16,4.53Z"/>
+        <path id="BW" class="cls-1" d="M121.84,165.21l.56.52.89,1.71,3.17,3.25,1.2.32v1.05s.83,1.9.83,1.9l2.17.46,1.79,1.36-3.97,2.22-2.52,2.26-.93,2.03-.84,1.15-1.53.25-.49,1.47-.29.96-1.79.72-2.28-.15-1.34-.86-1.18-.38-1.37.72-.69,1.48-1.33.93-1.4,1.39-2.01.32-.62-1.09.26-1.9-1.67-2.93-.75-.46v-8.86l2.76-.11.08-10.57,2.09-.09,4.32-1.03,1.08,1.21,1.78-1.15h.86s1.58-.67,1.58-.67l.5.22,1.08,2.35Z"/>
+        <path id="CD" class="cls-1" d="M136.38,102.87l-.17,3.26,1.12.37-.9.99-1.08.74-1.07,1.46-.59,1.29-.16,2.24-.65,1.06-.02,2.1-.81.78-.1,1.66-.39.21-.26,1.53.71,1.26.18,3.37.5,2.57-.28,1.46.56,1.62,1.63,1.57,1.51,3.55-1.1-.29-3.77.48-.75.33-.8,1.8.63,1.25-.5,3.35-.35,2.85.76.51,1.96,1.1.77-.51.24,3.08-2.15-.03-1.15-1.57-1.03-1.22-2.15-.4-.63-1.49-1.72.9-2.24-.4-.94-1.29-1.78-.26-1.31.07-.16-.88-.97-.07-1.28-.17-1.73.42-1.22-.07-.7.26.15-3.37-.93-1.05-.21-1.73.41-1.7-.56-1.09-.05-1.76-3.41.02.25-1.01h-1.43s-.15.5-.15.5l-1.74.11-.71,1.63-.42.71-1.55-.4-.92.4-1.86.22-1.07-1.47-.64-.91-.81-1.68-.69-2.09-8.27-.03-.99.33-.81-.05-1.16.38-.39-.87.71-.3.09-1.22.46-.72,1.02-.58.74.28.96-1.07,1.52.03.18.79,1.05.5,1.65-1.76,1.63-1.36.71-.89-.09-2.3,1.22-2.71,1.28-1.43,1.85-1.34.32-.89.07-1.02.46-.97-.15-1.58.35-2.47.55-1.74.84-1.49.16-1.68.25-1.95,1.1-1.42,1.5-.9,2.31.95,1.78,1.03,2.05.28,2.09.54.84-1.68.39-.22,1.27.28,3.13-1.39,1.1.59.91-.08.42-.68,1.04-.24,2.11.29,1.8.06.93-.29,1.69,2.31,1.26.33.75-.47,1.3.19,1.56-.59.67,1.19,2.47,1.86Z"/>
+        <path id="CF" class="cls-1" d="M92.76,91.92l2.32-.22.52-.72.46.06.7.63,3.53-1.07,1.19-1.1,1.47-.99-.28-.99.79-.26,2.71.18,2.64-1.31,2.02-3.09,1.43-1.14,1.77-.49.32,1.22,1.62,1.77v1.15l-.45,1.18.18.87.97.81,2.14,1.24,1.53,1.13.03.92,1.88,1.46,1.17,1.21.71,1.68,2.1,1.11.45.89-.93.29-1.8-.06-2.11-.29-1.04.24-.42.68-.91.08-1.1-.59-3.13,1.39-1.27-.28-.39.22-.84,1.68-2.09-.54-2.05-.28-1.78-1.03-2.31-.95-1.5.9-1.1,1.42-.25,1.95-1.8-.16-1.9-.47-1.67,1.48-1.47,2.6-.3-.81-.12-1.27-1.28-.9-1.04-1.44-.24-1-1.32-1.46.22-.83-.28-1.18.22-2.17.67-.51,1.41-2.84Z"/>
+        <path id="CG" class="cls-1" d="M86.36,125.98l-1.05-.96-.85.47-1.13,1.2-2.3-2.95,2.13-1.54-1.05-1.85.96-.7,1.89-.34.22-1.24,1.5,1.34,2.48.12.86-1.32.35-1.85-.31-2.18-1.32-1.64,1.21-3.23-.7-.55-2.08.22-.79-1.43.21-1.22,3.53.11,2.27.73,2.23.66.2-1.5,1.47-2.6,1.67-1.48,1.9.47,1.8.16-.16,1.68-.84,1.49-.55,1.74-.35,2.47.15,1.58-.46.97-.07,1.02-.32.89-1.85,1.34-1.28,1.43-1.22,2.71.09,2.3-.71.89-1.63,1.36-1.65,1.76-1.05-.5-.18-.79-1.52-.03-.96,1.07-.74-.28Z"/>
+        <path id="CI" class="cls-1" d="M41.91,98.72l-1.27.03-1.96-.55-1.79.03-3.33.49-1.94.81-2.78,1.02-.54-.07.21-2.3.27-.35-.08-1.11-1.19-1.17-.89-.19-.82-.77.61-1.24-.28-1.36.13-.82h.45l.16-1.23-.22-.54.27-.39,1.04-.34-.69-2.26-.65-1.16.23-.97.56-.21.36-.26.78.42,2.16.03.52-.83.48.06.81-.32.44,1.21.65-.36,1.16-.42,1.26.62.49.93,1.26.6.98-.71,1.32-.11,1.92.73.74,4.01-1.18,2.36-.73,3.17,1.21,2.41-.13,1.11Z"/>
+        <path id="CM" class="cls-1" d="M86.59,106.33l-.35-.15-1.66.36-1.71-.38-1.33.19-4.56-.07.41-2.2-1.1-1.84-1.28-.48-.57-1.25-.72-.4.04-.77.72-1.98,1.33-2.7.81-.03,1.67-1.64,1.07-.04,1.57,1.15,1.93-.95.26-1.16.63-1.14.43-1.42,1.5-1.16.57-1.97.59-.63.4-1.47.74-1.81,2.36-2.2.15-.95.31-.51-1.11-1.14.09-.9.79-.17,1.11,1.83.19,1.89-.1,1.89,1.52,2.57-1.56-.03-.79.2-1.28-.28-.61,1.33,1.65,1.65,1.22.48.4,1.17.88,1.93-.44.77-1.41,2.84-.67.51-.22,2.17.28,1.18-.22.83,1.32,1.46.24,1,1.04,1.44,1.28.9.12,1.27.3.81-.2,1.5-2.23-.66-2.27-.73-3.53-.11Z"/>
+        <path id="DJ" class="cls-1" d="M170.72,76.98l.66.88-.09,1.19-1.6.68,1.21.77-1.04,1.52-.62-.5-.67.2-1.57-.05-.05-.86-.21-.79.94-1.33.99-1.26,1.2.25.85-.7Z"/>
+        <path id="DZ" class="cls-1 dinosaurs" d="M83.57,45.34l-9.61,5.75-8.12,5.85-3.95,1.32-3.11.29-.03-1.88-1.3-.48-1.75-.85-.66-1.39-9.46-6.55-9.46-6.65-10.55-7.53.06-.61v-.21l-.03-3.75,4.53-2.36,2.8-.49,2.29-.86,1.08-1.62,3.28-1.29.12-2.41,1.62-.29,1.27-1.21,3.67-.56.51-1.28-.74-.71-.97-3.53-.16-2.05-1.06-2.18,2.69-1.87,3.04-.6,1.77-1.43,2.7-1.05,4.75-.62,4.64-.29,1.41.52,2.64-1.37,3-.03,1.14.81,1.91-.21-.57,1.79.45,3.28-.66,2.82-1.73,1.88.25,2.53,2.29,1.98.03.81,1.72,1.33,1.2,5.86.91,2.84.15,1.48-.49,2.59.2,1.44-.36,1.72.25,1.97-1.12,1.29,1.66,2.26.11,1.32.99,1.71,1.31-.56,2.22,1.42,1.23,1.91Z"/>
+        <path id="EG" class="cls-1 dinosaurs" d="M147.84,26.54l-.79,1.29-.6,2.4-.76,1.64-.66.56-.93-1.02-1.27-1.42-2-4.57-.28.29,1.16,3.37,1.72,3.18,2.12,4.88,1.03,1.68.9,1.74,2.52,3.4-.56.53.09,1.97,3.27,2.71.49.62h-33.27v-22.41l-.84-2.58.72-2-.43-1.39,1.01-1.57,3.73-.05,2.7.86,2.78.97,1.3.5,2.16-1.03,1.15-.93,2.48-.27,1.99.41.77,1.62.65-1.07,2.24.77,2.19.19,1.38-.82,1.84,5.55Z"/>
+        <path id="ER" class="cls-1" d="M168.67,77.43l-.96-.93-1.15-1.67-1.24-.92-.73-1-2.44-1.15-1.92-.03-.68-.61-1.64.68-1.7-1.31-.88,2.15-3.26-.6-.3-1.15,1.21-4.25.27-1.93.88-.9,2.07-.48,1.42-1.67,1.63,3.38.77,2.67,1.54,1.41,3.82,2.72,1.56,1.64,1.52,1.66.88.98,1.38.86-.85.7-1.2-.25Z"/>
+        <path id="ET" class="cls-1" d="M156.21,70.49l1.7,1.31,1.64-.68.68.61,1.92.03,2.44,1.15.73,1,1.24.92,1.15,1.67.96.93-.99,1.26-.94,1.33.21.79.05.86,1.57.05.67-.2.62.5-.61,1.01,1.04,1.56,1.03,1.36,1.07,1.01,9.17,3.34,2.36-.02-7.93,8.42-3.65.12-2.5,1.97-1.79.05-.77.88h-1.92l-1.13-.94-2.56,1.17-.83,1.16-1.87-.22-.62-.32-.66.07-.88-.02-3.55-2.38h-1.95l-.96-.91v-1.57l-1.46-.47-1.65-3.05-1.28-.65-.5-1.12-1.42-1.37-1.72-.2.96-1.61,1.48-.07.42-.86-.03-2.53.83-2.96,1.32-.8.29-1.16,1.2-2.17,1.69-1.42,1.14-2.81.45-2.47,3.26.6.88-2.15Z"/>
+        <path id="GA" class="cls-1" d="M81.03,123.74l-2.88-2.82-1.86-2.3-1.7-2.88.09-.92.61-.9.68-2.02.57-2.07.95-.16,4.07.03-.02-3.35,1.33-.19,1.71.38,1.66-.36.35.15-.21,1.22.79,1.43,2.08-.22.7.55-1.21,3.23,1.32,1.64.31,2.18-.35,1.85-.86,1.32-2.48-.12-1.5-1.34-.22,1.24-1.89.34-.96.7,1.05,1.85-2.13,1.54Z"/>
+        <path id="GH" class="cls-1" d="M52.9,96.1l-4.4,1.64-1.56.96-2.53.81-2.5-.79.13-1.11-1.21-2.41.73-3.17,1.18-2.36-.74-4.01-.39-2.13.07-1.61,4.87-.13,1.24.21.9-.46,1.3.22-.21.89,1.17,1.46v2.05l.27,2.22.7,1.03-.62,2.53.22,1.4.75,1.78.63.98Z"/>
+        <path id="GM" class="cls-1" d="M2.7,75.69l.36-1.27,3.05-.09.64-.67.89-.05,1.1.71h.87s.93-.47.93-.47l.56.82-1.21.65-1.22-.05-1.2-.61-1.04.66-.5.03-.68.4-2.55-.06Z"/>
+        <path id="GN" class="cls-1" d="M26.26,91.17l-.79-.07-.57,1.13h-.8s-.54-.61-.54-.61l.18-1.13-1.17-1.72-.73.31-.6.07-.77.16.03-1.03-.45-.74.09-.82-.61-1.19-.78-1.01h-2.24l-.65.53-.78.06-.48.61-.32.79-1.5,1.24-1.23-1.67-1.09-1.11-.72-.36-.7-.57-.31-1.25-.41-.62-.82-.47,1.25-1.38.85.05.73-.48h.62l.44-.38-.24-.94.31-.3.05-.97,1.35.03,2.02.7.62-.07.21-.31,1.52.22.41-.16.16,1.05h.45s.73-.39.73-.39l.46.1.78.72,1.2.23.77-.62.91-.38.67-.4.56.08.62.62.34.79,1.15,1.19-.58.73-.11.92.6-.28.35.34-.15.84.86.82-.56.21-.23.97.65,1.16.69,2.26-1.04.34-.27.39.22.54-.16,1.23h-.45Z"/>
+        <path id="GQ" class="cls-1" d="M76.54,109.83l-.53-.42.97-3.13,4.56.07.02,3.35-4.07-.03-.95.16Z"/>
+        <path id="GW" class="cls-1" d="M7.5,81.7l-1.5-1.19-1.18-.18-.64-.81v-.43s-.84-.6-.84-.6l-.18-.61,1.49-.47.93.09.75-.32,5.18.12-.05.97-.31.3.24.94-.44.38h-.62l-.73.48-.85-.05-1.25,1.38Z"/>
+        <path id="KE" class="cls-1" d="M164.86,115.04l1.66,2.29-1.96,1.12-.69,1.16-1.06.2-.39,1.97-.9,1.12-.55,1.86-1.13.92-4.02-2.79-.2-1.62-10.16-5.67-.48-.31-.02-2.95.8-1.13,1.38-1.84,1.02-2.03-1.23-3.2-.33-1.4-1.33-1.94,1.72-1.67,1.9-1.84,1.46.47v1.57l.96.91h1.95l3.55,2.38.88.02.66-.07.62.32,1.87.22.83-1.16,2.56-1.17,1.13.94h1.92l-2.45,3.17.03,10.15Z"/>
+        <path id="LR" class="cls-1" d="M28.3,100.48l-.74.02-2.89-1.33-2.54-2.13-2.39-1.53-1.89-1.81.67-.9.15-.81,1.26-1.53,1.31-1.31.6-.07.73-.31,1.17,1.72-.18,1.13.54.6h.8s.57-1.12.57-1.12l.79.07-.13.82.28,1.36-.61,1.24.82.77.89.19,1.19,1.17.08,1.11-.27.35-.21,2.3Z"/>
+        <path id="LS" class="cls-1 dinosaurs" d="M131.17,197.01l.98.96-.86,1.56-.48,1.05-1.56.5-.52,1.04-1,.32-2.1-2.49,1.49-2.03,1.52-1.25,1.31-.64,1.22.98Z"/>
+        <path id="LY" class="cls-1" d="M91.56,47.19l-1.98,1.12-1.58-1.66-4.43-1.31-1.23-1.91-2.22-1.42-1.31.56-.99-1.71-.11-1.32-1.66-2.26,1.12-1.29-.25-1.97.36-1.72-.2-1.44.49-2.59-.15-1.48-.91-2.84,1.37-.75.24-1.38-.3-1.35,1.93-1.26.86-1.05,1.37-.95.16-2.55,3.29,1.15,1.18-.29,2.34.56,3.72,1.47,1.31,2.92,2.52.64,3.95,1.36,2.99,1.61,1.37-.84,1.34-1.49-.65-2.51.88-1.6,2.02-1.55,1.93-.45,3.79.68.96,1.48h1.04s.89.57.89.57l2.79.39.68,1.08-1.01,1.57.43,1.39-.72,2,.84,2.58v28.37h-3.22s-.04,1.25-.04,1.25l-11.18-5.7-11.19-5.77-2.83,1.66Z"/>
+        <path id="MA" class="cls-1 dinosaurs" d="M46.03,15.57l-.97-3.53-.16-2.05-1.06-2.18-1.22-.04-2.9-.75-2.67.24-1.69-1.46-2.06-.02-.89,2.11-1.87,3.51-2.08,1.39-2.81,1.53-1.8,2.24-.38,1.74-1.07,2.82.7,4.03-2.34,2.68-1.4.85-2.21,2.17-2.61.35-1.41,1.21-.05.04-1.78,3.2-1.86,1.14-1.02,1.91-.06,1.65-.75,1.79-.94.49-1.56,1.94-.96,2.15.18,1.02-.92,1.57-1.08.82-.13,1.39-.12,1.27.61-1,10.98.02-.53-4.35.69-1.55,2.62-.27-.09-7.86,9.21.17v-4.73l.06-.61v-.21l-.03-3.75,4.53-2.36,2.8-.49,2.29-.86,1.08-1.62,3.28-1.29.12-2.41,1.62-.29,1.27-1.21,3.67-.56.51-1.28-.74-.71Z"/>
+        <path id="MG" class="cls-1 dinosaurs" d="M188.84,147.66l.74,1.21.69,1.89.46,3.46.72,1.35-.28,1.38-.49.86-.96-1.7-.52.86.53,2.14-.25,1.24-.77.67-.18,2.48-1.1,3.42-1.38,4.09-1.74,5.67-1.07,4.21-1.27,3.55-2.28.73-2.45,1.31-1.61-.79-2.23-1.1-.77-1.62-.19-2.71-.98-2.42-.26-2.17.5-2.16,1.29-.52v-.99s1.35-2.26,1.35-2.26l.25-1.88-.65-1.4-.53-1.85-.22-2.7.98-1.63.37-1.85,1.4-.1,1.57-.6,1.03-.52,1.24-.04,1.59-1.65,2.31-1.78.84-1.44-.38-1.23,1.19.35,1.55-1.99.05-1.72.93-1.28.98,1.23Z"/>
+        <path id="ML" class="cls-1" d="M15.8,71.48l.94-.52.47-1.7.89-.07,1.96.8,1.58-.57,1.08.19.43-.64,11.25-.04.62-2.03-.49-.36-1.35-12.68-1.35-13.06,4.29-.05,9.46,6.65,9.46,6.55.66,1.39,1.75.85,1.3.48.03,1.88,3.11-.29v6.75s-1.53,1.94-1.53,1.94l-.24,1.79-2.49.45-3.82.25-1.04,1.03-1.8.11h-1.79s-.7-.54-.7-.54l-1.55.41-2.62,1.2-.53.9-2.18,1.28-.38.74-1.17.58-1.36-.38-.77.7-.41,1.96-2.23,2.36.07.96-.77,1.21.19,1.64-1.16.42-.65.36-.44-1.21-.81.32-.48-.06-.52.83-2.16-.03-.78-.42-.36.26-.86-.82.15-.84-.35-.34-.6.28.11-.92.58-.73-1.15-1.19-.34-.79-.62-.62-.56-.08-.67.4-.91.38-.77.62-1.2-.23-.78-.72-.46-.1-.73.38h-.45s-.16-1.04-.16-1.04l.13-.89-.24-1.1-1.05-.81-.55-1.64-.13-1.79Z"/>
+        <path id="MR" class="cls-1" d="M15.8,71.48l-1.85-1.98-1.7-2.13-1.86-.77-1.34-.85-1.57.03-1.36.63-1.4-.25-.96.93-.24-1.56.78-1.44.35-2.75-.31-2.91-.34-1.47.28-1.47-.72-1.42-1.48-1.28.61-1,10.98.02-.53-4.35.69-1.55,2.62-.27-.09-7.86,9.21.17v-4.73l10.55,7.53-4.29.05,1.35,13.06,1.35,12.68.49.36-.62,2.03-11.25.04-.43.64-1.08-.19-1.58.57-1.96-.8-.89.07-.47,1.7-.94.52Z"/>
+        <path id="MW" class="cls-1 dinosaurs" d="M146.82,144.95l-.78,2.16.78,3.72.98-.04,1.01.92,1.17,2.08.24,3.72-1.21.61-.86,2.01-1.83-1.79-.2-2.04.59-1.35-.17-1.15-1.1-.73-.78.26-1.61-1.38-1.47-.74.85-2.66.88-.99-.54-2.36.56-2.3.48-.77-.71-2.4-1.32-1.26,2.74.52.57.78.95,1.32.78,3.86Z"/>
+        <path id="MZ" class="cls-1" d="M146.82,144.95l2.11-.23,3.37.8.74-.36,1.95-.07,1-.85,1.68.04,3.06-1.1,2.23-1.64.46,1.27-.12,2.83.35,2.5.11,4.48.49,1.4-.83,2.07-1.09,2.01-1.79,1.8-2.56,1.11-3.16,1.41-3.17,3.15-1.08.54-1.96,2.09-1.15.69-.24,2.12,1.33,2.25.55,1.76.04.9.49-.15-.08,2.96-.45,1.41.66.52-.42,1.27-1.17,1.09-2.31,1.04-3.37,1.66-1.23,1.15.24,1.3.71.21-.24,1.64-2.12-.02-.24-1.38-.42-1.39-.24-1.11.5-3.43-.73-2.17-1.34-4.26,2.95-3.41.74-2.15.43-.27.31-1.74-.45-.88.12-2.2.55-2.04v-3.69s-1.46-.94-1.46-.94l-1.34-.21-.6-.72-1.3-.61-2.34.06-.18-1.08-.27-2.05,8.51-2.38,1.61,1.38.78-.26,1.1.73.17,1.15-.59,1.35.2,2.04,1.83,1.79.86-2.01,1.21-.61-.24-3.72-1.17-2.08-1.01-.92-.98.04-.78-3.72.78-2.16Z"/>
+        <path id="NA" class="cls-1" d="M95.75,195.8l-2.08-2.39-1.1-2.3-.62-3.03-.69-2.25-.94-4.72-.06-3.63-.36-1.64-1.09-1.24-1.45-2.47-1.47-3.57-.61-1.85-2.29-2.87-.17-2.25,1.35-.55,1.68-.5,1.82.09,1.67,1.32.42-.21,11.37-.12,1.94,1.4,6.79.41,5.15-1.19,2.3-.67,1.82.17,1.1.66.03.24-1.58.66h-.86s-1.78,1.16-1.78,1.16l-1.08-1.21-4.32,1.03-2.09.09-.08,10.57-2.76.11v20.38s-2.51,1.63-2.51,1.63l-1.5.23-1.77-.6-1.26-.23-.47-1.36-1.11-.87-1.34,1.57Z"/>
+        <path id="NE" class="cls-1 dinosaurs" d="M55.96,79.14l.07-1.95-3.24-.65-.08-1.38-1.58-1.87-.38-1.31.22-1.4,1.8-.11,1.04-1.03,3.82-.25,2.49-.45.24-1.79,1.54-1.94v-6.75s3.94-1.32,3.94-1.32l8.12-5.85,9.61-5.75,4.43,1.31,1.58,1.66,1.98-1.12.69,4.67,1.05.78.05.95,1.16,1.02-.61,1.28-1.08,5.98-.14,3.79-3.58,2.74-1.21,3.8,1.17,1.06v1.85s1.79.07,1.79.07l-.28,1.34-.79.17-.09.9-.53.07-1.89-3.13-.66-.12-2.19,1.6-2.17-.83-1.51-.17-.81.4-1.65-.08-1.65,1.22-1.43.07-3.39-1.48-1.33.7-1.43-.05-1.05-1.08-2.82-1.07-3.01.34-.73.62-.39,1.65-.81,1.15-.19,2.54-2.14-1.64h-1.01s-.94.84-.94.84Z"/>
+        <path id="NG" class="cls-1" d="M73.76,99.34l-2.91,1-1.07-.14-1.08.62-2.24-.06-1.5-1.75-.92-2.02-1.99-1.84-2.11.03h-2.47l.16-4.53-.07-1.79.53-1.77.86-.87,1.36-1.75-.29-.76.55-1.14-.63-1.68.11-.95.19-2.54.81-1.15.39-1.65.73-.62,3.01-.34,2.82,1.07,1.05,1.08,1.43.05,1.33-.7,3.39,1.48,1.43-.07,1.65-1.22,1.65.08.81-.4,1.51.17,2.17.83,2.19-1.6.66.12,1.89,3.13.53-.07,1.11,1.14-.31.51-.15.95-2.36,2.2-.74,1.81-.4,1.47-.59.63-.57,1.97-1.5,1.16-.43,1.42-.63,1.14-.26,1.16-1.93.95-1.57-1.15-1.07.04-1.67,1.64-.81.03-1.33,2.7-.72,1.98Z"/>
+        <path id="RW" class="cls-1" d="M135.21,115.81l1.12,1.57-.17,1.64-.8.35-1.49-.18-.86,1.59-1.71-.22.26-1.53.39-.21.1-1.66.81-.78.68.29,1.67-.86Z"/>
+        <path id="SD" class="cls-1" d="M145.15,86.16l-.39-.05.05-1.41-.34-.97-1.44-1.12-.34-2.05.34-2.1-1.3-.19-.19.63-1.69.15.68.83.24,1.71-1.54,1.56-1.4,2.04-1.44.29-2.36-1.65-1.06.58-.29.83-1.44.53-.1.58h-2.79l-.39-.58-2.02-.1-1.01.49-.77-.25-1.44-1.65-.48-.77-2.03.39-.77,1.31-.72,2.52-.96.53-.86.31-.23-.14-.97-.81-.18-.87.45-1.18v-1.15l-1.62-1.77-.32-1.22.03-.69-1.03-.83-.03-1.66-.58-1.1-.99.17.28-1.05.73-1.2-.32-1.18.92-.88-.58-.67.74-1.78,1.28-2.13,2.42.2-.14-11.61.04-1.24h3.22v-5.97h33.27l.9,2.94-.61.54.41,3.06,1.03,3.52,1.06.73,1.54,1.08-1.42,1.67-2.07.48-.88.9-.27,1.93-1.21,4.25.3,1.15-.45,2.47-1.14,2.81-1.69,1.42-1.2,2.17-.29,1.16-1.32.8-.83,2.96v.33Z"/>
+        <path id="SL" class="cls-1" d="M17.85,93.7l-.76-.21-2.01-1.13-1.46-1.5-.49-1.03-.35-2.08,1.5-1.24.32-.79.48-.61.78-.06.65-.53h2.24l.78,1.01.61,1.19-.09.82.45.74-.03,1.03.77-.16-1.31,1.31-1.26,1.53-.15.81-.67.9Z"/>
+        <path id="SN" class="cls-1" d="M3.06,74.42l-1.16-2.24-1.4-1.02,1.24-.55,1.36-2.03.66-1.49.96-.93,1.4.25,1.36-.63,1.57-.03,1.34.85,1.86.77,1.7,2.13,1.85,1.98.13,1.79.55,1.64,1.05.81.24,1.1-.13.89-.41.16-1.52-.22-.21.31-.62.07-2.02-.7-1.35-.03-5.18-.12-.75.32-.93-.09-1.49.47-.46-2.19,2.55.06.68-.4.5-.03,1.04-.66,1.2.61,1.22.05,1.21-.65-.56-.82-.93.48h-.87s-1.1-.72-1.1-.72l-.89.05-.64.67-3.05.09Z"/>
+        <path id="SO" class="cls-1" d="M193.3,79.69l-.07-.79h-1.06s-1.33.99-1.33.99l-1.49.28-1.29.42-.89.06-1.6.1-1,.52-1.39.19-2.47.88-3.05.33-2.65.73h-1.39s-1.26-1.2-1.26-1.2l-.55-1.17-.91-.53-1.04,1.52-.61,1.01,1.04,1.56,1.03,1.36,1.07,1.01,9.17,3.34,2.36-.02-7.93,8.42-3.65.12-2.5,1.97-1.79.05-.77.88-2.45,3.17.03,10.15,1.66,2.29.63-.66.65-1.46,3.07-3.38,2.61-2.12,4.2-2.76,2.8-2.26,3.3-3.81,2.39-3.13,2.41-4.1,1.73-3.59,1.35-3.15.79-3.05.6-1.02v-1.5s.25-1.65.25-1.65Z"/>
+        <path id="SS" class="cls-1" d="M145.15,86.16l.03,2.2-.42.86-1.48.07-.96,1.61,1.72.2,1.42,1.37.5,1.12,1.28.65,1.65,3.05-1.9,1.84-1.72,1.67-1.73,1.28h-1.97l-2.26.65-1.78-.63-1.15.77-2.47-1.86-.67-1.19-1.56.59-1.3-.19-.75.47-1.26-.33-1.69-2.31-.45-.89-2.1-1.11-.71-1.68-1.17-1.21-1.88-1.46-.03-.92-1.53-1.13-1.91-1.1.86-.31.96-.53.72-2.52.77-1.31,2.03-.39.48.77,1.44,1.65.77.25,1.01-.49,2.02.1.39.58h2.79l.1-.58,1.44-.53.29-.83,1.06-.58,2.36,1.65,1.44-.29,1.4-2.04,1.54-1.56-.24-1.71-.68-.83,1.69-.15.19-.63,1.3.19-.34,2.1.34,2.05,1.44,1.12.34.97-.05,1.41.39.05Z"/>
+        <path id="SZ" class="cls-1" d="M139.85,190l-.57,1.39-1.64.33-1.68-1.69-.02-1.08.76-1.17.27-.9.81-.22,1.41.57.42,1.39.24,1.38Z"/>
+        <path id="TD" class="cls-1" d="M90.57,76.52l.28-1.34-1.8-.07v-1.85s-1.16-1.06-1.16-1.06l1.21-3.8,3.58-2.74.14-3.79,1.08-5.98.61-1.28-1.16-1.02-.05-.95-1.05-.78-.69-4.67,2.83-1.66,11.19,5.77,11.18,5.7.14,11.61-2.42-.2-1.28,2.13-.74,1.78.58.67-.92.88.32,1.18-.73,1.2-.28,1.05.99-.17.58,1.1.03,1.66,1.03.83-.03.69-1.77.49-1.43,1.14-2.02,3.09-2.64,1.31-2.71-.18-.79.26.28.99-1.47.99-1.19,1.1-3.53,1.07-.7-.63-.46-.06-.52.72-2.32.22.44-.77-.88-1.93-.4-1.17-1.22-.48-1.65-1.65.61-1.33,1.28.28.79-.2,1.56.03-1.52-2.57.1-1.89-.19-1.89-1.11-1.83Z"/>
+        <path id="TG" class="cls-1" d="M55.15,95.51l-2.25.59-.63-.98-.75-1.78-.22-1.4.62-2.53-.7-1.03-.27-2.22v-2.05l-1.17-1.46.21-.89,2.46.06-.36,1.5.85.83.98.99.1,1.39.57.58-.13,6.46.69,1.94Z"/>
+        <path id="TN" class="cls-1 dinosaurs" d="M76.51,23.95l-1.2-5.86-1.72-1.33-.03-.81-2.29-1.98-.25-2.53,1.73-1.88.66-2.82-.45-3.28.57-1.79,3.06-1.41,1.96.42-.08,1.77,2.38-1.29.2.67-1.41,1.71v1.6s.96.85.96.85l-.37,2.96-1.85,1.71.53,1.83,1.45.06.71,1.59,1.07.52-.16,2.55-1.37.95-.86,1.05-1.93,1.26.3,1.35-.24,1.38-1.37.75Z"/>
+        <path id="TZ" class="cls-1 dinosaurs" d="M144.98,115.29l.48.31,10.16,5.67.2,1.62,4.02,2.79-1.29,3.45.16,1.59,1.8,1.02.08.73-.77,1.7.16.85-.18,1.35.98,1.76,1.16,2.79,1.02.62-2.23,1.64-3.06,1.1-1.68-.04-1,.85-1.95.07-.74.36-3.37-.8-2.11.23-.78-3.86-.95-1.32-.57-.78-2.74-.52-1.6-.85-1.78-.47-1.12-.48-1.17-.71-1.51-3.55-1.63-1.57-.56-1.62.28-1.46-.5-2.57,1.16-.13,1.01-1.01,1.1-1.46.69-.58-.03-.91-.6-.63-.16-1.1.8-.35.17-1.64-1.12-1.57.99-.34,3.07.04,5.71-.22Z"/>
+        <path id="UG" class="cls-1" d="M139.27,115.51l-3.07-.04-.99.34-1.67.86-.68-.29.02-2.1.65-1.06.16-2.24.59-1.29,1.07-1.46,1.08-.74.9-.99-1.12-.37.17-3.26,1.15-.77,1.78.63,2.26-.65h1.97l1.73-1.28,1.33,1.94.33,1.4,1.23,3.2-1.02,2.03-1.38,1.84-.8,1.13.02,2.95-5.71.22Z"/>
+        <path id="ZA" class="cls-1 dinosaurs" d="M138.3,197.97l-.55.46-1.19,1.63-.78,1.66-1.59,2.33-3.17,3.38-1.98,1.98-2.12,1.51-2.93,1.3-1.43.17-.36.93-1.7-.5-1.39.64-3.04-.65-1.7.41-1.16-.18-2.89,1.33-2.39.54-1.73,1.28-1.28.08-1.19-1.21-.95-.06-1.21-1.51-.13.47-.37-.91.02-1.96-.91-2.23.9-.6-.07-2.53-1.84-3.05-1.41-2.74h0s-2.01-4.16-2.01-4.16l1.34-1.57,1.11.87.47,1.36,1.26.23,1.76.6,1.51-.23,2.5-1.63v-11.52l.76.46,1.66,2.93-.26,1.89.63,1.1,2.01-.32,1.4-1.39,1.33-.93.69-1.48,1.37-.72,1.18.38,1.34.87,2.28.15,1.79-.72.28-.96.49-1.47,1.53-.25.84-1.15.93-2.03,2.52-2.26,3.97-2.22,1.14.03,1.36.51.94-.36,1.49.3,1.34,4.26.73,2.17-.5,3.43.24,1.11-1.42-.57-.81.22-.26.9-.77,1.17.03,1.08,1.67,1.7,1.64-.34.57-1.39,2.13.03-.7,2.28-.33,2.62-.73,1.43-1.9,1.62ZM131.17,197.01l-1.22-.98-1.31.65-1.52,1.25-1.5,2.03,2.1,2.48,1-.32.52-1.03,1.56-.5.48-1.05.86-1.56-.97-.97Z"/>
+        <path id="ZM" class="cls-1" d="M141.78,138.47l1.32,1.26.71,2.4-.48.77-.56,2.3.54,2.36-.88.99-.85,2.66,1.47.74-8.51,2.38.27,2.05-2.13.4-1.59,1.15-.34,1.01-1.01.22-2.44,2.4-1.55,1.89-.95.07-.91-.34-3.13-.32-.5-.22-.03-.24-1.1-.66-1.82-.17-2.3.67-1.83-1.82-1.89-2.38.13-9.16,5.84.04-.24-.99.42-1.07-.49-1.33.32-1.38-.3-.88.97.07.16.88,1.31-.07,1.78.26.94,1.29,2.24.4,1.72-.9.63,1.49,2.15.4,1.03,1.22,1.15,1.57,2.15.03-.24-3.08-.77.51-1.96-1.1-.76-.51.35-2.85.5-3.35-.63-1.25.8-1.8.75-.33,3.77-.48,1.1.29,1.17.71,1.12.48,1.78.47,1.6.85Z"/>
+        <path id="ZW" class="cls-1 dinosaurs" d="M137.38,176.26l-1.49-.3-.95.36-1.35-.51-1.14-.03-1.79-1.36-2.17-.46-.82-1.9v-1.05s-1.21-.32-1.21-.32l-3.17-3.25-.89-1.71-.56-.52-1.08-2.35,3.13.32.91.34.95-.07,1.55-1.89,2.44-2.4,1.01-.22.34-1.01,1.59-1.15,2.13-.4.18,1.08,2.34-.06,1.3.61.6.72,1.34.21,1.45.94v3.69s-.54,2.04-.54,2.04l-.12,2.2.45.88-.31,1.74-.43.27-.74,2.15-2.95,3.41Z"/>
+      </g>
+    </g>
+  </svg>`,
+  antarctica: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="AQ" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1210.26 985.5">
+    <defs>
+      <style>
+        #AQ {
+          fill: #2a7035;
+          cursor: pointer;
+          stroke
+        }
+        #AQ:hover {
+          fill: #194520;
+        }
+      </style>
+    </defs>
+    <g data-name="Layer 2">
+      <g>
+        <path class="cls-1" d="M50.83,153.78c.25-1.08,1.34-1.33,1.34-2.67-2.8,2.92-4.08-1.59-4.01-5.35,1.23-.85,1.15,1.31,1.34,1.34,1.85.27,3.38-3.87,4.01-4.01,3.88-.89,6.1-.19,9.36,1.34-2.07,1.94-1.18,6.85-1.34,10.7-1.58,1.54-4.73,1.51-8.02,1.34,2.47-4.5-.35-.16-2.68-2.67v-.02Z"/>
+        <path class="cls-1" d="M0,164.48c3.42-.15,2.65-4.48,2.67-8.02.52-.7,6.49-1.98,2.67-2.67-.22-2.9,4.74-.61,6.69-1.34v5.35c-1.96,3.1-4.24-2.61-4.01,2.67-.82-.07-1.58-.2-1.34-1.34-1.35.06-.31,13.4-6.69,8.02v-2.67h0Z"/>
+        <path class="cls-1" d="M112.34,362.38c3.13,2.78-2.85,7.77-4.01,10.7-7.01-1.27-6.08-3.1-6.69-9.36,2.27.49,2.52-1.04,2.67-2.68h6.69c-.38,2.39,1.26,2.74,1.34,1.34h0Z"/>
+        <path class="cls-1" d="M346.35,402.49c.16-1.85,4.88-.87,4.01,0-.06.06-1.25.05-1.34,0-.07-.04.56,2.31,1.34,1.34,1.9-2.38,3.19,3.42,4.01,8.02-1.53-1.61-3.39-.52-1.34,0-2.02,4.04-5.87-1.11-9.36,0-.16-3.4-1.95-5.19-5.35-5.35-.25-1.14.52-1.26,1.34-1.34.03-1.88-1.7-.28-1.34,0-4.65-3.64-3.09-8.91-4.01-17.38h5.35c-.17,2.54.4,3.87,1.34,5.35.29.45.9,3.65,1.34,4.01.15.12,2.23.45,1.34,1.34-.04.04-1.31.04-1.34,0,1.03,1.5,7.34.17,4.01,4.01h0Z"/>
+        <path class="cls-1" d="M172.51,401.16c2.12-1.58,3.02,3.67,4.01,2.67,1.84-1.84.57-.1,4.01,1.34-.45,4.73-.08,6.89-1.34,8.02-2.39.38-2.74-1.26-1.34-1.34-.16-2.3-3.68-.47-5.35-1.34.1.05-1.29-4.69-4.01-4.01,1.33-1.78,3.62-2.62,4.01-5.35h.01Z"/>
+        <path class="cls-1" d="M521.51,768.88c-1.36.87-1.38,3.07-4.01,2.67.55-4.61-.66-.76-4.01-1.34-.21-.04.2-2.48,0-2.68-.11-.1-3.35,1.82-4.01,0-.1-5.69-2.24-9.35-1.34-16.05-.22-1.55.3-2.38,1.34-2.67-2.98-2.46-4.4-5.5-2.67-9.36h8.02c-.46,1.41,2.55,2.1,2.67,2.68.16.78-1.05,1.86-1.34,1.34.81,1.48,1.62.28,2.67,1.34-.35-.35,1.53,3.68,1.34,2.67.24,1.26-1.25,2.93-1.34,2.67.45,1.35,3.95,1.95,1.34,2.67.98,1.25,2.43,2.03,4.01,2.67-1.49,1.65-1.16,2.74,1.34,2.67v10.7c-4.35-4.21.04,4.49-4.01,0v.02Z"/>
+        <path class="cls-1" d="M837.08,985.5v-1.34c-3.94.11-5.53-4.02-6.69-5.35-4.24-4.89-13.79,2.01-18.72-1.34.35,3.53-5.23,3.59-8.02,4.01-5.78.86-13.15-.65-18.72,0-.3-1.04-1.12-1.55-2.67-1.34,2.4-4.56-.07-.68-1.34-1.34-.66-.34.25-3.2-1.34-2.67-.55,2.28-1.72-.28-4.01,0-.14-1.5,1.94-2.08,0-2.67-.65,2.38-1.43-.58-2.67-1.34-1.34,4.13-2.21-2.9-4.01-2.68-1.52,2.77-5.51.97-8.02,1.34-2.5.37-4.47,2.42-6.69,1.34-2.58-2.48-1.06-2.45-5.35-1.34-.14-1.5,1.94-2.08,0-2.68-.08,1-.2,1.34-1.34,1.34-1.03,0-1.3-1.33-1.34-1.34-2.02-.22-4.68-.27-6.69,0-1.45.19-.64.95-1.34,1.34-2.48,1.36-6.51,2.22-9.36,2.68-6.78,1.07-13.27-.79-21.4,0-.29-1.04-1.12-1.56-2.67-1.34,2.86-4.54-.74-.72-1.34-1.34-.76-.79-.71-3.19-1.34-4.01-.54-.71-1.99-1.96-2.67-2.67-.93-.97-3-.42-4.01-1.34-.4-.36.23-2.16,0-2.67-.05-.11-2.5-1.25-2.67-1.34.12.06-1.32,2.11-2.67,1.34-1.07-.61.52-2.48-1.34-2.67-.18-.02-1.21,2.4-2.67,1.34-.14-1.5,1.94-2.08,0-2.68-.29,1.04-1.12,1.56-2.68,1.34,2.69-4.66-.59.11-1.34-2.67.11-2.48-1.99.31-2.67,0-.67-.3-2.24-3.29-2.67-4.01-.59-.98-.6-1.88-1.34-2.67.81-7.94-.92-13.34-2.67-18.72,1.19-.78,7.63-2.81,4.01-4.01,0,.74-1.08,1.52-1.34,0,1.68-.55,2.73-1.72,2.67-4.01,2.72-.4,3.56-2.68,5.35-4.01,3.85-1.5,7.89-2.81,10.7-5.35,1.77-3.35-5.29-4.19-1.34-5.35,2.27-2.63,4.1-5.7,8.02-6.69,2.69-4.66-.59.11-1.34-2.67,2.79-2.53.88-6.69,1.34-9.36.28-1.62.73-3.03,1.34-5.35.39-1.47.54-3.38,1.34-5.35.48-1.18,3.84.67,1.34-1.34-.73-.59,2.79-2.1,2.67-2.67-.02-.13-2.24-.21-1.34-1.34,1.52-1.9,4.79-4.65,5.35-5.35.58-.73,1.85,1.22,1.34,1.34,1.49-.34,2.09-1.44,2.67-2.67-2.49-.65-2.05-3.93-2.67-5.35-.33-.75-2.94-2.68-2.67-4.01-.01.06,2.13-1.22,0-1.34-3.2-.17-3.14-7.67-1.34-8.02-4.41-.25-3.88-6.84-4.01-9.36-.06-1.15-1.28-1.09-1.34-1.34-.28-1.27.28-2.73,0-4.01-.25-1.14-1.28-1.73-1.34-2.67-.11-1.94,1.68-3.92,0-4.01-2.21-.12-2.23-8.52-2.67-9.36-.35-.65-2.95-2.79-4.01-4.01-1.1-1.27-2.51-5.4-4.01-2.67,1.5.28,2.08,1.48,2.67,2.67-1.96-1.87-7.6-5.62-8.02-6.69-.33-.84,1.68-2.29,1.34-2.67-.58-.65-3.34-.83-4.01-1.34-1.19-.91-6.7-6.99-6.69-6.69-.05-.88,1.81.07,0-1.34.6.47-.64,1.48-1.34,1.34-.9-.18-.75-3.3-1.34-4.01-.56-.68-2.3,1.71-2.67,1.34-.87-.88,1.36-1.42,1.34-1.34.41-1.58-6.39-7.48-5.35-4.01,1.5.28,2.08,1.48,2.68,2.67-2.34-1.22-4.65-2.48-4.01-6.69,5.33.43,9.88.07,12.03-2.67.14,1.5-1.94,2.08,0,2.67.02-2.66,3.65-1.7,2.67-5.35,2.93-.42,6.05-1.32,2.67-2.67.26.1-1.29,2.03-1.34,0,2.53-2.2,4.76-1.88,6.69-5.35-.3.53-1.49-.57-1.34-1.34.02-.1,1.1-.47,1.34-1.34,1.35-5.11-.15-7.24,1.34-10.7-3.6.92-2.59-2.77-4.01-4.01-.2-7.72-6.33-15.49-4.01-21.4-5.39-.05-3.31-6.2-4.01-9.36-.04-.19-1.1-.38-1.34-1.34-.48-1.91-.5-2.7-1.34-4.01-.77-1.2-1.24-2.24-1.34-4.01-.04-.83-1.35-1.51-1.34-1.34-.05-.82,1.57-2.27,1.34-2.67-.27-.47-3.2-3.2-5.35-5.35l-10.7-10.7c-1.9-1.9-4.25-4.57-5.35-5.35-1.62-1.15-2.46.1-4.01-1.34.65.6,1.24-2.54-1.34-2.67-4.25-.22-5.72-7.55-8.02-8.02-.55,2.28-1.72-.28-4.01,0,3.16-4.86-1.85.15-4.01-2.67-1.15-1.5,2.18-1.26,0-2.68.68.44-2.51.87-4.01,0-2.38-1.37-.24-5.71-2.67-2.67-.78.98-1.35-1.34-1.34-1.34-1.65-1.27-1.44-2.58-2.67-1.34-.82.82-2.53-2.4-4.01-1.34,2.41-4.19-.41-.69-4.01-2.67,2.33-4.6.02-.64-1.34-1.34-1.05-.54.52-2.53-1.34-2.67-.03,0,.06,1.28,0,1.34-1.43,1.43-2.39-3.65-2.67-4.01-.32-.4-2.08.71-1.34,1.34-2.56-2.15-2.22-2.81-4.01-4.01-.55,2.28-1.73-.28-4.01,0,.42-1.55,1.74-3.08,0-1.34-1.4,1.4-3.29-4.45-4.01-5.35-2.86-3.57-6.59-4.43-9.36-6.69-1.82-1.48-2.76-3.69-4.01-5.35.74,0,1.52-1.08,0-1.34-.76,3.2-3.82-3.24-1.34-4.01-4.91,0-5.12-4.68-8.02-6.69-1.72,3.73-5.34-1.64-6.69-2.67-.29,1.04-1.12,1.55-2.67,1.34-.14-1.5,1.94-2.08,0-2.67-1.34,4.13-2.2-2.9-4.01-2.67-.58,2.9-3.27-1.07-4.01-1.34-6.23-2.24-11.7-.57-17.38,0,.11,2.83,1.85,4.59,2.67,5.35.23.21-.2,1.09,0,1.34.49.61,3.33-2.46,4.01,0-.24,3.04,4.52,7.79,5.35,5.35-1.19-.39-5.85-3.62-2.67-4.01,3.17,3.92,6.75,7.83,9.36,10.7,1.11,1.22,3.61,1.93,4.01,4.01.18,3.3.2,6.44-1.34,8.02,2.86-.2,2.52,1.36,2.67,4.01.18,3.21,3.71,6.7,1.34,8.02,6,2.02,9.21,6.84,12.03,12.03,4.69-.24,4.4,4.52,8.02,5.35,1.82,1.8,2.55,9.34-1.34,9.36-4.16-7.88-11.07-13-17.38-18.72,0-.74-1.08-1.52-1.34,0,5.91,4.79,11.26,10.14,16.05,16.05-5.56-5.57-13.7-11.68-18.72-17.38-.98-1.11-.44-2.72-1.34-4.01-2.13-3.05-4.07-1.49-5.35-1.34-2.14.26-6.58-4.38-6.69-1.34,3.27,1.44-2.47,4.24-2.67,5.35.1-.58,1.25,1.95,1.34,2.67.61,5.07-.46,9.02,0,13.37-5.86-.97,1.08,2.66-2.67,2.67-1.43-4.55-.42,3.72-4.01,1.34-2.06-1.16-4.83-8.4-6.69-5.35.74,0,1.52,1.08,0,1.34-.7-1.98-1.98-3.37-2.67-5.35v-8.02c-1.93.71-3.26-5.21-4.01-1.34,3.34.92-2.48,3.57-1.34,6.69-2.94,1.52-4.44,4.47-8.02,5.35-.3,1.04-.04,2.64-1.34,2.67-.81,3.2,1.68,9.7-1.34,10.7.63-6.35-1.67-1.48-4.01-4.01-.23,2,.48,4.94-1.34,5.35-.07,3.98,3.93,6.78,1.34,9.36,2.3-.07,3.43,1.03,4.01,2.67-4.55,1.43,3.72.42,1.34,4.01-2.88,2.72-7.53-.33-9.36-2.67-1.85,3.05-9.31.5-13.37,1.34v12.03c-1.5-.14-2.08,1.94-2.67,0-1.77.02-1.22,2.34-2.67,2.67-3.42-1.02-6.15-1.97-8.02,1.34,3.11.5-1.55,1.62-1.34,2.67.05.23,2.17.08,1.34,1.34-1.71,1.86-6.31.82-8.02,2.67.78,4.74,3.3,12.78,1.34,16.05,4.71.45,7.67,7.41,5.35,10.7,6.22.38-2.08,2.42,2.67,2.67,1.89.1,1.52,3.12,2.67,4.01-2.19-.97-5.82,2.66-2.67,2.67,0-.74,1.08-1.52,1.34,0-1.57,2-5.56,1.58-5.35,5.35-.55.34-2.23-.45-2.67,0-.69.69,1.23,1.78,1.34,1.34-.32,1.37-1.63,3.38,0,1.34.58-.73,2.1,2.79,2.67,2.68.23-.05.08-2.17,1.34-1.34-4.26,3.23,8.06,5.31,4.01,8.02-.32-1.45-2.2-3.21-2.67-1.34,1.98,2.48,6.91,2,6.69,6.69v5.35c-2.19-.97-5.82,2.66-2.67,2.67,0-.74,1.08-1.52,1.34,0-4.07.83-2.61,7.19-6.69,8.02-2.21-3.14-7.8-2.9-13.37-2.67-.5-2.09.1-2.52-1.34-1.34-1.71,1.42-.09-2.98,0-2.67-.09-.32-2.48-.88-2.67-1.34-.71-1.65-1.14-1.77,0-2.67.98-.78-1.33-1.33-1.34-1.34.9,1.03-1.88.59-2.67,0-1.11-.83-1.64-3.03-2.67-4.01-2.12-2-4.94-3.02-6.69-6.69,3.47-3.03-4.2-2.93,0-6.69-5.96-2.9-12.76.29-18.72,1.34-3.08.41-1.49-3.86-4.01-4.01-.23,1.56-2.11,1.45-4.01,1.34-.81,3.63-2.65,1.21-4.01,4.01.2-.41,1.91,2.01,0,2.67-1.04.3-2.64.04-2.67,1.34-.74,0-1.52,1.08,0,1.34,3.48-2.77-2.27,1.49-1.34,4.01,1.12.46,1.05.5,0,1.34-1.67,1.34,3.14.08,2.67,0,1.41.24,2.44,8.69,4.01,4.01-3.65-2.08,6.9,1.01,4.01,6.69-3.08-2.65-1.49.34-4.01,1.34-4.27.79-4.14-2.51-6.69,0,.24-2.92-1.09-4.26-4.01-4.01-.28,1.62-1.48.82-2.67,0-1.43,1.25-.42,4.93-4.01,4.01-4.58.47-7.01-2.55-9.36-2.67-1.27-.07-.93-1.22-1.34-1.34-1.23-.36-2.76.27-4.01,0-4.64-1.01-14.49-3.29-17.38-4.01-1.57-.39-3.21-.22-2.67-2.67-5.7-.1-9.35-2.24-16.05-1.34-.88-1.79-2.86-2.49-5.35-2.67-.14-1.5,1.94-2.08,0-2.67.41,4.18-3.86-3.4-4.01,1.34-4.93,2.14-8.48,1.14-10.7,0-1.23-.63-2.73-.51-4.01-1.34-.92-.6-.53-3.41-1.34-4.01-1.52-1.14-1.21,2.12-2.67,0,.26.38,1.8-.44,1.34-1.34-.41-.8-2.12-.08-2.67-1.34-1.06-2.43-1.46-5.3,0-6.69-5.86.97,1.08-2.66-2.67-2.67-.78,3.2-2.81-3.24-4.01-4.01,1.5-.14,2.08,1.94,2.67,0-1.98-.7-3.37-1.98-5.35-2.67-2.45-4.23-6.07-7.3-9.36-10.7-1.54-1.58-1.51-4.73-1.34-8.02-3.14-.43-2.99-4.14-5.35-5.35-1.58,2.44-1.16-3.07-5.35,0-1.21-2.35-4.92-2.21-5.35-5.35-.43-3.14-2.11-5.02-1.34-9.36-4.33-.78-2.5-2.81-4.01-4.01,0,.74-1.08,1.52-1.34,0-.9-1.49-4.44.27-5.35-4.01.74,0,1.52-1.08,0-1.34-.76,3.2-3.82-3.24-1.34-4.01-1.77-.38-2.94-.06-1.34-1.34.97-.78-1.31-1.31-1.34-1.34.05.05.05,1.28,0,1.34-.77.97-1.28-1.27-1.34-1.34-3.48-4.18-6.26-7.24-9.36-10.7-.98-1.09-2.94-3.15-4.01-4.01,2.49,1.99-5.96-.84-4.01-2.67,2.38-.65-.58-1.43-1.34-2.67-.97-4.09,2.31-3.93,1.34-8.02-1.37-1.34-2.93-2.93-4.01-4.01-1.07-1.07-1.94.28-2.67-1.34.13.28,2.01-1.31,0-1.34-1.49,3.41-10.34-.54-12.04,2.67-1.19-2.38-3.73-3.4-2.67-8.02,1.34-2.84,5.4-6.48,8.02-4.01,1.15-4.65,5.56-6.03,8.02-9.36-2.07-1.49-2.79-4.35-6.69-4.01-1.23,2.34-4.18,2.95-8.02,2.67-1.84-1.72-1.32-5.81-1.34-9.36,2.5.07,2.83-1.03,1.34-2.67,1.91-.32,2.26-2.2,4.01-2.67.14,1.5-1.94,2.08,0,2.67,1.34-1.78,1.26-4.98,4.01-5.35,3.98-.05,10.69,3.98,12.03-1.34-6.65-.53,1.81-3.78-2.67-4.01-3.51-.18-1.75-9.24-1.34-8.02-.28-.83-2.37.21-2.67-1.34-.23-1.12,2.02-2.57,0-2.68-3.03-.16.33-3.35,0-4.01-.49-.98-2.21.26-2.68-1.34-.35-1.22,1.63-1.77,1.34-2.67-.25-.79-2.37.18-2.67-1.34-.26-1.29,1.29-1.63,1.34-2.67-1.78-1.34-4.98-1.27-5.35-4.01,1.5-.14,2.08,1.94,2.67,0-6.37-1.2-7.38-7.78-6.69-16.05-2.17-2.73-9.87.06-12.04-2.67-.3-2.53,1.51-2.95,1.34-5.35-4.74-.15,2.84-4.42-1.34-4.01-.25,1.98-3.23,1.23-5.35,1.34-2.31.81.1,6.34-2.68,6.69-2.66-1.61-5.71-1.1-6.69,0,2.38.65-.58,1.43-1.34,2.67,2.67,1.48-1.63,7.18,0,12.03-2.66.02-1.7,3.65-5.35,2.67-.28-1.61-1.48-.82-2.67,0-2.74-4.4-.72-13.54-1.34-20.06-1.04-.3-1.55-1.12-1.34-2.68-5.86.97,1.08-2.66-2.67-2.67.29,1.76-1.86.66-2.67,1.34-3.25,2.68-.33,10.59-1.34,14.71,1.62.28.82,1.48,0,2.67,2.92-.24,4.25,1.1,4.01,4.01-1.62.28-.82,1.48,0,2.67-3.52-3.54-2.83,3.96-6.69,4.01-.14-1.5,1.94-2.08,0-2.67-1.07,4.98-2.99,2.15-4.01,4.01-.22.4,1.83.09,0,1.34-1.2.82-2.24,1.14-4.01,1.34.07-2.5-1.03-2.83-2.67-1.34,3.02-3.51-1.11-1.44-1.34-4.01-.41-4.64,3.2-10.19,4.01-17.38.7-6.2-.52-11.98,0-18.72,2.5.07,2.83-1.03,1.34-2.67,2.3.07,3.43-1.03,4.01-2.68-1.64-.02.34-7.56,1.34-9.36-5.86.97,1.08-2.66-2.67-2.67-.28.87-.72,1.35-1.34,0-1.29-2.86.68-3.29-1.34-4.01-.17-.06-.71,2.09-1.34,0-1.29-4.31-2.13-10.07-1.34-16.05.66-4.98,1.15-5.47,0-9.36-.66-2.24-.78-3.17-1.34-5.35-.06-5.37-.12-9.78-1.34-10.7-2.02,1.99-5.87,2.15-9.36,2.67v-12.03c3.45.22,6.43-6.8,8.02-4.01-1.45.32-3.21,2.2-1.34,2.67.28-1.5,1.48-2.08,2.68-2.67,0-1.34,1.08-1.59,1.34-2.67.74,0,1.52-1.08,0-1.34,0,.74-1.08,1.52-1.34,0,4.43-.47,2.1-7.71,2.67-12.03,4.31-2.14,5.78-6.08,8.02-8.02.81-.7.23-1.39,1.34,0,.78.97,1.3-1.3,1.34-1.34-.04.04-1.29.04-1.34,0-.88-.88,3.03-1.9,2.68-2.67-1.04-.3-1.56-1.12-1.34-2.67,4.72-1.35,2.45-2.31,4.01-5.35,1.04-.3,2.64-.04,2.67-1.34.74,0,1.52-1.08,0-1.34,0,.74-1.08,1.52-1.34,0,1.78-1.34,4.98-1.27,5.35-4.01-2.95-.17-.57-5.67-1.34-8.02h-10.7c-.15-1.63-.4-3.17-2.67-2.67,2.83-3.46-.83-1.51-1.34-4.01-.49-2.42.35-4.41,0-8.02,1.67-.15,1.86-1.66,2.67-2.67.39-.49,2.06.88,1.34,1.34,2.04-1.28,2.04-3.26,4.01-4.01,1.55-.6,3.74.46,5.35,0,1.53-4.27,1.38-10.21,9.36-8.02-.57-7.11-.15-13.49-1.34-20.06-.57-3.15-2.59-6.05-2.68-9.36-.27-10.43,4.05-19.92,2.68-29.42-.68-4.71-3.85-7.55-4.01-12.04-2.51-.32.76-2.2-1.34-2.67,0,.74-1.08,1.52-1.34,0-1.71-3.35.68-4.19-1.34-5.35-2.24,4.37-3.97-.72-6.69-1.34.06.01-1.28,1.56-2.68,1.34-.08-.01.08-1.24,0-1.34-.26-.31-2.24.1-2.67,0-1.94-.45-4.28-.72-6.69-1.34-.11-.03.09-1.23,0-1.34-.28-.34-3.51.03-4.01,0-1.22-.06-1.01-1.25-1.34-1.34-7.24-1.88-14.32-1.52-21.4-2.67-.29-1.04-1.12-1.55-2.67-1.34,2.39-4.56-.07-.68-1.34-1.34-.66-.34.25-3.2-1.34-2.67-.03,2.05-6.31-.67-8.02-1.34-.02,2.32-.58,2.26-2.67,2.67-4.18.82-10.74-1.37-14.71-1.34,4.15-4.13-4.53-2.26-2.67-8.02,3.94,1.88,1.58.68,1.34-1.34-.27-2.22,1.6-1.08,0-2.67-.82-.82,2.4-2.54,1.34-4.01-1.58-1.54-4.73-1.51-8.02-1.34.24-3.04-4.52-7.79-5.35-5.35,1.19.39,5.85,3.62,2.67,4.01-1.55-3.35-5.33-4.48-5.35-9.36h-8.02c1-2.7-1.86-1.73-1.34-1.34-1.15-.85-1.69-1.7-2.67-2.67-.17,2.28-3.33.83-5.35,1.34-.71.18-1.39,1.35-1.34,1.34-1.21.26-2.56-1.35-2.67-1.34-1.76.11-1.23,2.27-2.67,2.67-2.86-4.27-6.29-7.97-9.36-12.03.02-1.77,2.34-1.22,2.68-2.67-1.85-2.16-3.79-4.23-5.35-6.69,1.5-.14,2.08,1.94,2.67,0-2.13-1.34-5.38-1.27-2.67-4.01-2.3-1.45-5.98-2.46-6.69-4.01-.57-1.26,1.29-1.13,1.34-1.34.3-1.48-.91-.76-1.34-1.34.23.31-.51,1.44-1.34,0-.83-1.45-2.74-3.76-2.67-8.02.02-1.77,2.34-1.22,2.67-2.67-1.62-.28-.82-1.48,0-2.68-5.86.98,1.08-2.66-2.67-2.67-2.52,3.37-1.26-4.67-2.67-4.01-1.85,4.98-.2-3.94-2.67-2.67-.14-1.5,1.94-2.08,0-2.67-.77,2.79-8.65,1.67-8.02-1.34-1.14-4.26,1.17-5.08,2.67-6.69-.74,0-1.52-1.08,0-1.34,1.43-4.99-3.32-3.81-2.67-8.02-.41-6.72,1.54-15.81-2.67-18.72,1.5-.14,2.08,1.94,2.67,0-3.18-.14-2.94-3.7-4.01-5.35-1.3-2.01-4.29-4-4.01-6.69.78-4.33,2.81-2.5,4.01-4.01-1.98-.25-1.23-3.23-1.34-5.35,1.78-2.4,2.08-2.98,2.67-5.35.35-1.38,1.79-2.18,1.34-4.01-.13-.54-1.38-1.63-1.34-1.34-.18-1.33,1.27-1.02,1.34-1.34.26-1.31-2.65-.32-4.01-1.34-1.12-.83-1.6-3.07-2.67-4.01v-10.7c1.49-2.16,9.34-4.8,6.69-6.69,0,.74-1.08,1.52-1.34,0,1.5-.28,2.08-1.48,2.67-2.67,2.33.1,4.16-.3,5.35-1.34-.92,5.38,1.28,7.63,1.34,12.04-1.37,1.31-2.7,2.64-4.01,4.01.74,0,1.52,1.08,0,1.34,2.25,2.88-.39,5.48,0,6.69.1.29,2.42-.25,2.67,0,.42.43-.81,2.05-1.34,1.34.85,1.15,1.7,1.69,2.67,2.67.06,1.84-.63,2.94-1.34,4.01,2.5-.07,2.83,1.03,1.34,2.67,5.75-.63,1.71,3.6,4.01,4.01-.17-.74,1.3-1.52,1.34,0,.14,5.1-3.45,14.25-2.67,22.73,2.86,2.94-1.84,2.88,1.34,4.01-.24-.09,1.27-2.04,1.34,0,0,1.34,1.08,1.59,1.34,2.68,1.79,3.56.43,10.27,4.01,12.04,1.58-2.44,1.16,3.07,5.35,0,.14,1.5-1.94,2.08,0,2.67,1.14-3.92,3.37,1.74,6.69,0,.07,1.71,1.71,1.85,1.34,4.01,1.5-.14,2.08,1.94,2.67,0-.74,0-1.52-1.08,0-1.34,3.52,3.4-.91,3.4-1.34,5.35.01-.04,1.56,1.26,1.34,2.67-.07.45-2.33,1-1.34,2.68,2.38.74,5.55.69,6.69,2.67-3.87,2.96,2.05-1.68,1.34,1.34-.18,7.23-5.22,1.29-10.7,2.67-1.16,5.17,3.27,4.76,4.01,8.02-.6-1-3.26-.46-2.67,0,2.68,2.13,4.86-1.61,9.36-2.68-.29,1.68-2.08,2.73,1.34,2.68-.41,4.6,2.47,11.74,2.68,13.37.15,1.26-1.53,1.72-1.34,2.67.04.21,2.45-.18,2.68,0,.66.55-1.61,2.23-1.34,2.67,1.63.15,3.17.4,2.67,2.67,4.14-.13,7.42.61,9.36,2.67,4.54-.08,7.35,1.57,9.36,4.01,1.88-2.14,4.61-3.42,8.02-4.01.97-4.09-2.31-3.93-1.34-8.02,1.76-1.36,4.93-1.31,6.69-2.68-.97,3.65,2.66,2.69,2.67,5.35,8.2,1.73,17.29-1.23,25.41,0,7.48,1.14,12.09,6.53,18.72,6.69,1.14-2.02,3.27.38,4.01,0,.78-.4-.08-4.11,2.68-2.67,2.72.4,3.56,2.68,5.35,4.01,2.02-4.04,5.87,1.11,9.36,0,.65,1.58,1.43,3.03,2.67,4.01.32-3.45,2.3,1.84,5.35,0,.32,1.91,2.2,2.26,2.67,4.01v5.35c4.85-.95,5.73-5.86,13.37-4.01.29,1.04,1.12,1.55,2.67,1.34-2.25,4.31.62,3.58,1.34,2.67,1.36-1.7,3.02,5.27,4.01,4.01,1.71-3.47,6.37-.6,6.69-2.67-.74,0-1.52-1.08,0-1.34,5.3-.39,8.17,1.64,10.7,4.01-1.62.28-.82,1.48,0,2.67-3.77-1.15-1.15,2.46-1.34,2.67-.81.95-4.54,1.34-5.35,2.67-.46.75,1.25,2.09,1.34,2.67-.02-.11-2.1,1.17,0,1.34.68-1.84,2.89.21,5.35,0,.07,0-.07-1.26,0-1.34.89-1.09,5.47.83,6.69,0,1.66,3.63,5.16,3.33,2.68,8.02,2.92,1.54,8.65.27,9.36,4.01-2.38.65.58,1.43,1.34,2.67-1,.08-1.34.2-1.34,1.34,0,.86,1.34,1.52,1.34,1.34.05,2.31-1.62,3.42,0,4.01.88-4.01,2.86,1.28,5.35,0,.21,1.57-.23,3.8,1.34,4.01.3-1.04.04-2.64,1.34-2.67.63,5.75-3.6,1.71-4.01,4.01-.06,2.08-9.49-2.89-8.02,4.01,3.78,1.1,3.46-1.89,6.69-1.34-1.17,5.18,5.84,2.18,4.01,8.02,4.12-.94,2.57-.94,6.69,0,2.25,5.78,4.14,11.91,5.35,18.72-.02,1.77-2.35,1.22-2.67,2.67,1.92.14.05.91,0,1.34-.36,2.98.27,9.16,0,13.37,1.52,2.94,4.47,4.44,5.35,8.02.21,1.57-.23,3.8,1.34,4.01.14,1.5-1.94,2.08,0,2.67.25-1.98,3.23-1.23,5.35-1.34-3.31,3.19,3.47,2.2,1.34,8.02-1.05,2.96-4.68,3.35-6.69,5.35-2,1.01.35,3.44,0,4.01-.64,1.05-3.16,1.58-4.01,2.67.31-.4,1.78.47,1.34,1.34-.34.66-3.2-.25-2.67,1.34,2.18-.6,2.18,7.29,0,6.69-.52-2.05-1.61-.2,0,1.34-2.47-.61-3.98.97-5.35,1.34-2.55.69-5.3.36-5.35,1.34-.17,3.23-7.3,2.66-9.36,2.67-.67,4.21,2.91,5.23,0,8.02,4.61.35,5.68,1.53,9.36,1.34.85-.05,1.48-1.35,1.34-1.34,4.04-.27,7.93-1.52,10.7,0-.86-2.91,2.89-3.36,5.35-4.01,3.04-.81,8.4-.67,9.36,0-.98-5.86,2.66,1.08,2.67-2.67-3.82-.7,2.16-1.98,2.67-2.67-2.59-5.82,5.83-7.4,2.67-9.36-.75,1.93-2.09,3.26-4.01,4.01,1.93-2.52,2.99-5.93,8.02-5.35,1.71.07,1.85,1.71,4.01,1.34,2.56,2.49-1.98,3.84,1.34,4.01,2.52.13-.77,2.34,1.34,2.67,1.21-2.07,4.92-.7,5.35,0,1.14,4.26-1.17,5.08-2.67,6.69,1.04.3,1.55,1.12,1.34,2.67-4.58-2.66-.65.46-1.34,1.34-1.79,2.28-6.67,5.19-9.36,8.02,2.28.55-.28,1.73,0,4.01-4.3-3.04-.56,1.39-2.67,4.01-.7.86-3.81-.08-4.01,1.34.14,1.5-1.94,2.08,0,2.67.28-1.5,1.48-2.08,2.67-2.67-1.55,3.35-5.33,4.48-5.35,9.36,5.62,0,11.18-.04,12.03-1.34-2.28-.55.28-1.72,0-4.01,3.96,2.36.28-.62,2.67-4.01.66-.93,1.62,1.39,1.34,1.34,1.31.26.32-2.65,1.34-4.01,2.84-3.82,11.31-5.14,14.71,1.34-1.12,3.33-1.75,7.16-1.34,12.04-4.09-3.56-2.55,4.84-5.35,1.34-.69-.86-2.57,5.58-4.01,4.01-.14-1.5,1.94-2.08,0-2.67-.7,1.98-1.98,3.37-2.67,5.35-1.69-.61-.75,2.26-1.34,2.67-1.06.75-2.42-2.42-2.67,0,4.01.88-1.28,2.86,0,5.35,4.2-.7,7.46-2.35,10.7-4.01,1.5-.14,2.08,1.94,2.67,0-4.13-3.27,3.84,1.47,2.67-1.34-.74,0-1.52-1.08,0-1.34,1.71-.07,1.85-1.71,4.01-1.34.32-2.51,2.2.76,2.67-1.34-.74,0-1.52-1.08,0-1.34,1.7-1.86,6.55-.58,6.69-4.01-1.15-.82-3.26-1.54-4.01-2.67-.64-.96,1.7-1.95,1.34-2.67-.33-.65-3.21.25-2.67-1.34,1.37-1.31,2.7-2.64,4.01-4.01-1.79-.92.66-3.57-1.34-6.69,1.5-.14,2.08,1.94,2.67,0,2.51-.32-.76-2.2,1.34-2.67,3.52,3.48,1.05-2.16,5.35-1.34,1.81.42,2.55,1.9,5.35,1.34-.29,1.68-2.08,2.73,1.34,2.67,2.96,1.93-1.68,3.26,1.34,4.01-.06-.74,1.15-1.53,1.34,0,.4,3.27-.49,15.27,1.34,14.71,2.11-1.55,2.86-2.54,4.01-5.35,1.06-2.59-.76-7.35,2.67-8.02,10.89,0,21.46.17,30.76,1.34,3.49.44,7.59,3.38,10.7,1.34,1.69-3.32-1.93-4.27-2.67-5.35-1.35-1.98-2.18-3.74-4.01-5.35.41,2.72-3.86-.27-4.01-1.34-1.93-.75-3.26-2.09-4.01-4.01,3.61-3.97,6.5-8.65,9.36-13.37,1.34,0,1.59,1.08,2.67,1.34-1.15-6.2,2.1-1.4,2.67-4.01-3.06-3.1,4.17.45,6.69,0-.41,4.18,3.86-3.4,4.01,1.34-3.04,1.86,1.11-.17,1.34,0,.49.36.92,3.5,1.34,4.01.84,1.05.68-3.4,1.34-2.67,1.16,1.27,1.19,3.04,2.67,4.01-.96-.63,2.7-.35,4.01,0,.11.03-.09,1.23,0,1.34,1.6,1.95,8.31-1.28,8.02-1.34.23.05.08,2.17,1.34,1.34.45-4.76,7.65-3.52,10.7-5.35,4.22.35,5.05,4.5,8.02,6.69.88.65,3.09-.6,4.01,0,.68.44-.27,3.11,1.34,2.67,3.93-3.13,1.43-10.63,2.67-13.37-3.92-.09-4.97-3.06-6.69-5.35,1.5-.14,2.08,1.94,2.67,0-3.17-2.14-4.82-2.46-8.02-6.69.21.27,1.94-1.23,0-1.34-2.56-.14.42-1.86,0-2.67-.52-1.02-2.24.33-2.67-1.34-2.08-8.02,0-16.43-2.67-26.74,6.22-.38-2.08-2.42,2.67-2.67,2.7-.14,1.02-4.81,4.01-4.01,3.24,2.1,5.71,4.99,12.04,4.01,2.3-.53,3.43,1.98,4.01,0-2.37-1.77-.04-5.92,0-9.36.02-1.99-2.27-4.39,0-5.35-2.5.07-2.83-1.03-1.34-2.67-5.86.98,1.08-2.66-2.67-2.67-3.27,4.13,1.47-3.84-1.34-2.67-.46,1.43-1.21,1.02-1.34-1.34-.04-.83-1.35-1.51-1.34-1.34-.05-.88,1.58-2.2,1.34-2.67-.4-.78-4.11.08-2.67-2.67-.54-2.69,1.99-2.56,2.67-4.01-.06.12-2.11-1.32-1.34-2.67.38-.67,3.16.26,2.67-1.34-2.57-1.38.55-8.26,2.67-9.36-1.5-1.18-.73-5.37,0-8.02.21-.77,2.45.22,2.67-1.34.02-.14-2.4-1.24-1.34-2.67,1.91-.32,2.26-2.2,4.01-2.67,3.05-1.84,5.03,3.45,5.35,0-2.9.23-.61-4.74-1.34-6.69,5,3.23-.99-2.09,4.01-2.67,1.21-3.47-1.18-7.15-1.34-10.7-.18-4.24,2.01-6.4,0-6.69-.37,3.15-4.98-.12-5.35-1.34,1.61.14.82-1.94,0,0-4.65-2.73,3.88-7,0-9.36-2.58.23-6.35.27-9.36,1.34-.54.19-2,2.32-2.67,2.67-.17.09-1.74,1.44-1.34,1.34-1.11.29-2.57-2.07-2.67,0-.08,1.52-.54.65-1.34,1.34-4.4,3.81-14.86-.33-17.38-1.34-.33-.13-.98.02-1.34,0-1.39-.07-.78-1.12-1.34-1.34-1.92-.76-5.31.71-5.35-2.67,1.5-.14,2.08,1.94,2.67,0-1.58-.65-3.03-1.43-4.01-2.67-4.13-9.36,1.13-18.29,2.67-28.08,1.02-6.48.89-12.88,1.34-20.06,1.58-.65,3.03-1.43,4.01-2.67-2.38-.65.58-1.43,1.34-2.67-1.04-.3-1.55-1.12-1.34-2.68,1.04-.29,1.56-1.12,1.34-2.67,3.9.79.98-1.98,1.34-2.67.48-.93,2.15.14,2.67-1.34-.21.6-1.34-.72-1.34-1.34,0-2.41,2.71-3.45,2.67-5.35-1.57-2-5.56-1.57-5.35-5.35,1.09-.77,1.22,1.3,1.34,1.34,1.7.56.12-1.14,0-1.34-1.45-2.42,4.12-1.82,5.35-2.68.95-.66,2.33-3.64,4.01-5.35,1.98-2,4.8-3.24,5.35-4.01,1.36-1.9-.06-3.7,1.34-5.35,3.45,1.45,1.58-2.84,1.34-2.67,1.05-.73,2.9,2.55,4.01,0,1.73-1.89,1.78-5.42,2.67-6.69.78-1.1,2.26.47,2.67-1.34.01-.05-2.4-1.31-1.34-2.67,1.86-.64,5.15-3.96,5.35-4.01,1.11-.29,2.56,2.07,2.67,0,.17-3.32,2.82-2.49,4.01-1.34-.07-2.3,1.03-3.43,2.67-4.01-3.93-8.96,6.15-12.74,4.01-24.07,1.91-.32,2.26-2.2,4.01-2.68,1.22-2.03,3.17.39,4.01,0,.34-.16-.27-2.35,0-2.67.55-.66,2.23,1.61,2.67,1.34,1.03-.63-.53-2.57,1.34-2.67,1.48-.08.67-1,1.34-1.34,1.1-.55,1.81,1.58,2.67,1.34,1.92-.55-.03-3.31,0-4.01,0-.06,1.26.06,1.34,0,1.09-.89-.83-5.47,0-6.69-3.33-.15-4.14-2.5-6.69,0-.7.56-3.85,1.88-4.01,2.68-.13.66,1.18,1.92,1.34,1.34,1.02-3.73,1.71,2.12-1.34,1.34-2.11-.12-1.05-3.41-4.01-2.67,3.58-3.44-3.58-1.91,0-5.35-4.85,1.28-1.36-5.77-4.01-6.69,5.78-.01,5.28-6.31,9.36-8.02.14,1.5-1.94,2.08,0,2.67.7-1.98,1.98-3.37,2.67-5.35,3.26,2.22.87-.4,2.67-1.34,3.13-1.62,2.82.86,4.01-1.34-1.04-.29-1.55-1.12-1.34-2.67,3.14-2.23-.49.12-2.67,0-1.12-1.55-1.76-3.58-1.34-6.69,3.52,3.48,1.05-2.16,5.35-1.34.22-6.02,12.7-9.36,9.36-12.03-.89-.71,1.37-1.6,1.34-1.34.24-1.77-.58-4.52,1.34-6.69-1.79-.88-2.49-2.86-2.67-5.35-2.97-2.97-7.36-4.36-8.02-8.02,1.07-2.94,2.68-5.34,6.69-5.35.45,4.73.08,6.89,1.34,8.02,0-.74,1.08-1.52,1.34,0,1.84.06,2.94-.63,4.01-1.34.82-.08,1.58-.2,1.34-1.34-1.38-.33-.12-4.39,0-6.69.06-1.15,1.28-1.1,1.34-1.34.78-3.58-.71-8.68,2.67-9.36h5.35c-1.35,5.13-1.02,2.05,1.34,4.01.61.51-1.51,2.24-1.34,2.67.13.31,2.38-.25,2.67,0,.66.55-1.61,2.23-1.34,2.67,3.45,1.46,6.43,3.38,8.02,6.69.23,2-.48,4.94,1.34,5.35.25-1.98,3.23-1.23,5.35-1.34-1-13.48.98-23.98,9.36-28.08.55,2.28,1.72-.28,4.01,0,.43,2.93,1.32,6.05,2.68,2.67-.1.26-2.03-1.29,0-1.34,3.61,4.28-1.03,8.87,0,14.71,10,1.53,12.05-4.89,21.4-4.01,2.72.4,3.56,2.68,5.35,4.01.25-1.98,3.23-1.23,5.35-1.34,1.95-.22,1.49-1.47,2.67,0,.78.98,1.35-1.35,1.34-1.34-1.75,1.33-.1-2.99,2.67-2.67.52,1.91,1.6,2.26,0,4.01,5.86-.97-1.08,2.66,2.67,2.67.32-2.05,2.2-.19,2.67,1.34,2.54-2.07,10.41-7.77,8.02-9.36-1.7,2.17-6.78,9.35-8.02,6.69,4.51-2.33,5.36-6.27,9.36-8.02,2.68-1.17,5.28.03,5.35-1.34.16-3.12,6.49-2.93,6.69-1.34-.62-4.19,3.36-3.77,5.35-5.35,5.67,1.19,9.47-3.51,13.37,0-.53-5.71,7.03-2.82,6.69-2.67.33-.14-.26-2.36,0-2.67.56-.68,2.24,1.66,2.67,1.34,1.11-.82-.4-2.14,1.34-2.67,4.98-1.54,11.1-.89,17.38-4.01,4.08.23,1.38,5.96,4.01,2.68.67-.83,3.85,3.94,4.01,4.01-.68-.31,2.96-.9,2.67-1.34-2.14-.94,1.12-7.69,4.01-8.02,4.44,2.24,9.32,9.67,6.69,14.71,6.55-2,.95,6.5,4.01,2.67.72-.9,1.59,1.38,1.34,1.34,5.67.9,7.2-2.5,10.7,0-.42-5.85,6.18-3.72,6.69-5.35-3.87-2.03,10.31-1.36,12.03-2.67.3,1.04.04,2.64,1.34,2.67,3.15,3.74-2.09-2.53,1.34-1.34.72.25,1.15,3.04,2.67,2.67,2.51.32-.76,2.2,1.34,2.67,1.59-2.61,3.09.24,5.35,0,3.08-.33,5.35-2.15,9.36-2.67,3.34-.44,8.14.33,12.03,0,0,2.8,7.28,5.03,4.01,6.69-.32-1.45-2.2-3.21-2.67-1.34,3.24,2.11,5.91,4.78,8.02,8.02,1.5-.14,2.08,1.94,2.67,0-.74,0-1.52-1.08,0-1.34.65,1.58,1.43,3.03,2.67,4.01-.07-1.34,2.54-.97,4.01,0,2.55,1.68,4.73,7.67,9.36,6.69-2.39,4.56.07.69,1.34,1.34.66.34-.25,3.2,1.34,2.67.64-2.37,1.44.56,2.67,1.34-.35-.22.41-1.81,1.34-1.34.93.48-.15,2.16,1.34,2.67-.31-.11,1.47-1.56,2.67-1.34.09.02-.08,1.24,0,1.34.47.58,4.49,1.31,5.35,1.34,3.15.11,6.61-2.62,10.7-1.34-.79,3.9,1.98.98,2.67,1.34.66.34-.25,3.2,1.34,2.67.43-2.21,2.82-.76,4.01,0,4.06,2.59,16.29,3.51,20.06,4.01,8.38,1.1,14.35,2.65,20.06,4.01.14,1.5-1.94,2.08,0,2.67,2.73-2.74,6.1,2.23,10.7,2.67,2.37.23,3.12-1.05,5.35-1.34,2.83-.37,6.16.04,8.02,0,.33,0-.11-1.63,1.34-1.34.37.07-.08,2.04,1.34,1.34.85-.42.33-4.85,2.67-4.01.14,1.5-1.94,2.08,0,2.67,1.34-1.78,1.27-4.98,4.01-5.35.14,1.5-1.94,2.08,0,2.67,1.53-5.15,6.93-6.44,13.37-6.69.7,1.98,1.98,3.37,2.67,5.35-1.32,4.47-1.96,9.63-5.35,12.03,2.08.35.93,2.93,0,4.01,2.7-1,1.73,1.86,1.34,1.34.85,1.15,1.7,1.69,2.67,2.67-1.19,5.67,3.51,9.47,0,13.37,3.92.09,4.97,3.06,6.69,5.35-.06,1.84.63,2.94,1.34,4.01.14,1.5-1.94,2.08,0,2.67-.65-1.77,6.13-.08,6.69,1.34,3.54-1.57,7.87,2.56,8.02,0-2.65-1.88,5.19-4.52,6.69-6.69,2.48-1.98,2-6.91,6.69-6.69,5.74.9,12.76,3.34,20.06,4.01,3,.27,6.76-.48,9.36,0,.56.11-.04,2.48,1.34,2.68.14.02,1.24-2.4,2.67-1.34.41,1.44,1.83,1.64,2.67,2.67.37.45,2.26-.42,2.67,0,.67.67-.26,2.93,1.34,2.67,1.63-2.4,8.21,1.5,13.37,0,3.52,1.38,4.5,5.3,8.02,6.69.34-2.78,5.87-.36,6.69-2.67.14,1.5-1.94,2.08,0,2.67,1.34-1.78,1.27-4.98,4.01-5.35-3.55,3.53,3.96,2.83,4.01,6.69-1.5.14-2.08-1.94-2.67,0,6.45.23,3.29,10.08,4.01,16.05,1.72,1.84,5.81,1.32,9.36,1.34,1.95-2.5.05-8.86,2.68-10.7,1.84.06,2.94-.63,4.01-1.34,4.28,3.74,2.62,13.43,6.69,17.38.32-3.45,2.3,1.84,5.35,0,2.59-3.2,2.15-9.44,6.69-10.7.07,1.41,1.71,1.05,1.34-1.34h5.35c2.88,2.6,8.98,10.45,4.01,2.67,1.92,1.34,4.12,2.84,5.35,4.01.63.6,1.53,4.24,2.68,4.01-.05,0,.43-2.33,1.34-1.34.62.68-1.5,2.1-1.34,2.67.24.85,7.77,7.23,6.69,4.01-1.19-.39-5.85-3.62-2.67-4.01,1.13,3.37,5.52,5.13,6.69,6.69.94,1.25.34,2.94,1.34,4.01,3.64,3.9,8.58,6.61,12.04,10.7-.66,4.22,2.41,4.72,1.34,9.36.16,1.78-2.58,1.21-2.68,2.67,0,.04,2.49,1.41,1.34,2.67-.75.82-2.66-2.47-2.68,0,2.29,2.55-.14,2.54,1.34,4.01.89.89-.68,4.25,0,5.35-5.75-.63-1.71,3.6-4.01,4.01-2.11-.12-1.05-3.41-4.01-2.67-1.58.65-3.03,1.43-4.01,2.67,1.12.46,1.05.5,0,1.34-1.28,1.02,7.24,5.48,2.68,5.35-.32-1.45-2.2-3.21-2.68-1.34,1.78,1.34,4.04,2.21,5.35,4.01,2.91,6.41-.82,11.26-1.34,16.05-.42,3.9,2.64,7.28,0,10.7,1.04.3,2.64.04,2.67,1.34-.74,0-1.52,1.08,0,1.34,0-.74,1.08-1.52,1.34,0,.46,2.4,2.94,4.02,2.68,5.35.01-.06-2.13,1.23,0,1.34,3.12.16,2.93,6.49,1.34,6.69,6.1.23,3.74,10.56,5.35,10.7,2.42-4.42,1.96,4.97,1.34,4.01.39.61,3.43.67,4.01,1.34.39.44,1.25,3.72,1.34,4.01.71,2.33-.33,5.38,0,8.02s2.41,8.32,2.67,9.36c.47,1.87.81,2.68,1.34,4.01.13.33-.02.98,0,1.34.06,1.24,1.24.98,1.34,1.34,1.8,6.59,1.22,11.82,2.67,16.05,3.24,2.11,5.91,4.78,8.02,8.02-2.71,3.08-5.31,6.28-8.02,9.36-12.69-1.55-18.06,4.22-30.76,2.67-1.33-6.38.11-1.45-4.01-4.01-.02,1.77-2.34,1.22-2.67,2.67,2.05.32.2,2.2-1.34,2.67-2.8,1.37-.93,4.08-4.01,4.01-.85-.02-3.72-5.16-4.01-5.35,1.5-.14,2.08,1.94,2.68,0-2.19-2.94-3.11.98-5.35,2.67-1.42,1.08-7.7,2.67-4.01,4.01,0-.74,1.08-1.52,1.34,0-1.68.55-2.73,1.73-2.67,4.01-1.91-.52-2.26-1.61-4.01,0-1.47,7-9.87,7.06-8.02,17.38-.74,2.73-4.02,2.62-1.34,5.35,7.39-.63,17.94,1.89,21.4-2.67.14,1.5-1.94,2.08,0,2.67.65-1.58,1.43-3.03,2.68-4.01,1.04-.52,6.17,2.25,6.69,2.67.45.37-.42,2.26,0,2.67,1.02.99,3.38-.09,4.01,1.34.37.84-1.69,2.29-1.34,2.67.63.68,3.3.8,4.01,1.34,2.35,1.8,2.62,5.13,4.01,6.69,1.06,1.18,2.97,1.4,4.01,2.67-.34-.41-1.78.48-1.34,1.34.5.97,2.18-.21,2.68,1.34.45,1.4-2.18,2.16,0,2.67,1.55-3.82,1.55,2.16,0,2.67,1.78,1.34,4.98,1.26,5.35,4.01-1.5.14-2.08-1.94-2.67,0,5.61,1.07,6.21,7.16,12.03,8.02-2.4,4.56.07.68,1.34,1.34.66.34-.25,3.2,1.34,2.67,0-.74,1.08-1.52,1.34,0h10.7c-.79,3.9,1.98.98,2.67,1.34.66.34-.25,3.2,1.34,2.67,1.18-3.62,4.8,1.06,5.35,1.34,1.85.95,3.63.3,5.35,1.34.37.22.69.92,1.34,1.34,2.51,1.6,4.38,3.06,5.35,5.35-1.5.14-2.08-1.94-2.68,0,1.78,1.34,4.98,1.27,5.35,4.01-1.21,2.74,1.2,4.8,1.34,6.69.17,2.33-1.71,3.41,0,4.01,2.51.32-.76,2.2,1.34,2.67,1.22-2.83,5.37,2.52,9.36,1.34.98,3.92,4.06,5.75,6.69,8.02.18,3.3.2,6.44-1.34,8.02,1.78,1.34,4.98,1.27,5.35,4.01.46,2.4,2.94,4.02,2.67,5.35-.06.29-2.12.01-1.34,1.34.64,1.09,2.31-.46,2.67,1.34,1.41,6.82-2.29,17.15,4.01,21.4-1.5.14-2.08-1.94-2.67,0,1.78,1.34,4.98,1.27,5.35,4.01-3.4,2.42,1.1-1.16,1.34,1.34.17,1.83-.72,2.95-1.34,4.01,1.28.23,2.33.12,2.67,1.34,1.06,3.76.09,10.61,2.68,13.37-5.69,5.57-.71,11.49,0,18.72.52,5.26,1.09,7.89,1.34,10.7.51,5.75-.77,9.14-1.34,12.03-.48,2.48.46,4.96-1.34,6.69,5.75-.63,1.71,3.6,4.01,4.01,0-.74,1.08-1.52,1.34,0,.22,1.95,1.47,1.49,0,2.67-.97.78,1.37,1.39,1.34,1.34-.27-.46.62-1.4,1.34,0,.11.21,1.44,1.72,1.34,1.34.26.99-1.68,2.1-1.34,2.67.94,1.59,6.15.14,4.01,4.01-1.75-.35-1.75,2.31-2.68,2.67-1.29.52-1.09-1.29-1.34-1.34-2.72-.54,2.49,4.66-4.01,2.67v8.02c-.55.34-2.23-.45-2.67,0-.57.57,1.07,1.94,1.34,1.34-.34.77-1.3,2.02-1.34,2.67-.06,1.09-1.31,1.16-1.34,1.34-.7,3.72.53,9.92,0,14.71,1.71.07,1.85,1.71,4.01,1.34-2.06,4.35-.41.99,1.34,1.34-.58-.12,1.35,1.34,1.34,1.34,1.22.83,1.64,1.72,2.67,2.67,1.06-3.08,6.29.69,6.69,2.67v5.35c3.91,1.21.39,4.92,2.67,5.35,3.55-3.38-.18,7.5,4.01,2.67v4.01c-2.39-.38-2.74,1.26-1.34,1.34-.66,2.59-2.23.16-2.68,0-3.73-1.36-.9,0-1.34-2.67-.07-.42-2.08.65-1.34,1.34-.14-.55-2.62-2.37-2.68-2.68-.21-1.27,1.27-2.9,1.34-2.67-.41-1.35-3.98-1.95-1.34-2.67.96-3.63-6.31.96-5.35-2.67-2.52,1-.93,3.99-4.01,1.34,3.1,1.96-2.61,4.24,2.68,4.01-.93,6.73,2.27,9.32,1.34,16.05-3.15.42-.79,6.34-5.35,5.35,2.27,5.06-2.68,12.96-8.02,14.71.26,1.52,1.34.74,1.34,0,3.13,1.52-3.43,4.47-5.35,5.35.25,1.52,1.34.74,1.34,0,2.59.89-.83,1.8-1.34,2.67-.16.27-.3,3.71-1.34,2.67-.05-.05-.05-1.27,0-1.34.01-.01-2.31.56-1.34,1.34.05.04,1.33.02,1.34,0-.05.14-1.25,2.64-1.34,2.67-.45.21-2.4-.37-2.67,0-.61.82,1.03,1.7,1.34,1.34-2.28,2.67-2.37,2.91-5.35,4.01.25,1.52,1.34.74,1.34,0,2.83.41-1.4,4.65-4.01,4.01,3.16,3.04-1.67,4-4.01,5.35.25,1.52,1.34.74,1.34,0,1.87.47.11,2.35-1.34,2.67.21,1.79-2.1,7.37-2.68,4.01-1.52.26-.74,1.34,0,1.34-.25,1.08-1.34,1.33-1.34,2.67-3.1-.43-5.13.22-6.69,1.34,2.09,4.78-1.68,8.49,5.35,9.36-2.31.81.1,6.34-2.68,6.69.59,3.1,5.9,3.58,2.68,6.69,1.96-1.92,4.24-1.03,4.01,2.67,3.1-3.23,3.58,2.09,6.69,2.67-.82,1.19-1.61,2.39,0,2.67-2.55,5.92-3.17,13.77-2.68,22.73-3.06,1.46-11.06,4.89-14.71,1.34-1.59-.52-.68,2.34-1.34,2.67-.81.42-2.54-2.56-2.67,0-.11,2.13-1.41-.01-1.34,0-.71-.13-1.9,3.44-2.67,2.67-1.84-1.84-.57-.1-4.01,1.34,2.36,2.24-1.19,4.48,0,9.36-1.19.59-2.39,1.17-2.67,2.67h-8.02c-1.19.82-2.39,1.61-2.68,0-2.02.04-.11,1.6,0,1.34-1.11,2.65-3.49.63-4.01,4.01-1.94-.59.14-1.17,0-2.67-3.47.81-2.27,3.96-4.01,5.35-1.11.89-1.27-1.34-1.34-1.34-2.97-.07-2.79,4.29-5.35,4.01.47,1.87,2.35.11,2.68-1.34,3.38.66-5.15,6.61-6.69,8.02-3.08,2.82-6.65,5.99-9.36,8.02.75,3.87,2.09-2.05,4.01-1.34,1.6-1.87.52-.11,0,1.34-1.76.47-2.1,2.35-4.01,2.67-1.41,1.39,1.85,7.84,0,9.36-1.69,1.38-.56,2.29-1.34,6.69-.02.11-1.26-.11-1.34,0-.78,1.12.43,4.02,0,5.35-.12.38-1.17.17-1.34,1.34-.35,2.51.58,6,0,8.02-.06.21-2.63-.2-2.67,0-.18.88,1.69,2.24,1.34,4.01.17-.87-1.77,2.98-1.34,2.67-.22.15-3.73.28-2.68,1.34,1.8,1.8-1.52,1.56-4.01,4.01.47,1.87,2.35.11,2.68-1.34,3.18.4-1.49,3.62-2.68,4.01,2.81,3.26-2.12,6.88,2.68,6.69.44,4.9-1.48,7.43-2.68,10.7-10.31,1.97-22.69.12-29.42,5.35-2.48,1.93-3.16,4.22-5.35,6.69,1.54,2.92.27,8.65,4.01,9.36-.59,1.94-1.18-.14-2.68,0,.66,3.8,6.99,1.92,5.35,8.02-1.14.25-1.26-.52-1.34-1.34-2.05.07-.08,1.57,0,1.34-1.32,3.75-4.17,1.55-5.35,5.35,3.68-3.27,2.82,2-1.34,1.34,1.49,1.65,1.17,2.75-1.34,2.67,2.96,1.44-1.18,3.7-1.34,6.69-.04.74-3.42.59-2.68,1.34,1.92,1.92-.12.78-2.68,2.67.25,1.52,1.34.74,1.34,0,3.51.7-1.22,1.73-1.34,4.01-.1,1.93-2.55,2.89-2.68,5.35-.13,2.52-1.87-.89-2.67,0-.67.74.99,1.7,1.34,1.34-5.22,5.44-11.5,6.2-14.71,14.71-6.51,1.52-15.39.65-17.38,6.69-3.87-.75,2.05-2.09,1.34-4.01-4.42,2.26-6.17,7.2-9.36,10.7-3.87-.75,2.05-2.09,1.34-4.01-3.04.97-4,4.02-5.35,6.69-5.49-1.41-8.76,6.29-12.03,2.67-2.02.04-.11,1.61,0,1.34-1.11,2.65-3.49.63-4.01,4.01-1.94-.59.14-1.17,0-2.67-1.65.58-2.75,1.71-2.67,4.01-15.68-.52-24.36,5.95-30.76,14.71-4.6-.59-7.49.53-10.7,1.34,1.49,1.65,1.16,2.75-1.34,2.67,1.55,1.98,1.55,3.37,0,5.35-3.31.17-2.27-1.64,0-1.34-3.34-3.35-9.96-3.41-8.02-12.03-8.07,3.12-16.95,2.92-29.42,4.01-.25-1.14.52-1.26,1.34-1.34.26-1.6-2.01-.67-2.67-1.34-1.11-1.12-.1-4.35-2.67-4.01-10.3.4-11.88,9.52-18.72,13.37-.71-.57-5.05-.66-5.35-1.34.22.5,1.54-4.22,0-2.67-1.38,1.38-.88.6-1.34-1.34-1.26-.84-1.11,1.29-1.34,1.34-1.47.29-.98-1.32-1.34-1.34-2.23-.09-7.04,1.21-8.02-1.34-.43.77-3.85,1.96-4.01,2.67-.17.76.99,1.89,1.34,1.34-1.45,2.28-3.39,2.23-4.01,5.35-1.09.87-1.32-1.34-1.34-1.34-1.3.1-1.99,4.22-2.67,1.34-1.39.12.14,4.71,0,9.36,0,.03-1.57.2-1.34,1.34,0-.02,2.14.35,1.34,1.34-1.33.01-2.67.01-4.01.01l-.08.1ZM1055.04,326.28c-6.04,6.04,2.45-.86,2.67-2.67-1.73.05-1.78,1.78-2.67,2.67ZM873.19,965.44c-6.04,6.04,2.45-.86,2.67-2.67-1.73.05-1.78,1.78-2.67,2.67ZM259.43,687.31c-5.52-2.82.51,1.76,1.34,4.01,5.52,2.82-.51-1.76-1.34-4.01ZM1033.65,343.66c2.63,1.74,14.44.4,12.03-2.67h-12.03c0-.74-1.08-1.52-1.34,0,1.67-.65.77,2.3,1.34,2.67Z"/>
+        <path class="cls-1" d="M698.02,962.77c-.38,2.36,6.3,4.14,2.68,5.35-.59-1.19-1.18-2.39-2.68-2.67-.82-1.19-1.61-2.39,0-2.67h0Z"/>
+        <path class="cls-1" d="M49.49,127.04c-2.89,3.23-5.21-1.77-5.35-2.67-.48-3.14,3.82-3.06,4.01-6.69.02-.45,2-3.35,2.67-2.67,2.63,2.64-.51-1.83,4.01-1.34,1.88,4.42-2.7,4.35-4.01,6.69,0,0-1.38,3.56-1.34,2.67-.07,1.39,2.26,2.11,0,2.67-3.07.77-2.45.93,0,1.34h.01Z"/>
+        <path class="cls-1" d="M8.04,121.69c1.52.25.74,1.34,0,1.34.02,3.14,3.65-.49,2.67-2.67,3.55-1.33.04,6.87,0,5.35.03,1.15,1.79,2.08,1.34,4.01-.06.26-2.5-.26-2.68,0-1.53,2.29.39,6.8-1.34,8.02-3.69.57-1.26-4.99-4.01-5.35-.72-1.44,2.58-.86,2.67-1.34.28-1.5-2.29-4.2-1.34-6.69.35-.91,3.04-.93,2.67-2.67h.02Z"/>
+        <path class="cls-1" d="M58.85,280.81c.82-1.19,1.62-2.39,0-2.67.83-2.3,6.85-.47,8.02,0,.15.06-.11,1.2,0,1.34.37.45,2.24-.41,2.68,0,.7.65-.87,1.63-1.34,1.34-.05-.03.56,2.31,1.34,1.34.04-.05.01-1.33,0-1.34.14.06,2.64,1.25,2.67,1.34.23.52-.4,2.31,0,2.67,1,.91,3.18.29,4.01,1.34.67.85-.42,2.94,1.34,2.67.82-1.94,1.62.14,0,0,.58,1.65,1.71,2.75,4.01,2.67-1.36,1.76-1.31,4.93-2.67,6.69h-5.35c1.77-5.29-5.29.43-1.34-2.67.33-2.54-2.92-1.56-4.01-2.67-.42-.42.45-2.3,0-2.67-1-.82-2.53-1-2.67-2.67-1.35-1.74-.64-.42,0,0,5.56,3.6-1.61-4.01-6.68-6.69v-.02Z"/>
+        <path class="cls-1" d="M377.1,279.47c9.22.14,20.12-1.4,24.07,4.01-1.74,3.32,3.92,5.55,0,6.69.76,1.25,3.72,2.03,1.34,2.67.83,5.06,5.59-1.45,5.35,5.35,2.43-2.46,4.21,1.21,6.69,1.34,1.55.08,1.93,1.71,2.67,2.67,1.05,1.36,2.87,1.48,4.01,2.67.87.91-1.29,1.52-1.34,1.34,1.01,3.68,6.39,7.33,9.36,8.02-1.18,6.44.72,21.45-5.35,26.74v14.71c-1.18,2.9-3.86-1.53-4.01,1.34-.14,2.64-6.11.8-12.04,1.34.49-2.27-1.04-2.52-2.67-2.67-.43-3.1.21-5.13,1.34-6.69.2-1.78-4.5-4.98-1.34-5.35-.98-1.25-2.43-2.02-4.01-2.67-.84-1.25,1.3-1.12,1.34-1.34.3-1.67-2.11-.59-2.67-1.34-.68-.9,1.16-1.63,1.34-1.34-.85-1.43-2.49-2.05-4.01-4.01-1.03-1.27,1.31-2.74,1.34-2.67-.34-.95-2-1.79-2.67-2.67-1.25-1.62-3.56-5.37-5.35-8.02-1.85-.04-.32,1.71,0,1.34-2.36,2.72-3.11-3.31-6.69-4.01.25-1.52,1.34-.74,1.34,0,3.62-1.2-2.82-3.23-4.01-4.01-.87-1.09,1.34-1.32,1.34-1.34-.14-1.86-2.14-.29-2.67-1.34.34.66,1.33-4.33,0-2.67-1.41,1.76-.74.27-1.34-1.34.05.14-2.29-.58-1.34-1.34,3.57-2.86-3.19.12,0-2.67.04-1.85-1.71-.32-1.34,0-4.22-3.69-4.35-12.47-2.67-18.72,3.6.92,2.59-2.76,4.01-4.01h-.02Z"/>
+        <path class="cls-1" d="M425.24,311.57c-1.17-2.33-4.48-5.15,0-2.67,1.17,2.33,4.47,5.15,0,2.67Z"/>
+        <path class="cls-1" d="M155.13,383.77c-5.43.53-9.32-.49-9.36-5.35,4.3.82,1.83-4.81,5.35-1.34-.14-3.43-4.98-2.15-6.69-4.01.07-8.02,1.7-9.78,4.01-17.38-1.23-2.33-4.18-2.95-8.02-2.67,2.45-5.35-1.9-9.67,1.34-14.71-4.31-.29-5.66,2.36-10.7,1.34v-6.69c-.33-1.45-2.66-.91-2.67-2.67-4.23,1.66-.25,3.89-4.01,6.69-4.64-1.07-5.14,1.99-9.36,1.34.24-5.13-1.1-.48-4.01-1.34,1.6-1.53.52-3.39,0-1.34-1.3-.04-1.04-1.64-1.34-2.67-3.27,3.29-5.8-.99-10.7,0,1.6-1.53.52-3.39,0-1.34-2.6,0-.16-3.1,0-2.67-.22-.58-2.69-1.46-2.68-1.34-.35-2.89,1.03-5.12,1.34-9.36,1.96-3.1,4.24,2.61,4.01-2.67,5.72-1.15,11.84,1.2,18.72,1.34,2.97.06,5.79-1.95,12.03-1.34,2.92.29,4.35.82,9.36,1.34,2.12.22,2.29,1.06,4.01,1.34,2.17.35,4.52-.39,6.69,0,.11.02-.11,1.26,0,1.34,1.13.79,4.01-.45,5.35,0,.48.16,0,1.27,1.34,1.34-.2-.01,3.9,1.22,4.01,1.34-.53-.55-1.44,4.12,0,2.67,1.82-1.82.89.17,2.67,2.67,1.52-.25.74-1.34,0-1.34.47-1.87,2.35-.11,2.67,1.34,2.27-.49,2.52,1.04,2.67,2.67,1.72-1.84,5.81-1.32,9.36-1.34-1.05,4.77,5.13,4.44,6.69,6.69,1.06,1.53.73,3.47,1.34,5.35.15.46,1.27.02,1.34,1.34.07,1.33-.06,2.68,0,4.01-4.78-3.37.54,2.56-4.01,4.01-.79,4.11,1.68,10.7-1.34,10.7-1.98,0,.53,2.15,0,4.01-1.53-1.61-3.39-.52-1.34,0,.1.85-1.8,1.59-2.67,1.34-1.31-.39-.66-2.44-1.34-2.67-1.08-.37-2.61,2.02-4.01,0,1.49-1.65,1.16-2.75-1.34-2.68.73-1.94-1.56-6.91,1.34-6.69-.57-2.1-4.21-1.14-2.67-5.35-6.32.34-8.79.55-12.03,1.34,1.54,1.58,1.51,4.73,1.34,8.02-2.65,3.42,1.39,6.15-1.34,8.02.32,3.45,2.3-1.84,5.35,0,.79,4.05-1.11,6.1-2.67,8.02-1.19,1.46-4.39,4.02-6.69,4.01,2.68,2.73-.6,2.62-1.34,5.35h0Z"/>
+        <path class="cls-1" d="M1032.31,340.98c.25-1.52,1.34-.74,1.34,0h12.03c2.41,3.07-9.41,4.42-12.03,2.67-.57-.37.33-3.32-1.34-2.67Z"/>
+        <path class="cls-1" d="M397.16,371.74c.25,1.52,1.34.74,1.34,0,2.8,2.99.93,7.59,1.34,12.03.32,3.41,2.35,7.12,1.34,10.7-3.9-2.36-1.54,4.1-2.67,2.67-1.28-1.6-.96-.44-1.34,1.34-6.64,1.35-4.26-4.23-9.36-1.34-1.15-4.64-5.28-6.3-8.02-9.36-.97-4.09,2.31-3.93,1.34-8.02,4.27.79,4.14-2.51,6.69,0-2.68-2.66.42-6.37,1.34-9.36h9.36c.24,1.14-.52,1.26-1.34,1.34h-.02Z"/>
+        <path class="cls-1" d="M197.92,386.45c3.16,1.79-4.32,9.37-9.36,9.36,1-6.06-3.67,2.04-2.67-4.01-3.78-1.11-3.46,1.89-6.69,1.34-3.58-5.98,1.95-12.76,9.36-12.04.25,1.14-.52,1.26-1.34,1.34.16,2.23,5.7-1.52,5.35,4.01.99-.4,2.28-1.8,4.01-1.34.24.07.79,4.53,1.34,1.34h0Z"/>
+        <path class="cls-1" d="M65.54,498.77c2.35.32,3.05,2.3,2.67,5.35-2.79.57-3.54-.92-5.35-1.34.47-1.76,2.35-2.1,2.67-4.01h.01Z"/>
+        <path class="cls-1" d="M161.82,542.9c-5.92,4.98,2.35,10.68,1.34,18.72-.77-.35-1.8,3.55-2.67,2.67-.05-.05-.05-1.27,0-1.34,0-.02-2.31.56-1.34,1.34,1.74,1.39.09.39,0,1.34-.44,4.58-.32,14.14,1.34,18.72-3.17-.89-1.73,1.93-1.34,1.34-.8,1.22-3.69,2.05-1.34,2.67-.04,1.3-1.64,1.04-2.68,1.34,1.55,1.98,1.55,3.37,0,5.35-2.89-.34-6.15-7.08-4.01-8.02-.76-1.25-3.72-2.03-1.34-2.67-.98-1.25-2.43-2.03-4.01-2.67.04-1.3,1.64-1.04,2.67-1.34-3.39-3.27,2.82-2.85,1.34-8.02-.25-1.08-1.34-1.33-1.34-2.67-4.2,1.48-3.53.63-6.69-1.34.73-5.33.73-4.03,0-9.36,1.53,1.6,3.39.52,1.34,0,.2-3.33,4.53-.39,5.35,0-.39-.82-3.33-5.15,0-5.35-.88-1.79-2.86-2.49-5.35-2.67.82-1.19,1.62-2.39,0-2.68,2.39-2.93,4.77-4.89,6.69-2.67,1.16,1.03.82-8.54,2.67-6.69,1.38,1.38,1.19.29,0,0,3.57-5.47,3.88,2.81,5.35,1.34,2.59-2.59-.05,2.16,4.01,2.67h0Z"/>
+        <path class="cls-1" d="M256.76,695.33c-3.13,1.52,3.42,4.47,5.35,5.35.85,2.92-3.79,4.25,1.34,4.01.18,3.3.2,6.44-1.34,8.02-1.52-.25-.74-1.34,0-1.34-1.42-4.21-2.32,2.43-2.67,2.67-1.22.85-1.14-1.3-1.34-1.34-1.07-.2-2.16,4.44-2.67,1.34-1.98.58.53,1.71,0,4.01-1.25.98-2.02,2.43-2.67,4.01-5.21,1.55-10.45-3.69-6.69-6.69.46-1.6-2.26-.66-2.68-1.34-.98-1.58,1.37-2.42,1.34-2.67-.24-1.68-2.35-.42-2.68-1.34-.53-1.51,1.34-2.66,1.34-2.67-.55-1.74-1.8-.23-2.68-1.34-.43-.54-1.32-5.17-1.34-5.35-.18-2.56,1.89-2.88,1.34-6.69,6.39-.73,13.73-2.25,17.38,2.68,3.87,1.5-2.05,1.44-1.34-1.34v.02Z"/>
+        <path class="cls-1" d="M262.1,723.41c.88-4.01,2.86,1.28,5.35,0v22.73c-1.64.58-2.75,1.71-2.67,4.01-1.65-1.49-2.75-1.16-2.67,1.34h-8.02c.48-4.73-.69-8.13,0-10.7.07-.26,2.44.23,2.67,0,.75-.75-1.1-1.69-1.34-1.34,2.1-3.09,4.05-3.33,6.69-5.35-.89-1.63-4.55.72,0-4.01,0-1.96-1.66-.19-1.34,0-3.77-2.19.27-2.96,0-6.69,1.08.25,1.33,1.34,2.67,1.34.24-1.14-.52-1.26-1.34-1.33Z"/>
+        <path class="cls-1" d="M676.62,811.67c-1.94.59.14,1.17,0,2.67,1.94.59-.14,1.17,0,2.67-4.2.63-3.53-3.6-6.69-4.01-3.15-.03-3.94,2.3-8.02,1.34-.04-1.3-1.64-1.04-2.67-1.34-.21-3.33,2.1-4.14,2.67-6.69,1.2-.86,1.18,1.32,1.34,1.34,1.42.14,1.76-3.55,2.67-4.01,1.09-.55,3.1.98,2.68,1.34,1.32-1.09.58-3.74,2.67-5.35,3.54-.56,2.89,5.71,4.01,9.36.28.92,4.2,5.67,1.33,2.68h.01Z"/>
+      </g>
+    </g>
+  </svg>`,
+  asia: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="continent" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 437.59 441.51">
+    <defs>
+      <style>
+      .cls-1 {
+        fill: #ccc;
+        stroke: #fff;
+        stroke-width: .5px;
+      }
+      .dinosaurs {
+        fill: #2a7035;
+        cursor: pointer;
+      }
+      .dinosaurs:hover {
+        fill: #194520;
+      }
+      </style>
+    </defs>
+    <g id="continent" data-name="Layer 1">
+      <g>
+        <path id="AE" class="cls-1" d="M48.82,342.55l.5-.15.11.84,2.19-.48,2.32.08,1.69.09,1.92-2.07,2.1-1.98,1.77-1.9.53,1.05.38,2.44h-1.43s-.23,2.01-.23,2.01l.5.42-1.27.6v1.25s-.83,1.26-.83,1.26l-.07,1.21-.57.64-8.42-1.52-1.08-3.08-.11-.71Z"/>
+        <path id="AF" class="cls-1" d="M75.83,305.73l2.86,1.3,2.11-.46.59-1.55,2.21-.52,1.58-1.05.56-2.79,2.36-.68.44-1.25,1.33.94.84.11,1.56.03,2.12.74.85.42,2.03-1.12.95.67.9-1.6,1.68.07.43-.52.3-1.43,1.21-1.23,1.51.8-.3,1.09.85.17-.27,2.95,1.11,1.15.98-.74,1.25-.34,1.74-1.57,1.93.26h2.9l.5,1.01-1.64.39-1.42.65-3.22.4-3.01.73-1.64,1.51.66,1.46.33,1.7-1.4,1.43.12,1.3-.77,1.22-2.67-.11,1.1,2.22-1.78.85-1.19,2,.15,1.98-1.1.92-1.03-.3-2.15.43-.3.91h-2.09l-1.56,1.84-.1,2.75-3.65,1.33-1.95-.28-.57.7-1.67-.4-2.81.48-4.69-1.64,2.54-2.93-.23-2.1-2.12-.55-.22-2.09-.92-2.64,1.2-1.83-1.22-.49.77-2.45,1.14-4.25Z"/>
+        <path id="BD" class="cls-1" d="M164.04,349.24l-.05,2.15-.98-.46.18,2.41-.8-1.56-.16-1.52-.54-1.45-1.17-1.76-2.58-.12.26,1.25-.88,1.67-1.2-.61-.41.55-.79-.33-1.08-.27-.44-2.48-.97-2.28.47-1.84-1.72-.82.62-1.12,1.75-1.15-2.02-1.63.99-2.11,2.22,1.34,1.34.16.25,2.15,2.66.42,2.61-.05,1.61.53-1.29,2.59-1.26.18-.86,1.73,1.53,1.58.46-1.94h.78s1.47,4.79,1.47,4.79Z"/>
+        <path id="BN" class="cls-1" d="M224.41,399.6l1.11-1.05,2.39-1.53-.13,1.38-.16,1.78-1.34-.09-.59.95-1.28-1.44Z"/>
+        <path id="BT" class="cls-1" d="M161.31,331.61l1.14,1-.2,1.93-2.29.09-2.36-.21-1.77.49-2.55-1.19-.05-.63,1.85-2.34,1.51-.8,2.01.73,1.48.08,1.23.85Z"/>
+        <path id="CN" class="cls-1 dinosaurs" d="M214.83,355.1l.63,1.13-1.23,1.3-.65,1.7-2.42,1.41-2.3-.91-.08-2.53,1.38-1.34,3.06-.83,1.61.07ZM278.16,258.43l-2.43,1.65h-4.26l-1.13-3.95-3.32-3.03-4.88-1.38-1.04-4.28-.98-2.73-1.05-1.94-1.73-4.61-2.46-1.71-4.2-1.39-3.72.13-3.48.84-2.32,2.31,1.54,1.1.04,2.52-1.56,1.45-2.53,4.73.03,1.93-3.95,2.74-3.37-1.63-1.37,3.25-1.98,4.23.72,1.71,1.59-.53,2.77.65,2.16-1.54,2.25,1.33,2.54,2.89-.31,1.45-2.21-.46-4.07.54-1.97,1.16-2.05,2.66-4.28,1.55-2.79,2.1-2.88-.8-1.58-.36-1.47,2.54.9,1.5.45,1.28-1.96,1.3-2.01,2.05-3.28,1.34-4.2.14-4.53,1.31-3.26,2.01-1.24-1.16h-3.39l-4.15-2.29-2.77-.57-3.73.53-5.79-.85-3.09.09-1.65-2.27-1.28-3.57-1.73-.43-3.39-2.45-3.78-.55-3.33-.68-1.01-1.73,1.08-4.73-1.93-3.32-4-1.57-2.36-2.23-.74-2.97-1.1.35-2.13,2.83-2.33.39-.13,4.19-1.56,1.86-5.56-1.35-2.02,7.26-1.44.89-5.55,1.58,2.52,6.75-1.92,1,.22,2.17-.39.85-4.42,2.03-1,1.48-3.6.44-1.06,2.35-2.97-.49-1.94.72-2.68,1.73.39.85-.8.83.71,3.32.92-.36,1.7.81-.1,1.38.42,2.01.5,1.01,2.07,1.63.83,2.66,4.42,1.33.14-.04,3.88-1.37,2.88,1.37-.91,2.76-1.6,2.23-1.24.1h-.05l-.18,1.74,1.12,1.71-.09,1.69-2.01-.44.79,3.63,2.75,2.07,3.9,2.26,1.16-.77,2.25.99,2.83,2.09,1.58.46.94,1.53,2.18.63,2.28,1.39,3.17.73,3.27.31,1.71-.66.24,2.48,1.85-2.34,1.51-.8,2,.73,1.48.08,1.23.85,2.26-.39,2.55-2.36,3.23-2.03,2.35.78,2-1.35,1.31,1.99-.95,1.33,3.02.47,1.64-.24.94,1.86,1.22.75.08,2.4-.11,2.57-2.66,2.58-.34,3.63,2.96-.51.67,2.8,1.78.59-.82,2.51,2.08,1.13,1.21.55,2.06-.87.08,1.24.25.7,1.5.08-.42-3.43,1.45-.44,1.5-.74,2.24.02,2.73-.35,2.39-1.62,1.35,1.14,2.56.55-.44,1.74,1.33,1.22,2.82.78,1.33-.49,3.76.96-.66,1.16.74,2.16,1.55-.17.96-3.15,2.97-.46,3.92-1.5,1.59-1.5.97.98,1.71-1.34,3.16-.35,3.9-2.55,3.86-2.82,2.6-3.68,2.27-4.09,2.05-3.41,1.57-.28.71-2.52.43-2.61-1.65-1-.67-1.73,1.76-.89.05-2.43-1.9-2.53-1.71-3.05-1.1-3.31-3.02-1.86,1.44-2.39,2.73-1.73,1.31-1.87,3.97-.97-.46-1.84-1.81-.09-2.49-1.37-3.14,2.51-2.22-1.02-.09-1.58-2.29-.58-1.48-2.41,1.43-1.68,2.75-.17,1.73-2.34,3.17-2.54,2.44-1.3,1.48,1.93-2.22,2.45.59,1.41-1.49,1.67,3.02-.98,2.06-1.69,3.92-1.06,2.28-2.35,3.09-1.98,1.93-2.64,1.33,1.17,2.42.14-.44-1.97,4.33-1.62,1.11-2.13,1.81,2.24-.02-1.93,1.43-.1.4-4.55-.74-3.36,2.41-1.4,3.4.7,1.89-3.89.96-4.46,1.09-1.51,1.47-3.76-4.63,1.21Z"/>
+        <path id="ID" class="cls-1" d="M242.67,440.89l-1.18.05-3.72-1.98,2.61-.56,1.47.86.98.86-.16.77ZM253.1,440.61l-2.4.62-.34-.34.25-.96,1.21-1.72,2.77-1.12.28.56.05.86-1.82,2.1ZM234.78,434.84l1.01.75,1.73-.23.7,1.2-3.24.57-1.94.38-1.51-.02.96-1.62,1.54-.02.75-1.01ZM248.81,434.83l-.41,1.56-4.21.8-3.73-.35v-1.03s2.22-.59,2.22-.59l1.76.84,1.87-.21,2.5-1.02ZM208.77,431.14l5.37.28.62-1.16,5.2,1.35,1.02,1.82,4.21.51,3.44,1.67-3.2,1.07-3.08-1.13-2.54.08-2.91-.21-2.62-.51-3.25-1.07-2.06-.28-1.17.35-5.11-1.16-.49-1.21-2.57-.21,1.92-2.68,3.4.17,2.26,1.09,1.16.21.4,1.02ZM281.95,429.56l-1.44,1.91-.27-2.11.5-1.01.59-.95.64.82-.02,1.34ZM260.99,421.85l-1.05.93-1.94-.51-.55-1.2,2.84-.13.7.91ZM270.03,420.84l1.02,2.13-2.37-1.15-2.34-.23-1.58.18-1.94-.1.67-1.53,3.46-.12,3.08.82ZM280.32,415.42l.78,4.51,2.9,1.67,2.34-2.96,3.22-1.68h2.49l2.4.97,2.08,1,3.01.53.05,9.1.05,9.16-2.5-2.31-2.85-.57-.69.8-3.55.09,1.19-2.29,1.77-.78-.73-3.05-1.35-2.35-5.44-2.37-2.31-.23-4.21-2.58-.83,1.36-1.08.25-.64-1.02v-1.21s-2.15-1.37-2.15-1.37l3.02-1,2,.05-.24-.74h-4.1s-1.11-1.67-1.11-1.67l-2.5-.51-1.19-1.38,3.78-.67,1.44-.91,4.5,1.14.45,1.02ZM255.36,408.26l-2.25,2.76-2.11.54-2.7-.54-4.67.14-2.45.4-.4,2.11,2.51,2.48,1.51-1.26,5.23-.95-.23,1.28-1.22-.4-1.22,1.63-2.47,1.08,2.65,3.57-.51.96,2.52,3.22-.02,1.84-1.5.82-1.1-.98,1.36-2.29-2.75,1.08-.7-.77.36-1.08-2.02-1.64.21-2.72-1.87.85.24,3.25.11,4-1.78.41-1.2-.82.8-2.57-.43-2.69-1.18-.02-.87-1.91,1.16-1.83.4-2.21,1.41-4.2.59-1.15,2.38-2.07,2.19.82,3.54.39,3.22-.12,2.77-2.02.49.61ZM265.03,409.06l-.15,2.43-1.45-.27-.43,1.69,1.16,1.47-.79.33-1.13-1.76-.83-3.56.56-2.23.93-1.01.2,1.52,1.66.24.27,1.15ZM234.71,407.12l3.14,2.58-3.32.33-.94,1.9.12,2.52-2.7,1.91-.06,2.77-1.08,4.27-.41-.99-3.19,1.26-1.11-1.71-2-.16-1.4-.89-3.33,1-1.02-1.35-1.84.15-2.31-.32-.43-3.74-1.4-.77-1.35-2.38-.39-2.44.33-2.58,1.67-1.85.47,1.86,1.92,1.57,1.81-.57,1.79.2,1.63-1.41,1.34-.24,2.65.78,2.29-.59,1.44-3.88,1.08-.97.97-3.17h3.22l2.43.47-1.59,2.52,2.06,2.64-.49,1.28ZM200.9,428.54l-3.1.06-2.36-2.34-3.6-2.28-1.2-1.69-2.12-2.27-1.39-2.09-2.13-3.9-2.46-2.32-.82-2.39-1.03-2.17-2.53-1.75-1.47-2.39-2.11-1.56-2.92-3.08-.25-1.42,1.81.11,4.34.54,2.48,2.73,2.17,1.89,1.55,1.16,2.66,3,2.85.04,2.36,1.91,1.62,2.33,2.13,1.27-1.12,2.27,1.61.97,1.01.07.48,1.94.98,1.56,2.06.25,1.36,1.76-.7,3.47-.16,4.32Z"/>
+        <path id="IL" class="cls-1" d="M4.36,315.65l-.49,1.05-1.02-.46-.58,2.2.7.36-.71.46-.13.86,1.32-.45.07,1.27-1.4,5.17-1.84-5.55.81-1.08-.19-.19.74-1.53.57-2.5.4-.84.08-.03h.94l.26-.58.75-.05.04,1.37-.38.5.06.02Z"/>
+        <path id="IN" class="cls-1 dinosaurs" d="M177.31,331.26l-.22-1.2-3.02-.47.95-1.34-1.32-1.98-2,1.34-2.35-.78-3.23,2.03-2.55,2.35-2.26.4,1.14,1-.2,1.93-2.29.09-2.36-.21-1.77.49-2.55-1.19-.05-.63-.24-2.48-1.71.67-.22,1.35.37,1.99-.32,1.24-2.33.05-3.38-.73-2.16-.29-1.62-1.59-3.84-.41-3.66-1.77-2.64-1.55-2.72-1.2,1.09-2.99,1.78-1.46-3.9-2.27-2.75-2.06-.79-3.63,2.01.44.09-1.69-1.11-1.71.18-1.74h.05l1.24-.1,1.6-2.23.91-2.76-2.88-1.37-3.88,1.37-.14.04.19.06-4.61-1.39-.83-2.66-2.07-1.63-1.64.39-1.42.65-1.28.16-.12.11-3.08,2.34.36,1.15h.96s3.04,2.6,3.04,2.6l-1.87,1.68.58,4.8,2.38.99.09-.03v.04s2.28,1.63,2.28,1.63l-2.39,1.91.04,2.33-2.72,3.24-1.76,3.26-2.93,3.32-3.26-.24-3.09,3.3,1.84,1.4.32,2.39,1.57,1.56.56,2.64h-6.17s-1.87,2.03-1.87,2.03l3.29,2.58.83,1.18-1.35,1.09,3.66,3.63,1.98.36,4.08-1.79.54,2.8v3.58s.83,3.77.83,3.77l1.16,5.64,2.55,3.97.49,1.8.69,3.58,1.49,2.74.98,1.34,1.08,2.86,1.3,3.96,2.65,2.63,1.13-.81.94-1.92,2.56-.8-.85-.93,1.27-2.15,1.45-.14.02-4.83,1.18-2.71-.15-2.37-.58-3.75.84-2.2,1.31-.16,2.53-1.03,1.4-.72v-1.34l2.79-1.91,2.11-1.85,3.14-3.46,4.03-2,1.5-1.76-.16-2.25,3.46-.62,1.9.04.41-1.1-.44-2.48-.97-2.28.47-1.84-1.72-.82.62-1.12,1.75-1.15-2.02-1.63.99-2.11,2.22,1.34,1.34.16.25,2.15,2.66.42,2.61-.05,1.61.53-1.29,2.59-1.26.18-.86,1.73,1.53,1.58.46-1.94h.78s1.47,4.79,1.47,4.79l1.39-.72-.3-1.28.63-1.03.11-3.15,2.19.7,1.25-2.52.15-1.5,1.54-2.6-.08-1.78,3.63-2.16,2,.57-.23-1.93.99-.58Z"/>
+        <path id="IQ" class="cls-1" d="M31.56,304.6l1.83,1.04.22,2-1.42,1.17-.65,2.64,1.95,3.18,3.43,1.82,1.45,2.5-.46,2.37h.89l.03,1.73,1.55,1.69-1.66-.15-1.89-.27-2.06,3.08-5.21-.26-7.9-6.49-4.18-2.29-3.38-.89-1.13-4.04,6.21-3.5,1.06-4.12-.27-2.52,1.54-.86,1.44-2.18,1.2-.55,3.26.46.99.89,1.34-.59,1.82,4.14Z"/>
+        <path id="IR" class="cls-1" d="M55.39,300.36l2.47-.68,1.99-2.02,1.87.1,1.23-.66,2,.33,3.1,1.79,2.24.39,3.2,3.09,2.09.12.25,2.91-1.14,4.25-.77,2.45,1.22.49-1.2,1.83.92,2.64.22,2.09,2.12.55.23,2.1-2.54,2.93,1.38,1.68,1.13,1.93,2.68,1.4.08,2.77,1.34.51.23,1.44-4.04,1.61-1.06,3.6-5.27-.93-3.06-.71-3.16-.41-1.2-3.81-1.34-.56-2.16.56-2.82,1.51-3.43-1.03-2.83-2.41-2.7-.9-1.87-3.01-2.07-4.27-1.51.52-1.78-1.07-1.05,1.26-1.55-1.69-.03-1.73h-.89l.46-2.37-1.45-2.5-3.43-1.82-1.95-3.18.65-2.64,1.42-1.17-.22-2-1.83-1.04-1.82-4.14-1.53-2.83.54-1.09-.87-4.12,1.92-1.03.44,1.37,1.42,1.66,1.92.47,1.02-.1,3.31-2.66,1.05-.27.82,1.07-.96,1.78,1.75,1.86.69-.17.89,2.61,2.66.73,1.95,1.76,3.98.6,4.38-.92.27-.82Z"/>
+        <path id="JO" class="cls-1" d="M3.87,316.7l.49-1.05,3.12,1.32,5.49-3.54,1.13,4.04-.53.49-5.62,1.65,2.8,3.26-.93.54-.46,1.09-2.14.44-.67,1.16-1.22.98-3.12-.51-.09-.46,1.4-5.17-.07-1.27.42-.96v-2.01Z"/>
+        <path id="JP" class="cls-1 dinosaurs" d="M281.71,310.84l.36,1.15-1.58,2.03-1.15-1.07-1.44.78-.74,1.95-1.83-.95.02-1.58,1.55-2,1.59.39,1.15-1.42,2.07.72ZM299.48,300.56l-1.06,2.78.49,1.73-1.46,2.42-3.58,1.6-4.93.21-4,3.84-1.88-1.29-.11-2.52-4.88.75-3.32,1.59-3.28.06,2.84,2.46-1.87,5.61-1.81,1.37-1.36-1.27.69-2.96-1.77-.96-1.14-2.28,2.65-1.03,1.47-2.11,2.82-1.75,2.06-2.33,5.58-1.02,3,.7,2.93-6.17,1.87,1.67,4.11-3.51,1.59-1.38,1.76-4.38-.48-4.1,1.18-2.33,2.98-.68,1.53,5.11-.08,2.94-2.59,3.6.05,3.63ZM307.71,274.63l1.97.83,1.98-1.65.62,4.35-4.16,1.05-2.46,3.76-4.41-2.58-1.53,4.12-3.12.06-.39-3.74,1.39-2.94,3-.21.82-5.38.83-3.09,3.29,4.12,2.17,1.3Z"/>
+        <path id="KG" class="cls-1" d="M103.17,281.94l.63-1.66,1.84-.54,4.62,1.31.43-2.24,1.59-.8,4,1.61,1.02-.42,4.65.1,4.16.4,1.4,1.37,1.73.55-.39.86-4.42,2.03-1,1.48-3.6.44-1.06,2.35-2.97-.49-1.93.72-2.68,1.72.39.85-.8.83-5.3.55-3.47-1.17-3.04.28.27-2.1,3.05.61,1.03-1.13,2.13.36,3.59-2.64-3.32-1.96-2,.93-2.07-1.4,2.35-2.43-.83-.37Z"/>
+        <path id="KH" class="cls-1" d="M194.39,382.43l-1.14-1.48-1.41-2.94-.67-3.45,1.8-2.38,3.62-.55,2.63.41,2.31,1.13,1.27-1.99,2.49,1.06.65,1.92-.35,3.42-4.71,2.19,1.23,1.73-2.94.2-2.43,1.14-2.35-.41Z"/>
+        <path id="KP" class="cls-1" d="M270.5,281.45l.39.67-1.06-.23-1.22,1.27-.84,1.28.11,2.67-1.45.81-.5.65-1.06,1.08-1.87.6-1.21.98-.09,1.57-.33.4,1.12.58,1.59,1.58-.41.86-1.19.23-1.98.17-1.09,1.6-1.26-.12-.17.32-1.36-.67-.34.66-.82.29-.1-.66-.73-.32-.75-.57.77-1.57.66-.42-.25-.65.71-1.94-.19-.59-1.63-.4-1.32-.97,2.28-2.35,3.09-1.98,1.93-2.65,1.33,1.17,2.42.14-.44-1.97,4.33-1.63,1.12-2.13,1.81,2.24Z"/>
+        <path id="KR" class="cls-1" d="M264.08,295.36l2.42,4.18.69,2.27.02,3.98-1.05,1.88-2.54.66-2.24,1.41-2.53.29-.31-1.85.52-2.57-1.24-3.6,2.08-.59-1.92-3,.17-.32,1.26.12,1.09-1.6,1.98-.17,1.19-.23.41-.86Z"/>
+        <path id="KW" class="cls-1" d="M38.72,324.59l.58,1.41-.25.73.9,2.41-1.98.08-.7-1.51-2.5-.31,2.06-3.08,1.89.27Z"/>
+        <path id="KZ" class="cls-1 dinosaurs" d="M103.17,281.94l-1.61.7-3.69,2.61-1.23,2.65-1.05.02-.76-1.75-3.57-.12-.57-3.06-1.37-.03.21-3.8-3.35-2.8-4.81.3-3.29.56-2.68-3.5-2.29-1.48-4.35-2.84-.52-.35-7.22,2.35.11,14.13-1.44.18-1.96-2.93-1.9-1.06-3.18.79-1.24,1.25-.16-.92.69-1.57-.53-1.32-3.25-1.3-1.27-3.47-1.54-.98-.1-1.28,2.73.37.11-2.88,2.38-.64,2.45.59.51-3.91-.5-2.51-2.81.2-2.38-1-3.25,1.79-2.61.86-1.43-.66.29-2.1-1.79-2.76-2.08.11-2.38-2.83,1.62-3.22-.82-.87,2.23-4.77,2.89,2.53.35-3.19,5.78-4.85,4.38-.12,6.19,3.1,3.31,1.79,2.98-1.87,4.44-.08,3.59,2.29.82-1.31,3.93.19.71-2.11-4.55-3.09,2.69-2.23-.52-1.25,2.69-1.21-2.02-3.2,1.28-1.62,10.49-1.66,1.37-1.19,7.02-1.79,2.52-2.04,5.04,1.06.88,5.01,2.93-1.16,3.6,1.63-.23,2.58,2.69-.27,7.02-4.49-1.02,1.5,3.58,3.66,6.26,11.58,1.5-2.33,3.86,2.56,4.03-1.14,1.54.8,1.35,2.55,1.96.84,1.2,1.85,3.61-.58,1.49,2.63-2.14,2.83-2.33.4-.13,4.18-1.56,1.86-5.56-1.35-2.02,7.26-1.44.89-5.55,1.58,2.52,6.75-1.92,1,.22,2.16-1.73-.55-1.4-1.37-4.16-.4-4.65-.1-1.02.42-4-1.61-1.59.8-.43,2.24-4.62-1.31-1.84.54-.63,1.66Z"/>
+        <path id="LA" class="cls-1" d="M199.22,372.04l.91-1.3.13-2.44-2.27-2.53-.18-2.87-2.13-2.38-2.12-.2-.56,1.02-1.65.08-.84-.51-2.95,1.74-.07-2.62.69-3.11-1.89-.13-.16-1.78-1.22-.92.6-1.1,2.39-1.94.25.7,1.49.08-.42-3.43,1.45-.44,1.64,2.37,1.26,2.72,3.45.03,1.09,2.59-1.79.77-.81,1.07,3.36,1.76,2.33,3.46,1.77,2.57,2.12,2.02.71,2.04-.51,2.88-2.49-1.06-1.27,1.99-2.31-1.13Z"/>
+        <path id="LB" class="cls-1" d="M4.64,313.76l-.75.05-.26.58h-.94l1-2.73,1.39-2.38.06-.12,1.26.18.46,1.32-1.53,1.27-.69,1.83Z"/>
+        <path id="LK" class="cls-1" d="M133.52,391.2l-.42,2.92-1.17.8-2.44.64-1.34-2.23-.49-4.03,1.27-4.58,1.93,1.57,1.31,1.98,1.35,2.93Z"/>
+        <path id="MM" class="cls-1" d="M183.31,354.78l-1.64,1.28-1.98.14-1.28,3.19-1.18.53,1.36,2.57,1.78,2.13,1.14,1.92-1.02,2.52-.97.53.67,1.45,1.87,2.28.32,1.6-.05,1.33,1.1,2.6-1.54,2.65-1.36,2.91-.27-2.1.86-2.18-.94-1.68.23-3.11-1.14-1.48-.91-3.44-.51-3.66-1.21-2.4-1.85,1.46-3.19,2.06-1.57-.26-1.74-.67.97-3.61-.58-2.74-2.2-3.39.34-1.07-1.64-.38-1.99-2.42-.18-2.41.98.46.05-2.15,1.39-.72-.3-1.28.63-1.03.11-3.15,2.19.7,1.25-2.52.15-1.5,1.54-2.6-.08-1.78,3.63-2.16,2,.57-.23-1.93.99-.58-.22-1.2,1.64-.24.94,1.86,1.22.75.09,2.4-.12,2.57-2.65,2.58-.34,3.63,2.96-.5.67,2.8,1.78.59-.82,2.5,2.08,1.13,1.22.55,2.05-.87.09,1.24-2.39,1.94-.6,1.1-1.6.69Z"/>
+        <path id="MN" class="cls-1 dinosaurs" d="M150.24,253.71l2.96-.74,5.35-3.74,4.27-2.07,2.43,1.35,2.93.06,1.87,2.05,2.8.15,4.06,1.09,2.72-3.03-1.14-2.6,2.91-4.66,3.14,1.87,2.54.53,3.3,1.15.53,3.32,3.99,1.84,2.65-.81,3.54-.57,2.81.58,2.75,2.09,1.7,2.2,2.6-.04,3.53.69,2.58-1.06,3.69-.71,4.11-3.06,1.68.47,1.47,1.46,3.34-.36-1.36,3.25-1.98,4.22.72,1.71,1.59-.53,2.76.65,2.16-1.54,2.25,1.33,2.54,2.89-.31,1.45-2.21-.46-4.07.54-1.98,1.16-2.05,2.66-4.28,1.55-2.79,2.1-2.88-.8-1.58-.36-1.47,2.54.89,1.5.46,1.28-1.97,1.3-2.01,2.05-3.27,1.33-4.21.15-4.53,1.31-3.26,2.01-1.24-1.16h-3.39l-4.15-2.29-2.77-.57-3.73.53-5.79-.85-3.09.09-1.64-2.27-1.28-3.57-1.73-.43-3.39-2.45-3.78-.55-3.33-.68-1.01-1.73,1.08-4.73-1.93-3.31-4-1.57-2.36-2.23-.74-2.97Z"/>
+        <path id="MY" class="cls-1" d="M187.6,394.9l.22,1.44,1.85-.33.92-1.15.64.26,1.66,1.69,1.18,1.87.16,1.88-.3,1.27.27.96.21,1.65.99.77,1.1,2.46-.05.94-1.99.19-2.65-2.06-3.32-2.21-.33-1.42-1.62-1.87-.39-2.31-1.01-1.52.31-2.04-.62-1.19.49-.5,2.28,1.22ZM236.79,399.73l-2.06.95-2.43-.47h-3.22l-.97,3.17-1.08.97-1.44,3.88-2.29.59-2.65-.78-1.34.24-1.63,1.41-1.79-.2-1.81.57-1.92-1.57-.47-1.86,2.05.96,2.17-.52.56-2.36,1.2-.53,3.36-.6,2.01-2.21,1.38-1.77,1.28,1.45.59-.95,1.34.09.16-1.78.13-1.38,2.16-1.95,1.41-2.19h1.13s1.44,1.41,1.44,1.41l.13,1.22,1.85.78,2.34.84-.2,1.1-1.88.14.49,1.35Z"/>
+        <path id="NP" class="cls-1" d="M151.28,331.28l-.22,1.35.37,1.99-.32,1.24-2.33.05-3.38-.73-2.16-.29-1.62-1.59-3.84-.41-3.66-1.77-2.64-1.55-2.72-1.2,1.09-2.99,1.78-1.46,1.16-.78,2.25,1,2.83,2.09,1.57.46.94,1.53,2.18.63,2.28,1.39,3.17.73,3.27.31Z"/>
+        <path id="OM" class="cls-1" d="M69.24,352.01l-1.05,2.04-1.27-.16-.58.71-.45,1.5.34,1.98-.26.36h-1.29s-1.75,1.09-1.75,1.09l-.27,1.43-.64.62-1.74-.02-1.1.74v1.18s-1.35.81-1.35.81l-1.55-.27-1.88.98-1.3.16-.92-2.04-2.19-4.84,8.41-2.96,1.87-5.97-1.29-2.14.07-1.22.82-1.26v-1.25s1.28-.6,1.28-.6l-.5-.42.23-2h1.43s1.26,2.08,1.26,2.08l1.57,1.11,2.06.4,1.66.55,1.27,1.74.76,1,1,.38v.67s-1.03,1.79-1.03,1.79l-.45.84-1.17.99ZM62.32,337.47l-.37.56-.53-1.06.82-1.06.35.27-.27,1.29Z"/>
+        <path id="PH" class="cls-1" d="M258.54,388.69l.29,1.87.17,1.58-.96,2.57-1.02-2.86-1.31,1.42.9,2.06-.8,1.31-3.3-1.63-.79-2.03.86-1.33-1.78-1.33-.88,1.17-1.32-.11-2.08,1.57-.46-.82,1.1-2.37,1.77-.79,1.53-1.06.99,1.27,2.13-.77.46-1.26,1.98-.08-.17-2.18,2.27,1.34.24,1.42.18,1.04ZM251.83,383.43l-1.01.93-.88,1.79-.88.84-1.73-1.95.58-.76.7-.79.31-1.76,1.55-.17-.45,1.91,2.08-2.74-.27,2.7ZM236.47,386.15l-3.73,2.67,1.38-1.97,2.03-1.74,1.68-1.96,1.47-2.82.5,2.31-1.85,1.56-1.48,1.95ZM245.95,378.85l1.68.88h1.78l-.05,1.19-1.3,1.2-1.78.85-.1-1.32.2-1.45-.43-1.35ZM256.09,378.08l.79,3.18-2.16-.75.06.95.69,1.75-1.33.63-.12-1.99-.84-.15-.44-1.72,1.65.23-.04-1.08-1.71-2.18,2.69.06.76,1.07ZM244.95,375.49l-.74,2.47-1.2-1.42-1.43-2.18,2.4.1.97,1.03ZM244.37,359.75l1.73.84.86-.76.25.75-.46,1.22.96,2.09-.74,2.42-1.65.96-.44,2.33.63,2.29,1.49.32,1.24-.34,3.5,1.59-.27,1.56.92.69-.29,1.32-2.18-1.4-1.04-1.5-.72,1.05-1.79-1.72-2.55.42-1.4-.63.14-1.19.88-.73-.84-.67-.36,1.04-1.38-1.65-.42-1.26-.1-2.77,1.13.96.29-4.55.91-2.66,1.7-.02Z"/>
+        <path id="PK" class="cls-1" d="M110.47,314.47l-.58-4.8,1.87-1.68-3.04-2.61h-.96s-.36-1.14-.36-1.14l3.08-2.34.12-.11-1.94.24-3.01.73-1.64,1.51.66,1.46.33,1.7-1.4,1.43.12,1.3-.77,1.22-2.67-.11,1.1,2.22-1.78.85-1.19,2,.15,1.98-1.1.92-1.03-.3-2.15.43-.3.91h-2.09l-1.56,1.84-.1,2.75-3.65,1.33-1.95-.28-.57.7-1.67-.4-2.81.48-4.69-1.64,1.38,1.68,1.13,1.93,2.68,1.4.08,2.77,1.34.51.23,1.44-4.04,1.61-1.06,3.6,3.95-.44,4.56-.05,5.16-.58,2.17,2.34.84,2.2,2.05.77,1.87-2.04h6.17s-.56-2.63-.56-2.63l-1.57-1.56-.32-2.39-1.84-1.4,3.09-3.3,3.26.24,2.93-3.32,1.76-3.26,2.72-3.24-.04-2.33,2.39-1.91-2.27-1.63v-.04s-.1.03-.1.03l-2.38-.99Z"/>
+        <path id="PS" class="cls-1" d="M3.87,316.7v2.01l-.42.96-1.32.45.13-.86.71-.46-.7-.36.58-2.2,1.02.46Z"/>
+        <path id="QA" class="cls-1" d="M46.67,340.99l-.19-2.24.76-1.62.76-.34.85.97.05,1.81-.61,1.81-.78.22-.84-.61Z"/>
+        <path id="RU-asia" class="cls-1 dinosaurs" d="M302.52,99.46l3.98,3.81.34,2.57-4.25.07-5.75-1.08-.49-.52,2.66-3.92,3.51-.93ZM198.82,47.17l.83-5.72-7.11-8.34-2.11-.98-2.3,1.7-5.12,18.6,15.81-5.26ZM323.59,90.86l3.24-4.25-7.04-2.88-5.23-1.68-.67,3.59,5.21,4.27,4.49.95ZM408.9,127.48l3.66-.52,2.89-2.06.24-1.19-4.06-2.51-2.38-.02-.36.37-3.57,3.64.5,2.73,3.08-.44ZM298.46,89.17l10.33.3,2.21-8.14-10.13-6.07-7.4-.51-3.7,2.18-1.51,7.75,5.55,7.01,4.65-2.52ZM165.84,30.9l4.65,5.73,7.81,4.2,6.12-1.8.69-13.62-6.46-16.04-5.45-9.02-6.07,4.11-7.28,11.83,3.83,3.27,2.16,11.34ZM431.73,158.02l-7.61-1.07.65,5.15-1.89-1.74.24-4.44-7.35-7.34-6.87-5.86-3.92-3.44-8.06-3.84-5.83.49-8.95-2.29-1.25,3.62,2.27,5.07-3.47,2.48-4.88-6.99-5.31.89-5.29-1.57-4.97.21-3.75,1.64-3.45-2.29.34-6.02-2.33-3.5-5.58-1.41-11.32,1.62-7.34-6.65-2.39-5.35-25.32-6.06-3.7,4.07,2.02,8.4-4.59-1.24-2.07,2.47-5.43-2.72-4.78,2.38-4.5-4-2.71,9.18-4.41-3.49-3.52-6.98,1.66-3.84-1.29-6.04-4.53-5.14-4.48.05-5.95-1.74-.16,7.47-11.72-1.43-.68-4.58-9-1.65-4.48,1.57-1.23,2.56-1.43-6.39-2.52,1.91-4.15-2.55-3.48-1.43,2.11-3.08,7.37-5.92,3.1-3.24.7-5.85-2.25-4.35-6.32-5.84-8.2-.16-2.56,2.94-.76-6.03-6.35-1.92,3.82-3.13-4.81-4.21-6.62,5.31-2.68,5.33-.77,5.24-5.15-.2-6.29,6.24-2.29-2.61-7.36,1.08-.94,3.15-7.4,1.51-5.45,5.51-3.22.3-3.25,7.02,2.28,5.38-6.08,1.32-6.74-.44-4.88,2.02.28,10.28,2.49,7.62-5.18-5.18-5.82.49-4.69,3.58,1.28,6.37-3-1.59,1.11-8.67-1.47-5.19-1.4.22.65,6.62-5.02,6.04,3.64,7.03-2.24,8.28.67,4.42,3.07.64-1.31,5.08,1.63,4.26-2.43,3.49-.74,3.55-3.08,1.82-1.12,2.51-3.21-1.02,5.49-10.2,1.2-5.01-3.09-4.73.64-11.07-.9-5.94-1.74-2.78,2.66-7.28-.58-5.18-7.42-2.51-2.08,1.88-1.84,8.42-5.17,7.99.08,2.74,1.5,6.49-.92,3.83,3.38.78.08,1.68,2.85,4.11-1.88,3.97-1.6-1.39.92,7.34-2.03,4.04-3.76,3.86-3.93,3.13-4.82,5.44-5.23,4.07-1.89,19.97-1.34,8.84-1.77,5-1.78,3.54-2.17,2.73-.56,2.42,2.11,1.76.31,2.15-1.66,2.67-2.57.64-3.08.37,1.3,2.44,1.2,1.83,1.81,2.32,1.9-1.5,1.84-1.12,1.89,1.07.71,1.75-1.33,1.39-1.13,2.2.09,2.65,1.88.67.8,1.88.3.11,2.33-1.92-.53-1.25,2.69-1.2-2.02-3.21,1.29-1.62,10.49-1.66,1.37-1.19,7.02-1.79,2.52-2.03,5.04,1.06.88,5.01,2.93-1.17,3.6,1.63-.23,2.58,2.69-.27,7.03-4.49-1.03,1.5,3.58,3.66,6.27,11.58,1.49-2.33,3.86,2.57,4.03-1.14,1.55.8,1.35,2.55,1.96.85,1.19,1.84,3.61-.58,1.49,2.63,1.1-.35,2.96-.74,5.35-3.74,4.26-2.07,2.44,1.35,2.92.06,1.87,2.04,2.8.16,4.05,1.09,2.73-3.03-1.14-2.6,2.9-4.66,3.14,1.87,2.54.53,3.3,1.15.53,3.32,3.98,1.84,2.65-.81,3.55-.57,2.81.58,2.75,2.09,1.7,2.2,2.6-.04,3.53.7,2.58-1.06,3.69-.71,4.11-3.06,1.68.47,1.47,1.46,3.35-.36,3.37,1.63,3.95-2.74-.03-1.93,2.53-4.72,1.56-1.45-.04-2.52-1.54-1.1,2.32-2.31,3.48-.84,3.72-.13,4.2,1.39,2.46,1.71,1.73,4.61,1.05,1.94.98,2.73,1.04,4.28,4.88,1.38,3.32,3.03,1.13,3.95h4.26l2.43-1.65,4.63-1.24-1.47,3.76-1.09,1.51-.96,4.46-1.89,3.89-3.4-.7-2.41,1.4.74,3.36-.4,4.55-1.43.1.02,1.93.39.66.44-1.26,3.76-2.79,1.76,1.86,1.77-.05,3.74-2.25,1.81-2.28,3.78-4.53,3.81-4.66.94-2.82,4.23-6.01,1.26-6.85.24-5.27,2.19-4.51-.09-3.92-4.05-5.19-3.08-.31-1.79,2.38-2.72-1.05-1.38-3-4.42-.61,10.75-11.78,9.08-10.33,9.22-1.62,8.57.94,3.47-2.7,4.36.84-.2,3.95,4.33-.56,6.26-1.42-2.32-3.38,7.02-9.56,7.24-2,2.3,7.14,7.11-6.36,1.68-4.92,3.41-.51-2.25,8.37-5.04,4.57-4.83,5.73-5.01,6.78-4.36,1.18-.15,2.44-2.37,3.07-1.35,6.95,1.56,10.65,1.2,6.69,1.04,3.08,4.04-4.18.84-4.66,4.18-1.14.97-5.4,4.9-2.47-1.17-2.1,1.2-4.15,2.6-.19.38-7.46-3.19-1.17-.1-2.14,3.36-5.22.9-3.63,3.75.76,2.7-2.39,1.28,2.08,7.31-4.42,4.01,3.9,1.03-2.55,4.07-3.5,4.29-4.1,2.49-.69,7.84-4.51,5.23,1.32.72-1.6-.33-2.53-1.29-1.67-1.67-5.2-2.53-3.44,3.63.48,3.6-2.89.02-.04,1.59-2.83-1.26-3.19,3.36-1.64-.61,2.55,1.52,2.37,3.19-.87,2.85,1.1.63,2.89,3.73,1.92,2.14,2.26,2.63.19,1.12-1.35.07-6.46,4.6-.7,2.78-2.97-5.53-6.58ZM309.79,255.08l-2.82-7.68-1.16-4.51.07-4.5-.97-4.5-.73-3.15-1.25.67,1.11,2.21-2.59,2.17-.25,6.3,1.64,4.41-.12,5.85-.65,3.24.32,4.54-.31,4.01.52,3.4,1.84-3.13,2.13,2.44.08-2.84-2.73-4.23,1.72-6.11,4.15,1.41Z"/>
+        <path id="SA" class="cls-1" d="M24.15,366.05l-.36-1.24-.85-.88-.22-1.17-1.44-1.04-1.5-2.46-.79-2.41-1.94-2.04-1.25-.48-1.86-2.85-.32-2.08.12-1.79-1.61-3.36-1.31-1.19-1.52-.63-.92-1.76.15-.69-.78-1.6-.82-.69-1.09-2.32-1.71-2.52-1.43-2.16h-1.39s.43-1.73.43-1.73l.13-1.11.34-1.28,3.12.51,1.22-.98.67-1.16,2.14-.44.46-1.09.93-.54-2.8-3.26,5.62-1.65.53-.49,3.38.89,4.18,2.29,7.9,6.49,5.21.26,2.5.31.7,1.51,1.98-.08,1.1,2.73,1.38.71.48,1.11,1.91,1.31.17,1.29-.28,1.03.36,1.04.8.87.38,1.01.42.75.84.61.78-.22.53,1.17.11.71,1.08,3.08,8.42,1.52.57-.64,1.28,2.14-1.87,5.97-8.41,2.96-8.08,1.13-2.62,1.32-2.01,3.07-1.31.48-.7-.97-1.07.15-2.71-.29-.52-.3-3.23.07-.76.27-1.15-.76-.75,1.43.29,1.23-1.23.93Z"/>
+        <path id="SY" class="cls-1" d="M12.97,313.43l-5.49,3.54-3.12-1.32-.06-.02.38-.5-.04-1.37.69-1.83,1.53-1.27-.46-1.32-1.26-.18-.26-2.61.68-1.41.75-.75.75-.76.16-1.94.91.68,3.09-.97,1.49.65h2.31s3.22-1.32,3.22-1.32l1.52.06,3.19-.54-1.44,2.18-1.54.86.27,2.52-1.06,4.12-6.21,3.5Z"/>
+        <path id="TH" class="cls-1" d="M191.84,378.01l-2.52-1.31-2.4.06.41-2.25-2.47.02-.22,3.14-1.51,4.15-.91,2.5.19,2.05,1.82.09,1.14,2.57.51,2.43,1.56,1.61,1.7.33,1.45,1.45-.91,1.15-1.86.34-.22-1.44-2.28-1.23-.49.5-1.11-1.07-.48-1.39-1.49-1.59-1.36-1.33-.46,1.65-.53-1.56.31-1.76.82-2.71,1.36-2.91,1.54-2.65-1.1-2.6.05-1.33-.32-1.6-1.87-2.28-.67-1.45.97-.53,1.02-2.52-1.14-1.92-1.78-2.13-1.36-2.57,1.18-.53,1.28-3.19,1.98-.14,1.64-1.28,1.6-.69,1.22.92.16,1.78,1.89.13-.69,3.11.07,2.62,2.95-1.74.84.51,1.65-.08.56-1.02,2.12.2,2.13,2.38.18,2.87,2.27,2.53-.13,2.44-.91,1.3-2.63-.41-3.62.55-1.8,2.38.67,3.45Z"/>
+        <path id="TJ" class="cls-1" d="M103.32,289.45l-1.03,1.13-3.05-.61-.27,2.1,3.04-.28,3.47,1.17,5.3-.55.71,3.33.92-.36,1.7.81-.09,1.38.42,2.01h-2.9l-1.93-.26-1.74,1.57-1.25.34-.98.74-1.11-1.15.27-2.95-.85-.17.3-1.09-1.51-.8-1.21,1.23-.3,1.43-.43.52-1.68-.07-.9,1.6-.95-.67-2.03,1.12-.85-.42,1.57-3.57-.6-2.66-2.06-.86.73-1.59,2.34.17,1.33-2.01.89-2.35,3.75-.86-.58,1.71.4,1.02,1.16-.1Z"/>
+        <path id="TL" class="cls-1" d="M254.6,437.08l.33-.66,2.41-.63,1.96-.1.87-.35,1.06.35-1.03.76-2.92,1.23-2.35.82-.05-.86-.28-.56Z"/>
+        <path id="TM" class="cls-1" d="M75.83,305.73l-.25-2.91-2.09-.12-3.2-3.09-2.24-.39-3.1-1.79-2-.33-1.23.66-1.87-.1-1.99,2.02-2.47.68-.52-2.49.41-3.73-2.19-1.22.72-2.48-1.86-.22.62-3.09,2.64.91,2.47-1.19-2.05-2.23-.8-2.14-2.26.96-.28,2.73-.88-2.41,1.24-1.25,3.18-.79,1.9,1.06,1.96,2.93,1.44-.18,3.16-.05-.46-1.88,2.4-1.3,2.36-2.2,3.78,2,.3,2.99,1.07.77,3.03-.17.94.67,1.38,3.79,3.21,2.51,1.83,1.69,2.93,1.75,3.73,1.52-.08,2.16-.84-.11-1.33-.94-.44,1.25-2.36.68-.56,2.79-1.58,1.05-2.21.52-.59,1.55-2.11.46-2.86-1.3Z"/>
+        <path id="TW" class="cls-1" d="M245.65,342.1l-1.69,4.87-1.2,2.48-1.48-2.55-.32-2.25,1.65-3,2.25-2.32,1.28.91-.49,1.86Z"/>
+        <path id="UZ" class="cls-1 dinosaurs" d="M90.71,299.78l.08-2.16-3.73-1.52-2.93-1.75-1.83-1.69-3.21-2.51-1.38-3.79-.94-.67-3.03.17-1.07-.77-.3-2.99-3.78-2-2.36,2.2-2.4,1.3.46,1.88-3.16.05-.11-14.13,7.22-2.35.52.35,4.35,2.84,2.29,1.48,2.68,3.5,3.29-.56,4.81-.3,3.35,2.8-.21,3.8,1.37.03.57,3.06,3.57.12.76,1.75,1.05-.02,1.23-2.65,3.69-2.61,1.61-.7.83.37-2.35,2.43,2.07,1.4,2-.93,3.32,1.96-3.59,2.64-2.13-.36-1.16.1-.4-1.02.58-1.71-3.75.86-.89,2.35-1.33,2.01-2.34-.17-.73,1.59,2.06.86.6,2.66-1.57,3.57-2.12-.74-1.56-.03Z"/>
+        <path id="VN" class="cls-1" d="M207.16,350.7l-3.74,2.56-2.34,2.81-.62,2.05,2.15,3.09,2.62,3.82,2.54,1.79,1.71,2.33,1.28,5.32-.38,5.02-2.33,1.87-3.22,1.83-2.28,2.36-3.5,2.62-1.02-1.81.79-1.91-2.08-1.61,2.43-1.14,2.94-.2-1.23-1.73,4.71-2.19.35-3.42-.65-1.92.51-2.88-.71-2.04-2.12-2.02-1.77-2.57-2.33-3.46-3.36-1.76.81-1.07,1.79-.77-1.09-2.59-3.45-.03-1.26-2.72-1.64-2.37,1.51-.74,2.23.02,2.73-.35,2.39-1.62,1.35,1.14,2.57.55-.45,1.74,1.34,1.22,2.82.78Z"/>
+        <path id="YE" class="cls-1" d="M53.11,365.16l-2.03.79-.54,1.28-.07.99-2.79,1.22-4.48,1.35-2.51,2.03-1.23.15-.84-.17-1.64,1.2-1.79.55-2.35.15-.71.16-.61.75-.74.21-.43.73-1.39-.06-.9.38-1.94-.14-.73-1.67.08-1.57-.45-.85-.55-2.12-.81-1.19.56-.14-.29-1.32.34-.56-.12-1.26,1.23-.93-.29-1.23.75-1.43,1.15.76.76-.27,3.23-.07.52.3,2.71.29,1.07-.15.7.97,1.31-.48,2.01-3.07,2.62-1.32,8.08-1.13,2.2,4.84.91,2.03Z"/>
+      </g>
+    </g>
+  </svg>`,
+  australia: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="AU" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 442.51 415.55">
+    <defs>
+      <style>
+        #AU {
+          fill: #2a7035;
+          cursor: pointer;
+          stroke: #2a7035;
+          stroke-width: 0;
+        }
+        #AU:hover {
+          fill: #194520;
+          stroke: #2a7035;
+        }
+      </style>
+    </defs>
+
+    <g data-name="Layer 1">
+      <path id="AU" class="cls-1" d="M195.55,27.83l.63-1.15-1.5-1.01.76-1,1.53.05-.03-2.3,5.08,1.63,4.76-.32,1.42.45,1.75-1.2,1.49.03.61,1.16-.37.8.66-.86v-1.68l1.45-.52,1.85.36-1.2-.63-.12-.51.08-3.17.61-.96-2.19-2-1.27-.43-3.1.96-1.26-1.76-1.14-.61-1.02.09,1.12-1.54,1.2-.55,1.15,1.13.82,1.94.76-.56-.96-2.37,1.41.26.7-.37,1.77,1.48,1.83,2.68.59.13,1.61-2.05.47.04,2.88,4.18,1.88-.18.55.8,1.54.65,1.26-.15,1.46-1.11,1.18.2-.75,1.3,1.12.64,1.17-.75,1.47,1.12.23,1.27,1.77-.26,1.59.32,2.77-.78,1.08,1.5,2.31,1.49,1.6.31,1.05-2.02,2.8-.15.83-.67-.87.23.12-1.1,1.74-.25,1.38-1.1-.04-.77h.4s.37.44.37.44l-2.18,2.1.46.54-1.29,1.62.72.99,1.61-1.42,1.51-.44-.92,1.16.37.56.8-.37.19.33-.87,1.26.61,1.04,2.18-.18.58-.45.76-1.93-2.02-.8,2.97-2.35,1.21-.22-.25.83,1.46,3.03.91-.31-.69-.71.88-.22,1.1.56,1.11,1.5-2.53,2.05-1.1,1.77-.25,1.34-1.37-.51.55,3.05-.95,2.19-.1-.35-.88.3-.12-2.14-1.7,2.05-1.12-.6-1.84,1.03-.39.67.47.65-.77,1.05-.2,1.25.89,1.61,1.07-.68-1.4,5.06-2.99,3.28-.9,2.33-1.8.92,1.13,2.63,7.99,4.81,1.17,2.34,2.7,1.29,2.1,2.03,2.87.12,3.68,2.39,4,1.39,1.32,2.1,1.54,1.31h0v109.55h0l-97.64.07h0l-.07-128.49h0l2.01,1.22.06,2.26.71-.88-.29-2.96,2.52,1.02,2.33,3.05.6-.23-1.15-1.79.23-2.05,1.67.29,1.55-1.07-1.94.61-.98-.63-.96-1.54,1.27-.29.75-.74-1.22.39-1.3-.24-1.97-1.52.13-.89,1.33-2.13,2.61-1.64,1.04-5.31.77-.8.48.91,1.05-.19,2.42-2.05-1.61-1.97.27-2.46.45-.2.47.43,1.11-.58.29-2.45,1.68-.75,1,1.14.96-.07-.92-.27-.31-1.17.37-1.59-.33-.25,2.07.27,1.37,2.34ZM196.23,20.09l-5.05-3.27-1.4-4.28.3-.99,1.96,2.05,1.43.45,2.11-.91,1.14.49,1.6-1.01.78,1.55-.33-1.21.85-1.12,1.59.64,1.2,1.58.09.83-2.77,2.91-3.5,2.29ZM260.82,44.9l.64-.31-.28,1.4-.57.22-5.55-.91.55-3.66.31-.53,1.81-.55-.31-.7.91-.76.47.03-.5.81.4,1.16,1.03.18.8-1.13.4.63-1.64,1.47-.64,2.51,1.76.49.41-.35ZM191.17,18.98l-1.8-.69-2.51.64-.78-.39.52-1.37.91.32.4-.39-.21-2.01.89-1.55.96-.2,1.6,4,1.58.93-.25.55-1.31.16ZM256.42,14.77l2.94-4.79-.53,2.04-1.77,2.48-.64.27ZM214,13.54l-1.02-3.23.83-.32.52,1.71-.33,1.84ZM262.4,63.33l-1.3-1.47.82-1.19.9,1.94-.42.72ZM253.14,41.35l-.31-1.01-.11.72-.62-.21.1-.93.95-.75.95.73-.96,1.45ZM252.94,17.35v-.51l.99-.79.82-.04,1.44-1.12-1.19,1.38-2.06,1.08ZM256.65,61.27l-.06-1.08.93-.28-.19,1.23-.68.13Z"/>
+      <path id="AU" class="cls-1" d="M14.73,181.05l.2-1.84-2.43-2.57-2.47-5.41-1.15-.7-.98-2.78.03-1.61-2.44-3.85-.06-2.35,1.61-5.6.75-1.5,1.54-1.21.27-1.01-.09-2.91.63-1.92-.12-1.14-1.64-3.85,3.01-7.33,1.05-1.56,1.44-.5.14.28-1.07,4.09.7,1.48-.34,2.16.47.48,2.37-1,3.02-7.02,3.17-1.76,1.19-.09,4.33-1.9,1.82-2.15,2.03-1.1,1.29-1.94,2.55-1.19.97-1.12,2.46-.73,2.52-1.81,1.83.79,2.07-1.11,2.99,1.26,3.72-.63,1.79-.78,3.56-2.99.61.29,6.15-1.02,1.94-2.81,1.09-.83,1.18-.11,4.27,1.28,2.21-1.14,5.99-1.07,7.49-3.28,2.71-2.26,2.3-3.11,1.64-3.53,1.5-2.02.25-1.78,1.51-.37,4.32-3.6.26-1.43-1.18-.54-.44-.7-.12-2.14-.64-1.66.37-3.56,2.22-2.79.74-.47h1.38s-.55-1.24-.55-1.24l.46-.62,2.02-.31.09-1.82,1.45-1.11.28-.99.77-.28.74.74-.77.23-.36,1.25,1.52,1.46,5.2,9.95-.13-2.6.6-1.82-.49-1.5.2-.86.69-.02,2.17,2.39.65.02-.91-.81-.39-1.56,1.03-1.18-1.36-.06-1.72-2.42.06-1.23-1.22.02-.83-.75,2.34.34.74-1.12-.02-1.01-1.6-.58.12-1.31,1.69-.37.87.68-.8.68,1.72,1.34.79-1.5,1.91.45.98,1.34,1.12.06.56-.64.99.48,4.53.12-3.02-.89-2.33.02-.27-1.21.59-1.34.44-.1.52.79.81-.56.24-2.2,1.22-1.23-1.2-.07-.88,1.5-1.46-.81-.31-2.2.93-2.13,1.35.43.81-.42.08-2.42.46-.11,5.17,2.99-1.14-1.01.23-1.44-2.01.38.28-1.2h.73l.43-.74-1.76.85-.62-.86.82-.67.87.12,1.06-1.24.94,1.82-.03-1.44,1.74,1.36,1.23-.26-2.73-2.06.65-.52-1.16-1.72,2.17-2.5,2.74.29-.1-2.68.42-.96,1.2.57-.94,3.98.87-2.52,1.31.67-.2,1.2.84.79,1.46-1.43,1.18-4.29-.71-1.82-.69-.44,2.17.39-.09.84-.56.24.37,1.22,1.04.72.5-2.03,1.83-.95-.51,1.25,1.58,1.69,2.05-2.87-.47-1.96,1.22-.49,1.06-.25.67.54,1.2,2.17.28-.79,2.94.6,1.04,1.47,1.49,1,3.25,4.52.47-.23,1.84,1.58.2,1.09-1.29,3.36.04,3.31-.57,1.14.83-.72.38-2.46,1.36,1.03.32,1.06.05-1.08-1.08-2.67,1.24-1.7.51,1.47,1.03.03-.66-2.76,1.61-.46,5.04,1.26h0l.06,128.49h0l-.06,70.13h0l-8.88,4.31-9.21,3-7.03.63-5.51-.99-1.2.59-1.05-.23-4.96,3.58-2.31.67-6.03,3.84-5.04,1.13-1.91,1.98-1.63,5.43-2.02,1.76-.38,1.1-2.11,1.64-1.87-.14-2.18,1.61-.73-1.72-.9-.38-1.88.64-5.33.08-1,.69-.12.66-1.44.15-.42-.14-.09-1.62-.84-.89-2.31.97-3.47-1.07-6.5.51-3.5.79-1.34.74-4.47-.59-2.08.64-2.14,1.59-1.61,3.03-1.54,1.51-1.27.75-3.22-.67-1.87,1.27.16.75-.38.43-2.63,1.1-2.12,2.39-2.17,1.16-2.65.48.21.55,1.5.33-.67.38-2.25-1.02-1.02.28-.25.91-3.01-1.58h-4.16s-.82.53-.82.53l-4.48-.75-2.64-1.76-2.64-.61-1.73-2.78-3.46-2.94-2.76-1.12-1.29.85-1.23-1.31-.6-7.43.47-2.18,2.43,1.68,1.92-.41,3.43-4.63-1.28-7.64,1.09-1.9-.68,1.62.94,1.56.47-2.54-.61-10.92-7.15-15.29-.26-2.89-.94-2.67.44-7.12-.98-3.49-.57-1.35-2.34-2.78-.38-2.93-.88-1.78-3.78-4.87-.65-2.93.5-1.76-.21-1.27-2-4.92-3.81-6.15-3.31-3.52-.61-1.93.69-2.54.5,3.02.34-1.92,1.61,1.99.35,2.7,1.05,1.77,1.53-.61.78-1.16.06-2.06-3.37-3.15-.63-2.6-1.1-1.78.85-2.03,2.47,3.05.4,1.13-.57,1.75.35,1.95.92-.34.87-1.68.87,3.6,1.2,1.61,1.36-1.07.29-.97-.54-1.32.37-2.33.35.09ZM3.5,182.78l-2.75-4.33-.49-1.88.51-1.67.54.13-.17.86,2.39,5.66-.03,1.23ZM126.35,59.02l-.58-1.35.98-.72.94,1.57-1.34.5ZM26.98,120.78l-.72-.14v-.61s1.48-1.72,1.48-1.72l.23,1.22-.99,1.25ZM133.04,50.12l-.57-.31.07-.85.76-1.17.5.56-.76,1.77Z"/>
+      <path id="AU" class="cls-1" d="M392.54,306.28l-1.26-.64h0l-1.73-2.94.02-1.84.52-2.56,2.85-2.1,3.51,2.19-2.31.69-.52.72-.62,2.49.16,3.14h0l-.62.85Z"/>
+      <path id="AU" class="cls-1" d="M402.69,327.75l-19.25-9.7-.98-.08.99-2.48-1.68-2.93-.56-4.42-1.18-.75-1.82-.78-.93.36-1.53-.29-1.01.17-1.01,1.26-2.8.73-1.46-.1-4.52-1.69-2.2.99-5.35-.78-3.1-2.19-1.77.7-1.31-.39-2.7.35-.58,1.62.17,1.36-2.34.43-1.34-.78-2.28-2.67-.56-1.28-2.81-2.59-3.52-2.26-1.89-.68-.35-1.69-1.69-.72-.5-1.1-.16-4.01-2.31-1.26-2.89-.62-.82-.81-.81.48-.84,2.17-.92-.18-.93-2.25-.94-.87-.05-1.88-1.51-2.54-2.12-1.16-1.7.35-1.61-.42-2.1,1.48-3.19-1.83-1.87-.14-.84-.6h0l.05-63.7h86.31l5.57-4.95,1.2-.33,1.4.69,6.37-1.08,1.88,1.44,3.45-.29,2.74,1.25,2.77,2.39.13,1.88.68,1.09,1.06-.68.98-2,1.87-.83,1.07,1.1,2.08-1,.35-2.01-.98-2.22,2.27-1.44,2.16-.71,1.28-1.38,1.19.8,1.78.39,3.82-.09.83-1.1,2.64-.73.81-.66h0l.78.71.15.81-.35,2.67.7,2.08-.2,2.48-1.71,2.24-1.2,3.52.3.87-1.21,6.89-2.55,7.71-.14,3.35.75,1.43-2.46,8.95-3.66,6.04.19,3.48-4.37,3.57.19,1.13-3.61,1.52-2.58,4.88-1.3.64-.11.59.46.41-.35,1.39-1.24,1.66-.73,4.43-1.88.58,1.04.56-2.83,3.59-1.02,2.97.44.54-.54,2.46-.92,1.21,1.14,1.9-.17.88-.29.25-.36-.39-.02-.88-.98.4.2,1.12-1.77,1.32-4.1,9.49-.29,5.19-.63.82-1.9,7.2.05,1.97,1.08,1.25.14.94v.45s-.78.24-.78.24l.35,3.13ZM393,302.29l.62-2.48.52-.73,2.31-.68-3.5-2.19-2.85,2.1-.52,2.56-.02,1.85,1.73,2.94h0l1.35.59.53-.8-.17-3.16Z"/>
+      <path id="AU" class="cls-1" d="M254.27,283.55l-1.08,1.74-.88.39-.93,1.59-1.12.7-.67,1.33-.77,2.33.75.06,1.45-1.02.15.69-.17,2.59-.52.22-1.84-1.95-.68.1-.48,1.13h-.39s-3.44-4.33-3.44-4.33l-1.2-.78-1.12.39,1.04-2.08.05.82,1.42,1.52.57.39,1.08-.35-1.23-.85-1.31-7.14-2.47-3.1-2.09-1.7.19-2.15-1.63-3.81-1.24-.48-2.87.81-.66-.74-2.11-5.47.87.06.58,1.05.96-1.67-.22-1.14-1.76-1.73-2.84,1.13.33-1.76.73-.08-1.35-1.88-1.78-1.26.12-.47-.74-.23-1.31,1.43-3.46-.24-4.14-2.99-2.18-.26-2.79,1.3-1.43-.14-4.63-3.88-6.74-3.15-3.85,1.73-7-.39-12.48,1.44h0l.06-70.18h0l97.64-.07h32.57l.03,36.55h0l-.05,63.7h0l-.28-.37-.04,54.4h0l-2.93.11-1.83-.98-1.89-1.9-2.55-4.52-1.38-.52-2.63-3.98.34-.7-.54-1.81.2-1.25,1.19-1.71-.42-3.24-1.9-4.63-3-4-3.14-3.08,3.49,2.76,2.78,4.4-1.02-2.95-5.07-5.1-.02-1.18.67-.27.82.71.07,1.9,1.02-.11.23-4.21-1.56-.77-.84.83-1.79.39-.11.89.83.46-.49.63-2.93-.25-2.01,1.52-3.92.21-.89-.67.65-1.15,1.44-.72,1.66-1.79.28-3.05.53-1.16-.2-3.9-2.4-3.27-.6-2.16-1.38-2.35-.93,1.6.13,1.14-1.03,1.06-.38,1.31-.15,2.44-1.38,5.56-.8.65-2.75-.81-2.04.77-.72.93-1.51-.23-1.47.97-.66-.69,1.14-1.38.95-3.24.91.36,1.89-.3,1.12.72.54-.29.7-3.69-.06-4.88-.45-1.65,1-.85.64-2.76,3.75-4.22-1.46-3.64.33-.98,2.19-.94-.08-.66-.93-.89-1.5-5.68-.58-.44-.09,2.09.5.58.07,1.37-.36,1.86-1.95.48-1.16,1.26-3.02,6.97-.74.62-1.32.05-.77-.47-.93,1.64-4.89,3.05-1.13,2.03ZM271.7,303.73l1.65.45.54,1.42-.79.71-1.29-.63-1.74-.13-1.53.72-.06,1.2-1.73,1.01-1.06-1.08-1.43-.27-.82.8-4.75.17-1.89-1.86.46-1.97,5.85-1.43,2.14-.95,2.27.19.64.61.14,1.28,2.1.75.64.05.66-1.04Z"/>
+      <path id="AU" class="cls-1" d="M398.36,330.79l-2.92.82-10.86.36-5.97,2.2-4.2,3.51-5.42,5.71-4.8.91-2.3-.02-.38.57,1.21,2.21,1.25-.88.23-.84.39.18.16,3.76-.97.99-.81-.69-.79-2.07-1.14-1.43-.84-.36-.68.6-1.07-.09-.96-2.84-.96-.15-1.31.58-2.08-1.91.18-1.29,1.34-1.08-1.21-2.03-2.05.15-.46,1.54-1.98,2.04-1.4.12-1.52-1.77,2.43-.51,1.59-2.8-.35-1.29-1.97-2.39-4.24,3.19-1.72.35.06.5,1.5.26,1.35-.65.75.21.05,1.01-.56.86-3.21.54-3.27,2.21-2.38,2.93-1.2.33-2.05,1.93-5.37-3.08-2.66-.85-2.91-2.25-1.69-.67-2.38.5-2.12-1.49-2.32-.34-1.42.75.33,1.08-.86.24-1.27-.62-2.83-2.92-2.44-1.32h0l.04-54.4.28.37h0l.84.6,1.87.14,3.19,1.83,2.1-1.48,1.61.42,1.7-.35,2.12,1.16,1.51,2.54.05,1.88.94.87.93,2.25.92.18.84-2.17.81-.48.82.81,2.89.62,2.31,1.26.16,4.01.5,1.1,1.69.72.35,1.69,1.89.68,3.52,2.26,2.81,2.59.56,1.28,2.28,2.67,1.34.78,2.34-.43-.17-1.36.58-1.62,2.7-.35,1.31.39,1.77-.7,3.1,2.19,5.35.78,2.2-.99,4.52,1.69,1.46.1,2.8-.73,1.01-1.26,1.01-.17,1.53.29.93-.36,1.82.78,1.18.75.56,4.42,1.68,2.93-.99,2.48.98.08,19.25,9.7h0l-2.09.65-1.22,1.86-1,.54ZM352.52,340.34l-.78-.3.13-1.52,1.74.36.44.52-1.53.94ZM352.5,342.26l-.9-.68h-1.64l.76-.74,1.01-.2.77,1.62Z"/>
+      <path id="AU" class="cls-1" d="M402.69,140.41l.66.82-1.28-3.39.85-2.22.86.03,1.32,2.67,4.26,2.27-.64-2.37.3-.61.79.14.77,1.15-.24,1.31,1.47,2.09-.69,4.67.61,1.26-.13,1.68.72,1.46,1.69.61,1.57,2.78,1.13.41,1.66,1.78,1.33.68-.09.6.83-.4.28-.89.48.14,2.71,2.71,2.15,4.63,3.65,2.42.43,2.21,1.78,2.95,2.57.52.42,1.76-.58,1.51.29,2.07.69.44.58,1.42,1.6.99-1.23,4.03.92,9.36-1.33,1.14.33,1.66,2.08,2.02.68,2.33.92,1.29.32.87-.24,1.63,1.04,2.06h0l-.81.66-2.64.73-.83,1.1-3.82.09-1.78-.39-1.19-.8-1.28,1.38-2.16.71-2.27,1.44.98,2.22-.35,2.01-2.08,1-1.07-1.1-1.87.83-.98,2-1.06.68-.68-1.09-.13-1.88-2.77-2.39-2.74-1.25-3.45.29-1.88-1.44-6.37,1.08-1.4-.69-1.2.33-5.57,4.95h-86.31l-.02-36.54h-32.57v-109.55h0l3.15,2.11,3.79.42,2.21,1.09,1.82.26,1.42,1.97.23,1.46,1.06,1.58,1.92.41,1.95,1.6,2.15.62,2.22,1.56,4.1-.45,4.74-2.51,1.34-5.09,2.58-3.35,1-2.26,1.52-4.82-.28-1.89.65-4.14,2.23-5.7-1.07-2.94-.49-3.15.86-3.76-1.33-2.17-.09-1.34.79-3.28,1.59-3.09-1.03-3.28,1.67-1.47.53-1.47.65.2.82,2.19.09-.47-.61-1.78-1.8-2.33.32-.26-.93-.21.25,1.11-.63.2-.36-.42,1.97-3.51.8-2.48.47-.21.27,1.32,1.16-.23-.84-1.58,1.91-6.08.27-4.66,2.12-.59,1.1-2.04,1.83.45-1.05,1.18-.08,1.01,1.08-.85,1.47,1.25.47,1.03.84,3.42-.07,4.87,1.18,1.04,1.46-.26,1.04.93-1.05,1.69-.27,2.14,2.15.87.04,1.16,1.64,1.1-.73,2.57,1.65.33.08,4.71.79,1.42-.69,3.6.77,2,.92.97,1.05,4.34,1.14.82h1.35s4.59-2.59,4.59-2.59l.45-1.01.43.1.68,1.02.39,2.47.68.81,2.39.68.68,1.51,3.87,2.14-.88,2.11.6,2.2-.44,1.29,1.21,3.23v2.03l1.12,1.88-.65,4.02,4.39,5.32,1.36-.47.26.32-.79,1.95,1.62,3.31.79,3.7-.11,2.99-1,2.12.08,1.09,2.2,2.88,1.32.5-.61,4,3.12,2.93,2.15.49,1.33,1.33,2.02.45.73.69,3.18.11-.33-1.19,1.67,2.6.43,2.2.81,1.12,1.09.02-.09-1.33h.66s.98,2.11.98,2.11l3.34.61.86,1.04-.26.44,1.51,1.49.57-.25-.15-1.39,2.39,1.75,1.04.19,1.8,3.48-1.56-.9-.53.22-.8.99.67,1.8,5.27,4.19.27,2.3,2.2,3.52-.14,2.21.99,3.54,1.91,3.19-.29,1.69,1.58-1.3,1.69,2.11ZM436.39,178.59l-.91-.7-.59-2.15,1.32-2.63-.16-1.29,2.07-2.83-.03-.94-.86-1.18,1.33-1.38.09,2.44.88,1.31-2.65,6.24-.49,3.11ZM285.13,73.8l.18-1.81,1.54-1.55,2.47-.55,2.09.34.06.66-.47.22-1.18-.23-1,.4-.51,1.46-3.18,1.06ZM416.26,154.69l-1.75-2.14-.52-1.75.48-.11,1.73,1.31.83,2.26-.77.43ZM362.66,93.42l-.75-.29-.21-1.04-1.25-1.39,1.99.61.9,1.3-.68.81ZM439.92,201.99l.29-3.84,1.17.02-.94,3.76-.52.06ZM440.12,197.51l-.72-2.02v-1.64l.93-.51-.21,4.17ZM318.08,7.23l-.68-.55-.05-.69,1.1-.52.46,1.21-.83.55ZM319.13,1.69l-.98-.67.42-.64.6-.11.61.61-.65.81ZM288.17,78.14l-.31-.65,1.11-1.04.56.4.18.83-1.28-.03-.26.49Z"/>
+      <path id="AU" class="cls-1" d="M369.46,414.76l-.9.45-.2-.45-1.4-.22-.94-.71-3.29-.32-.7-.68-1.31.4-.78-.26-1.15-1.81.64-.45,2.61.67.02-.98-.77-.64-.42.76-3.12-.54-2.62-4.71-.91-.2-.56-.9-1.15-3.62-1.02-.68-.59-5.21.2-.26,2.2,1.98.47,2.18.9-2.52-1.02-.36-1.93-2.12-.35-2.13-3.18-4.67-2.92-7.17-.73-4.09.88-.8.02-1.66.66-.44,2.43.29,1.42,1.05,1.72-.28,9.97,5.07,3.32-.1.94.53-.05-.5,1.62-1.17.57.18.24.95,1.48.27-1.3-.84-.05-.69,2.15-1.2,1.41.47,2.27-.46.66.59.78-.47,1.34-2.04.69-.17,1.77.86.69-.4.76-1.69,1.17.34,2.1,1.89.6,1.35-.69,2.46.57,2.41-.47,1.72.44,2.29-.54,2.92,1.08,5.94-.57.81-.39-.33.37-1.68-1.24-2.84-1.26,2.5-.82,5.93-.65,1.34.12,1.52-.78.81-.44,1.48.62.2.12-.68.96.98-.52,2.59.48,2.15-1.04-.66-1.19.94-1.72-2.69-.14-.71.95-1.17.82,1.67,1.28-.36-.8-1.4-2.91-1.5-.6.45.55,1.75-.29.58-1.07.35-.21-2.25-.98-.63.34.86-1.16,3.39.06,2.01-.75.19-.87-.56-.94-1.87-.46.27v1s1.36,1.25,1.36,1.25l-1.96,4.9ZM382.62,366.12l-.98-.54-.22-1.36-1.24-2.33-1.55-1.38,1.32-1.74.93-.4,2.15,2.87,1.33.92.47,3.17-2.21.79ZM336.95,364.17l-.38-.28.06-1.87-.62-.81.03-.77.19-2.27.85-.63.02-1.18.46-.13,1.42,1.24.41,3.76-.44,1.51-2,1.43ZM384.9,369.37l-.23-.97-2.52.32-.97-.42.02-.55.72-.59,1.49.19,1.52-.77,1.48,1.76-1.51,1.03ZM373.65,413.2l-1.38-.17-.57-1.19,1.92-2.37.7,1.83-.67,1.9ZM374.29,409.47l-.75-1.57.7-1.22.82,2.48-.77.31ZM381.38,401.93l.05-1.96,1.19-.43.44,1.11h-.76s-.92,1.28-.92,1.28ZM347.45,372.59l-.25-.83.65-.77.98,1.12-1.38.48Z"/>
+
+    </g>
+  </svg>`,
+  europe: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="continent" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 659.06 444.64">
+    <defs>
+      <style>
+        .cls-1 {
+          fill: #ccc;
+          stroke: #fff;
+          stroke-width: .5px;
+        }
+        .dinosaurs {
+          fill: #2a7035;
+          cursor: pointer;
+        }
+        .dinosaurs:hover {
+          fill: #194520;
+        }
+      </style>
+    </defs>
+    <g id="continent" data-name="Layer 1">
+      <g>
+        <path id="NL" class="cls-1" d="M98.74,288.33l-2.18,1.93-3.16-.41-.56-1.42,5.9-.1ZM96.88,284.51l.85.99-2.54-.66,1.69-.33ZM103.19,269.83l-1.19.39,1.19-1.65v1.26ZM98.74,288.33l-5.23-1.66,5.56.75-2.21-3.68,3.61-5.45,1.93-7.01,8.71-5.23,5.08-.38,2.57,1.79-1.11,7.05-2.17.37.26,1.7,1.93.92-2.1,3.3.51,1.23-5.73,1.8,1.68,3.82-.46,3.26-1.84,1.25,1.47,1.22-.45,1.76-2.6-.9.84-1.67.16-.73-4.99-4.4-5.42.89ZM106.15,266.29l-.91-.07,2.64-.52-1.73.59Z"/>
+        <path id="CY" class="cls-1" d="M302.43,435.97l-3.51,2.48.73,2.5-2.36.15-4.66,3.27-3.76-1.3-.89-1.84-.11-1.06,3.9-.8.41-1.73,4.49.3,5.76-1.97Z"/>
+        <path id="UA" class="cls-1" d="M327.7,332.62l-4.52.16-5.04,3.55-6.52.88-5.47,5.05,1.79-1.69-.34-1.57-2.56,2.45,1.16,4.72,2.93,3.69,7.53-.74-1.22,3.14-6.21-.32-9.42,6.39-4.2-1.22.71-5.16-7.06-2.92,7.8-5.23-.48-1.45-7.53.13-4.35-1.93,1.2-1.44-3.06-1.22,5.43.78,1.46-1.38-3.59-.26-1.92-5.61.76,5.54-2.29-.15.21-1.1-1.08,1.46-4.09.75-3.89,6.65-3.98,1.4.53,4.43-2.04-1.53-4.33,1.78-3.69-2.07,1.93-.65,3.02-5.13.07-3.97,7.9.35-1.71-3.98-2.06-1.33-.22-3.03-2.72-2.17-.06-4.75-10.62-5.18-6.27,2.21-2.11.56-.96,2.13-7.98,2.69-2.69-2.07-9.27-1.54-2.2,1.38-5.02-4.62,2.74-6.81,2.12.1-.99-5.61,9.32-9.68.04-3.64-3.3-6.92,5.1-3.8,9.59-.61,9.13,1.87,3.76,2.97,3.26-1.41,3.69,1.89,2.5-2.1,1.65,2.65,5.48-1.03,2.59,2.29,1.42-6.82,2.18-1.99,4.61-.26,2.41.55,2.11-2.81,8.76-.42,4.46,6.18-1.86,1.1.62,4.58,5.73.55,1.82,1.96,1.73,6.93,3.54-.42,3.39,2.09,5.41-2.13,4.21,5.17,1.42-1.39,12.28,4.96.18,3.36-2.84,2.5,2.14,1.91-1.42.14-1,2.21,2.11,3.27-1.21,3.83-5.92.33-3.58,2.45-1.04,5.15ZM285.92,341.31l.93.48-4.32-2.13,3.39,1.65Z"/>
+        <path id="TR" class="cls-1" d="M258.99,380.96l1.24,3.74,5.79,2.91-.68,1.97-9.82.32-8.73,7.93.34-2.1,3.63-2.76-4.63.14-.45-1.03,3.95-6.01-2.05-2.83,1.99-2.24,4.23-1.16,1.96,1.56,3.23-.44ZM349.9,385.03l6.45.7,1.93-1.25,4.62,4.22,1.91,3.45-1.03,2.1.65,3.13,4.2.76,3.22,2.95.33.46-1.55-1.03-1.33,3.02-2.47.39,2.88,9-1.61,3.64,2.55,1.68,1.18,4.8-3.25,1.38-1.13-2.73-9.03-.59-2.8,2.22-1.05-1.59-4.63,1.76-5.45-.07-9.09,3.49-7.85-1.84-5.09,2.16-5.24-1.33-.14,4.75-3.44,3.34-1.58-.7-.55-3.27,2.55-2.92-.94-2.11-4.42,2.81-4.65-2.02-6.79,5.3-6.07,1.21-10.47-6.55-4.01-.37-1.34,4.98-5.1.94-3.13-1.4-1.71-3.26-4.49-.81-1.29,1.39-3.85-.32,4.73-2.9-6.6.45,1.83-1.58-3.14-4.18,1.1-2.73-6.34-2.55,1.01-3.12,1.57,2.62,3.17-.99-2.56-2.22,1.68-1.52-2.24-3.51,1.47-2.24-5.3.72.46-4.57,3.75-3.61,6.67.63.99-1.62.59,1.26,7.04-.18-1.48-1.27,1.15-.85,5.97-.96-4.92-1.77.23-2.52,14.19,1.01,7.08-5.58,6.59-2.48,9.88.43,1.72-.96,1.97,3.02,5.08.42,2.38,3.66,2.51-.8,10.8,3.92,7.04-1.62,5.65,1.29,8.38-4.97ZM245.22,397.3h-2.03l1.68-.9.35.9Z"/>
+        <path id="SE" class="cls-1" d="M233,98.8l-3.13-.39-3.97,1.53-2.4-2.23-1.4,1.27-.92-1.12-.99,4.34-1.13-.22-1.4,3.04-2.11.25.29,2.37-1.34-.9,1.11,3.08-2.94,5.05,2.57,5.43-5.1,9.21-7.01,6.14-2.48-.21-2.15,3.6-2.88.89-1.98,2.7,1.27.11-.26,1.38-1.97,1.57-1.78-1.44.88,4.21-1.36,1.43-2.86-.08,1.5,4-1.74,5.27.61,2.59-1.81-.57-.45,2.11.81,12.21,2.77,2.27,1.98-.75,6.04,7.67.8,3.6-6.78,5.29-1.29-.72-.09-2.74-.49,1.05-2-.57-.12,1.18-5.38-1.54-3.57,1.77,5.86.43,2.64,2.29,8.11-1.79.73,1.07-11.04,8.79-5.15.23,4.78,1.86-1.83.75v11.85l-2.04,9.81-3.52,7.11-7.04-.45-.18,1.56-2.38.68-1.54,6.92-8.67-.17.59-4.01-3.38-6.53,2.22.32-.98-2.15,1.36-.14.17-2.02-3.11-3.54-6.55-15.19-.12-2.84-1.23-.37-.68-8.03.32-1.17,1.3.55.56,1.66,1.15-.22,1.05-4.77-.79-4.01,1.69-3.62,3.74-3.27.69-4.67-1.99-7.61,2.78-.79,1.17-4.09-4.88-5.21,1-8.1-2.05-14.82,1.18-4.59,4.16-6.16,2.77-1.16,5.1.94,1.22-2.45-.43-4.52-2.88-1.84,5.59-11.45.42-13.57,6.34-2.95-.41-3.09,6.61-9.66-1.86-6.46,3-3.41,1.42-4.9,3.64-3.76,3.99,2.5,2.6-10.74,11.23,3.65,1.82-2.22-1.83-1.18,2.56-5.69-1.56-3.23,3.4-.3,9.27,9.61,5.77,2.81,5.29,7.48-1.24,8.78,1.87.66-.62,5.12,2.34,5.49-1.94,9.44,3.07,7.42ZM198.78,213.19l-1.77,1.64.64,3.86-5.13,5.93.93-2.01-1-5.9,2.7-3.45,2.45-1.08,1.18,1.01ZM181.62,232.33l-.79-3.39,4.13-9.58-3.34,12.97Z"/>
+        <path id="SI" class="cls-1" d="M181.53,338.43l-1.31-.34-.63,1.58-3.99,1.67.11,3.28-2.52,1.25.42,2.55-3.68-.11-1.51-1.72-1.34,1.69-5.34-.34.96-.68,1.04-.26-2.66-3.8,1.03-1.63-1.71-.82,2.16-2.52,5.73,1.17,2.31-2.01,7.17-.7.91-1.83,1.29.06,1.56,3.51Z"/>
+        <path id="SK" class="cls-1" d="M222.1,312.67l-2.74,6.81-2.76.59-1.82-2.08-6.48.25-3.63,3.85-2.19-.77-5.62,2.24-.45,2.14-6.48.17-4.14-2.36-1.9-3.85.59-2.15,1.23-2.47,4.19-.48,2.71-3.79,4.53-2.62,2.13,1.15,1.97-2.05,2.24,2.34-.12,1.73,2.03.23,2.06-2.1,8.6-.27,6.05,3.49Z"/>
+        <path id="RS" class="cls-1" d="M206.63,342.23l3.59,3.47.13,2.71,4.69,3.06-.92,1.32,1.21,1.03-1.17.7,4.94,2.7,3.7-1.03.62.76-1.61,1.27,1.42,1.86-2.26,4.26,4.09,5.5-3.44,3.17.38,3.67-1.21,1.16-5.26.6,1.28-3.84-2.44-.75-3.97-4.69-3.08,3.99-7.74-6.5,2.02-1.02-1.68-3,2.28-.73-3.14-2.97,1.55-4.92-2.3.11.38-2.55,2.27-.72-2.66-1.78-.67-5.12,4.77-2.3,4.23.58Z"/>
+        <path id="RO" class="cls-1 dinosaurs" d="M260.33,348.58l3.69,2.07,4.33-1.78,2.04,1.53-1,3.96-3.43.82.31-2.06-1.37.53-2.06,11.06-10.1-3.97-5.86,1.49-4.85,3.14-17.36-1.52.73-2.28-2.17-1.49-1.42-1.86,1.61-1.27-.62-.76-3.7,1.03-4.94-2.7,1.17-.7-1.21-1.03.92-1.32-4.69-3.06-.13-2.71-3.59-3.47,5.93-1.68,5.91-12.03,5.91-4.42,2.2-1.38,9.27,1.54,2.69,2.07,7.98-2.69.96-2.13,2.11-.56,2.44,1.05,7.35,11.73,1.13,3.32-1.11,10.1.93,1.43Z"/>
+        <path id="PT" class="cls-1 dinosaurs" d="M20.38,422.74l-2.88,1.46-7.84-.22,1.24-3.36-.45-8.66,1.43.19-.88-.8-2.79.6-.25-1.79,1.54-.78,1.55-2.86-2.32,2.9-1.49.38-.83-.86,5.36-17.08-.62-10.64,3.44-1.78.76,2.94,10.34-1.18.51,2.44,2.22,1.26-4.82,4.67.8,5.91-1.5,1.54.38,3.24-3.77,1.19,3.62,5.26-2.32,5.16,2.59,2.31-3.27,3.91.25,4.65Z"/>
+        <path id="PL" class="cls-1" d="M202.34,254.02l21.3,1.19,4.07,1.8.77,2.99,2.9,13.19-4.99,5.34,3.22,2.7-.32,5.69,3.3,6.92-.04,3.64-9.32,9.68.99,5.61-2.12-.1-6.05-3.49-8.6.27-2.06,2.1-2.03-.23.12-1.73-2.24-2.34-1.97,2.05-2.13-1.15-1.82-3.83-6.3-2.48.5-2.01-1.93.56-3.6-1.82.73,2-2.36,1.42-2.89-3.38,1.41-1.59-.92-.87-1.85.47-6.83-4.29-1.24,1.66,1.39-4.22-2.79-6.27,1.01-2.72-.89-4.91-3.31-3.89,1.9-3.59-1.02-5.99,2.19,1.02-.17-2.09-2.33-.54v-.91s13.29-3.91,13.29-3.91l2.52-3.04,8.64-3.06,3.24-.25,2.93,1.79-2.17-.7,2.69,4.35,5.18-1.04Z"/>
+        <path id="NO" class="cls-1" d="M242.63.29l2.42.78,1.27,2.26-3.71.72-1.8-1.91,1.82-1.85ZM278.23,27.68l-.06,4.75-4.59-1.88-.62,3.91-4.71,2.58-2.85,5.23-.8-2.93,3.28-5.65-1.3-3.83-9.38-7.71-4.18,3.12-4.06-.17-3.05,4.37-2.18,13.4-5.44,7.39-7.32-3.94-3.58,2.91-6.15-1.31-5.5-10.42-3.56,1.14v3.26l-2.99.09-3.4.3,1.56,3.23-2.56,5.69,1.83,1.18-1.82,2.22-11.23-3.65-2.6,10.74-3.99-2.5-3.64,3.76-1.42,4.9-3,3.41,1.86,6.46-6.61,9.66.41,3.09-6.34,2.95-.42,13.57-5.59,11.45,2.88,1.84.43,4.52-1.22,2.45-5.1-.94-2.77,1.16-4.16,6.16-1.18,4.59,2.05,14.82-1,8.1,4.88,5.21-1.17,4.09-2.78.79,1.99,7.61-.69,4.67-3.74,3.27-1.69,3.62.79,4.01-1.05,4.77-1.15.22-.56-1.66-3.73-1.94-1.61-7.69-2.8,9.98-2.27.67-1.92-2.02.67,1.85-10.05,10.66-4.71,1.59-3.11-.05-.86-1.61-1.93.68.47-2.11-4.08-1.44-3.62-4.53.64-3.73,3.28,1.86,1.78-1.71-1.78.64-1.41-1.91.42-2.64,3.04-3.47-7.01,5.22-1.56-.8,1.17-5.5,3.15-.25,2.99-2.1-3.27-.6,2.76-5,2.54-2.37.06,3.46,3.16-4.88-5.67,2.26-6.79,9.52.4-6.02,3.25-.47-2.71-1.12-1-3.26,3.44-3.33-2.72,1.63-.86-.91-.73-5.56,11.92-1.45,1.76,2.63v-1.92s3.81-1.66,3.81-1.66l-1.35-.17-.39-1.09.65-1.89-.75.88-.37,2.68-2.25.29-2.31-1.81-1.45,2.18-7.14.36-2.04-1.99-.13-2.57,2.27-.73-2.77-3.19v-2.39s3.62-.27,3.62-.27l3.71,1.56,4.82-1.17-9.87-.94-.83-3.22,1.45.11,3.7-3.82,1.18.96,3.35-.84.76-.88-3.75.88,1.46-2.97,8.2.92.81-.55-1.02-1.26,3.75-.88-9.17.16,1.39-3.09,4.25-2.51,3.56.13,3.52,3.62-3.13-4.68.76-1.86,2.46-.84-1.85-2.36,1.45-1.54,3.77.11.13,2.03,3.65-2.5,2.18,3.53,4.98-1.06-.24-2.48,4.35-2.73-1.31-1.44,1.9-1.6-1.02-.7-2.64,1.96.14,2.31-5.93,3.9-3.29-2.93,6.73-10.97,7.19-6.21-2.27.93,1.3-3.5,4.52-3.25.99,1.49,2.91-2.06,1.19-2.01-3.36,2.5-1.99-2.88,3.82-8.81,2.32-.89-1.68-2.39,6-1.33,2.42-1.97-6.16,1.12.62-6.91,2.76-2.62h2.28l-2.07-1.93,3.08-3.66,8.8-1.44-6.56-1.2,3.5-5.32,4.26,3.97.66-3.04-2.98-1.42.37-2.9-3.03,1.81-.37-2.59,2.26-2.93,3.28.44-2.06-2.17,4.66-2.89,2.05,6.26.54-3.78-1.27-4.09,1.24-1.33,3.8.64,4.05-1.31-.86-1.02-5.66.29-.48-1.22,5.91-4.94,2.11-5.64,2.67-1.07,1.07-5.99,4.04,2.99-1.65-3.38,2.54-1.36,1.39-3.62,3.29-1.11-.3,7.37,2.15-7.75,2.45-2.37.09,6.41-1.55,5.3,2.55-3.86,1.7.42-1.39-3.33.6-4.32,3.65.48,1.82-2.44,3.65,3.51-1.16-4.55-3.01-3.32,5.81-1.51.69.89,2.45-2.21,4.51,7.78.17-5.23,4.44-4.77,2.57-4.4-1.06-2.54,3.38-3.76,3.37,3.41,1.15-1.41,2.24,1.21-5.21,12.84.33,2.17,1.13-.54,9.77-16.26-.52,10.79,2.73-2.04,1.31-4.73,2.44-1.22-2.1-2.95,2.44-2.99,5.36,2.41-.44,3.09-2.89,3.01,2.52.21-.54,8.45,4.31-12.46,1.82.06,4.28,4.38,2.21-1.15.99,3.25,2.58.39,2.23,2.47.12,2.54-4.59,2.98-9.98-.41,5.52,3.33.63,4.55,2.65.52,1.01-2.82,1.28,2.73.38-1.41,2.6.22ZM228.18,7.04l-4.12,5.58-3.17.54-2.45-2.88,6.53-1.1,1.62-2.14h1.59ZM232.07,12.1l-2.03.11.42-3.77,2.02,1.97-.41,1.69ZM229.36,12.46l-.46,2.85-2.62,2.52-1.46-3.25,4.08-3.49.46,1.37ZM203.44,19.14l2.13,1.75-.53,1.04-1.51-.02-1.23-3.75,1.14.98ZM199.98,22.12l2.38.94-1.84,3.9-2.2,1.18-1.5,3.51-4.42.41,1.48-4.08,2.19-.27.08-2.13,2.46-2.89.55-4.09.82,3.52ZM210.25,21.66l-2.12.26.19-2.51,1.98-.32-.05,2.57ZM272.07,27.42l-1.28.57.47-2.7,1.48,1.32-.67.81ZM188.18,31.31l3.38,1.76-.36,5.87-3.12.02-2.73,3.46-1.83-1.07,1.29-5.53,2.41-.37-.88-1.85,1.84-2.29ZM176.44,50.85l3.83-5.85,1.28,4.53-3.66,4.24-3.62,1.64-1.07-1.12-4.38,3.2-2.5.14,2.21-3.82,3.43-.76,2.14-3.21.47-7.98,3.81-4.9-1.94,13.89ZM172.72,43.75l1.27,2.98-1.17,3.1-5.51-.86,1-2.9,1.68.52.47-2.29,1.11.36v-1.99s1.15,1.08,1.15,1.08ZM163.73,56.27l1.51.85-5.85,4.03.48-2.97,3.86-1.91ZM157.66,63.33l-1,.94,1.65-4.48-.65,3.54ZM153.94,94.86l.72-2.78.65,1.47-1.37,1.31ZM154.54,97.2l-.22-1.24,1.82-.57-1.6,1.81ZM150.9,101.72l-1.28.36.15-1.29,1.15-.29-.02,1.22ZM145.93,113.96l-3.31-.07,1.9-1.72,1.41,1.79ZM127.34,132.56l-1.24-.3,3.22-1.79.14,1.54-2.12.55ZM124.86,137.54l-2.01-1.15,1.81-.85.2,2ZM103.68,170.1l-1.08.03.42-1.55.66,1.52ZM104.53,180.8l.03,1.61-.89-3.52.86,1.91Z"/>
+        <path id="ME" class="cls-1" d="M199.58,366.65l7.74,6.5-1.95.62.06,1.94-1.85.65-.9-1.39-2.52,4.15.41,2.75-5.55-5.11-.55-1.16.16-4.01,2.63-3.21,1.19.5-.52-2.17,1.65-.06Z"/>
+        <path id="MK" class="cls-1" d="M220.79,377.84l4.45,5.19-.59,3.62-2.11,1.76-4.11.08-2.81,2.33-4.12.17-3.2-3.77.52-5.39,1.42-1.79,5.29-1.6,5.26-.6Z"/>
+        <path id="MD" class="cls-1" d="M260.33,348.58l-.93-1.43,1.11-10.1-1.13-3.32-7.35-11.73-2.44-1.05,6.27-2.21,10.62,5.18.06,4.75,2.72,2.17.22,3.03,2.06,1.33,1.71,3.98-7.9-.35-.07,3.97-3.02,5.13-1.93.65Z"/>
+        <path id="LV" class="cls-1" d="M254.53,217.07l3.21,2.93-1.27,5.54,1.41-.09,1.71,3.77.3,4.89-3.85,4.14-3.53-.38-3.09,1.95-6.27-5.25-3.99-1.15-1.55-2.56-4.85,1.79-13.71-1.73-7,4.07.17-9.18,4.43-9.28,5.56-1.93,4.94,7.93,2.43,1.47,4.95-3.46-.41-7.79,6.47-2.26,4.77,2.67,3.17,3.7,3.4-.81,2.6,1.02Z"/>
+        <path id="LU" class="cls-1" d="M111.48,301.79l2.5,3.37-.97,3.59-3.73-.88-.31-3.97,1.56-2.59.95.48Z"/>
+        <path id="LT" class="cls-1" d="M249.42,239.82l-.92,3.87,2.15.82-6.18,4.16-2.1,6.83,1.36,2.01-1.61-.03-.33-1.53-7.7,4.58-5.61-.53-.77-2.99-4.07-1.8.39-5.98-1.73-2.2-8.97-2.42-1.28-9.62,7-4.07,13.71,1.73,4.85-1.79,1.55,2.56,3.99,1.15,6.27,5.25ZM211.45,244.44l-.39-.09,1.45-3.92-1.06,4.01Z"/>
+        <path id="LI" class="cls-1" d="M134.81,332.95l-.62-.05.27-2.06.35,2.11Z"/>
+        <path id="KV" class="cls-1" d="M215.53,378.44l-5.29,1.6-1.42,1.79-.55-3.17-2.84-2.95-.06-1.94,1.95-.62,3.08-3.99,3.97,4.69,2.44.75-1.28,3.84Z"/>
+        <path id="JE" class="cls-1" d="M56.68,311.04l-1.47.56.11-.92,1.36.36Z"/>
+        <path id="IE" class="cls-1" d="M21.65,246.64l-2.24,3.8-2.43.81,1.06,1.2-2.45,2.09,3.44,3.12,1.9.12,2.14-3.15,2.41,4.01,2.91-.35-.88,1.69,2.16,11.46-2.94,6.45.94,1.08-3.81.96-.51-.99-7.36,4.64-2.36-.69v1.92s-5.98,2.32-5.98,2.32l-3.63.16,1.72-2.23-3.64.96,3.51-2.97-5,.82,2.91-3.54-3.01-.92,3.93-.47-.9-1.7,1.85-1.58,5.71-1.47-1.39-.84-1.25,1.34-4.99.72,4.16-5.88,2.49-1.23-7.83-2.31-.17-1.54,2.67-.63-1.22-1.4,2.18-.89-2.26-.66.39-2.65-1.6-.7.25-1.17,10.18.19-.53-1.22,3.3-3.4-4.24-.47,2.6-2.43.69-3.02,4.09-1.3.06,3.36,2.36-4.65,2.34,1.51-1.73,1.72ZM3.26,260.3l-2.14-.73,1.81-.3.33,1.03Z"/>
+        <path id="IM" class="cls-1" d="M40.55,257.18l-2.51,1.29,2.43-3.84.08,2.55Z"/>
+        <path id="HU" class="cls-1" d="M219.36,319.48l5.02,4.62-5.91,4.42-5.91,12.03-5.93,1.68-4.23-.58-4.77,2.3-3.69,1.72-3.71-.36-8.7-6.88-1.56-3.51-1.29-.06,2.43-1.41-.07-3.89,1.57-1.35-1.71-1.38,4.34-.33.55-2.99,4.14,2.36,6.48-.17.45-2.14,5.62-2.24,2.19.77,3.63-3.85,6.48-.25,1.82,2.08,2.76-.59Z"/>
+        <path id="HR" class="cls-1" d="M181.53,338.43l8.7,6.88,3.71.36,3.69-1.72.67,5.12,2.66,1.78-2.27.72-.38,2.55-2.32-1.98h-5.72l-6.02-1.9-4.22,2.55-3.4-1.62-.35,3.93,3.8,6.04,6.56,6.32,2.1,4.68-4.6-4.19-6.18-1.18-5.39-6.08-.42-.79,2.34-.14-3.3-3.12-.85-4.53-2.05-2.07-1.6-.38-3.04,4.76-1.91-6.48,5.34.34,1.34-1.69,1.51,1.72,3.68.11-.42-2.55,2.52-1.25-.11-3.28,3.99-1.67.63-1.58,1.31.34ZM170.04,353.1l-2.42-.98.81-1.38,1.61,2.36ZM189.29,372.51l5.18,3.09.55,1.16-9.92-5.33,4.19,1.08ZM188.89,373.69l.92.63-2.7-.83,1.78.2Z"/>
+        <path id="GR" class="cls-1" d="M247.58,383.24l2.05,2.83-3.95,6.01-6.29-2.38-9.04,2.2.7,2.91,3.21,2.39-4.14-1.61,1.47,3.2-3.51-2.62,1.36,2.99-4.92-4.19.17-1.68-1.97.84-.26,4.04,4.96,7.52-1.17.63.05-1.35-1.62-.43.98,2.33-3.35,1.47,4.61,1.78,5.19,4.47-.03,3.93-3.49-3.03-3.13,1.33,3.05,3.72-2.21.89-2.94-1.76,2.93,9.2-2.98-2.89-1.95,2.66-2.34-4.64-1.27,2.45-2.08-2.88.64-2.59-3.73-4.27,1.88-2.61,2.84-1.12,7.38,3.16,1.54-1.86-4.9-2.25-6.4,1.01-.95-1.43-1,1.22-2.79-4.55,2.31-.19.05-1.15-2.73-.04-4.8-5.86,1.66.27.43-2.64,2.33-1.21,2.52-4.45-.45-2.02,4.12-.17,2.81-2.33,4.11-.08,2.11-1.76,10.59-1.96,5.14,2.79,4.53-.61,1.42-.66-.33-2.86,1.58-.11ZM237.16,393.07l-1.74-.28.73-1.29,1.01,1.57ZM241.63,398.64l-.54,1.53-1.99-.38-.02-1.29,2.55.14ZM205.53,403.45l-2.91-2.92,1.88-.06,1.03,2.98ZM248.18,404.34l1.25,2.44-.85.65-2.43-.93,1.11-1.01-1.35.88-1.54-.9,3.81-1.13ZM228.01,407.56l4.8,2.67,2.75,5.72-3.34-3.5-7.88-4.13,3.67-.76ZM209.63,410.57l-.89.23.94-2.26-.05,2.03ZM246.06,413.92l-1.37-.22-.31-2.83,2.12.29-.44,2.76ZM209.13,412.51l1,2.67-2.75-.93,1.75-1.74ZM210.99,417.44l.71.83-1.18.36-1.34-1.61,1.81.42ZM250.98,417.39l1.55.87-3.19-.13,1.64-.74ZM242.36,424.52l-1.24-.86,1.1-1.07.14,1.93ZM257.84,433.19l-.86-2.01,3.48-2.19-2.62,4.2ZM230.95,436.45l2.12-.5,1.26,1.95,9.27.09.42,1.85,3.56-1.58-1.04,2.44-9.2.69-2.27-1.85-5.86-.8-.16-2.29,1.13-.99.77.99Z"/>
+        <path id="GG" class="cls-1" d="M53.35,308.32l-.24.68-.61-.23.85-.45Z"/>
+        <path id="GE" class="cls-1" d="M383.05,381.67l-1.67,2.11,3.3,3.32-.93,1.77-8.44-3.23-1.89,1.42-10.52,1.64-4.62-4.22-1.93,1.25-6.45-.7,1.71-4.07-1.85-6.27-10.17-6.99,1.15-1.39,3.36.33,6.28,2.91,7.95.46,6.88,3.88.3,1.6,4.58-1.62,1.78,1.21.68-1.28,5.62,2.36-.45,2.67,5.33,2.84Z"/>
+        <path id="FR" class="cls-1 dinosaurs" d="M109.28,307.87l3.73.88,2.64,3.02,4.81.08,4.61,1.84-3.49,8.3v5.65l-1.84,1.59-2.52-.2-6.73,12.17,1.55.22,1.54-2.32,2.22.15,1.77,4.75-1.55,1.79,2.39,3.45-3.44,2.99,2.41,2.29-.62,4.64,2.81,1.85,2.15-.25-.97,3.71-9.28,6.43-7.11-3.43-2.34.66-5.4-1.76-5.79,5.98,1.08,4.42-7.94.71-2.2-1.37-1.87-.84-4.93-2.28-.44,1.42-8.21-1-5.89-2.32.34-1.73-2.58-1.55,2.08-1.45,2.75-10.54-1.13.22,1.1-8.25,3.59,5.08-1.63-4.46-2.73-2.37,1.1-.25-.77-5.53-4.31-1.98-1.84-2.89-.93-3.48,1.15-1.1,1.92.58-1.57-.94-3.56-.02.51-1.57-2.31-.43-.11-1.11-10.27-1.98-2.47-2.17,2.35-1.32-1.67-1.21,2.26-.14-3.22-.6v-1.79s10.03-3.07,10.03-3.07l3.63,3.1,1.66-1.13,2.98.67.66-1.17,3.56.45-3.23-10.6,4.03.03.8,3.03,6.57.94,3.91-1.56-1.93-.62.38-2.02,7.13-3.08,2.34-2.67.54-6.7,5.74-2.27,1.58,3.7,2.34-.3,3.93,5.01,3.27.62-.17,2.89,2.66.11,1.84-2.02.33,3.82,4.31,2.88,1.91-.28ZM62.34,344.21l-.24.85-1.18-2.26,1.42,1.41ZM134.14,373.35l.47,6.16-2.45,6.71-2.55-1.83-1.63-6.95,1.67-2.28,3.36-.96.34-2.79.79,1.94Z"/>
+        <path id="FO" class="cls-1" d="M27.12,153.44l-.93.77-.07-2.18,1,1.41ZM25.6,153.88l-.16,1.94-1.25-.37.78,2.41-3.01-4.81,3.64.83ZM21.86,155.16l.47,1.33-1.77-.4-.29-.94h1.59Z"/>
+        <path id="FI" class="cls-1" d="M265.4,42.27l-3.72,2.21,2.42,1.2-1.98,5.61.55,3.41.84,2.94,4.44,2.32,4.34,7.04-6.21,13.55,6.88,18.55.05,1.72-2.56.94-.72,6.09,1.47,1.66-1.5,2.82,3.41,3.74-.83,3.28,3.37,4.48.26,2.46-3.6,5.24,8.01,7.95,2.38,4.79-4.03,8.23-21.14,25.18-2.26.97-1.73-1.07-4.52,1.78.24-2.89-1.29,2.74-2.84-.69.53,1.81-1.87,1.01-.68-.89-8.16,4.22-6.61.47-2.98,2.29,1.19-2.77-1.93-2.54-1.09,2.07-1.94.37.83-4.76-7.74-2.95-.51-5.12,1.65-8.74-2.36-5.68.59-4.14-1.61-5.03,2.49-6.06,1.19-.09-.71-2.45,5.22-1.59v-2.92s8.62-8.18,8.62-8.18l6.46-11.86,4.92-.94.57-2.37-.17-7.56-4.53-3.12-.32-3.09-3.18.89-3.07-7.42,1.94-9.44-2.34-5.49.62-5.12-1.87-.66,1.24-8.78-5.29-7.48-5.77-2.81-9.27-9.61,2.99-.09v-3.26l3.56-1.14,5.5,10.42,6.15,1.31,3.58-2.91,7.32,3.94,5.44-7.39,2.18-13.4,3.05-4.37,4.06.17,4.18-3.12,9.38,7.71,1.3,3.83-3.28,5.65.8,2.93ZM237.66,111.97l-1.81.2-.02-1.02,2.66-.2-.83,1.02ZM213.21,138.98l1.33.66-1.09.68-1.15-1.88.91.54ZM214.78,177.77l-1.02.68-.51-1.93,1.53,1.25ZM219.66,179.94l1.61.91-.37,1.87-1.9-1.64.66-1.14ZM217.35,183.06l-.93.35.47-1.14.46.79Z"/>
+        <path id="EE" class="cls-1" d="M258.98,191.85l.93,1.45-4.83,7.7.46,7.29,1.86,4.46-2.87,4.32-2.6-1.02-3.4.81-3.17-3.7-4.77-2.67-6.47,2.26,1.39-6.16-2.79,1.13-2.34-1.22-1.74-3.83,1.16-1.67-1.68-1.74.42-3.6,3.97-1.27,2-2.39,7.16-.64.45-1.57,9.87,2.51,6.18.48.81-.93ZM224.7,200.48l-2.57,1.78-.89-2.25-2.39-1.05,4-1.87,1.75,1.25.1,2.14ZM227.53,204.07l-1.58-1.41,1.5.14.08,1.27ZM222.63,203.15l4.76,2.2-6.41,2.99-2.53,3.64,1.29-2.83-2.06-1.38-.13-3.02,5.08-1.6Z"/>
+        <path id="ES" class="cls-1 dinosaurs" d="M58.19,367.81l2.58,1.55-.34,1.73,5.89,2.32,8.21,1,.44-1.42,4.93,2.28.13,1.44,1.74-.6,2.2,1.37,7.94-.71.24,4.42-1.63,1.59-15.43,8.45,1.19.89-1.99.96-6.22,9.61.82,3.97,2.74,2.62-4.86,3.79-1.98,4.67.62,1.18-6.19,2.07-3.17,5.13-15.19.49-5.42,2.47-3.06,3.31-2.8-1.35-2.31-3.75,1.12-2.32-1.2.69-3.15-3.76-3.66.84-.25-4.65,3.27-3.91-2.59-2.31,2.32-5.16-3.62-5.26,3.77-1.19-.38-3.24,1.5-1.54-.8-5.91,4.82-4.67-2.22-1.26-.51-2.44-10.34,1.18-.76-2.94-3.44,1.78.31-4.27-.54-2.09-1.5.43.72-1.87-2.08-1.64.39-1.81,9.97-5.47,4.29,1.96,8.18-.85,8.91,2.13,6.19-.97,3.77,1.37,8.43-.33ZM99.2,399.88l-2.97-1.94,2.51.27.46,1.67ZM91.46,400.33l2.13.81-2.62,3.45-4.73-2.36,5.31-3.48-.09,1.58ZM80.01,407.9l-1.5.13,2.3-1.88-.8,1.75Z"/>
+        <path id="GB" class="cls-1 dinosaurs" d="M63.1,295.78l-1.26,1.08-2.1-.82,1.69-1.14,1.67.88ZM42,267.02l1,.17-2.19,1.93-1.31-2.84,2.5.74ZM28.39,258.29l-2.91.35-2.41-4.01-2.14,3.15-1.9-.12-3.44-3.12,2.45-2.09-1.06-1.2,2.43-.81,2.24-3.8,5-1.76,2.96,1.14,2.15,3.84-1.09,2.05,2-.26.75,1.9-1.35-.58.43,3.21-2.77,2.54-1.34-.43ZM35.88,242.42l-1.52-.38.09-2.72,1.43,3.1ZM28.99,236.67l.49,2.82-1.68,1.05.13-1.97-1.38.9.19-1.33,2.25-1.47ZM30.06,238.06l-.69-.94,2.34-2.71-1.65,3.65ZM31.35,231.67l-3.61.62,1.18-2.4-.99-1.48,3.54,1.48-.12,1.78ZM21.44,222.21l-1.09-3.3.97.11.12,3.19ZM28.88,217.36l.06,2.38,3.12.77-1.86,2.57-.58-1.93-1.94-.02-2.95-2.99,2.72-2.82,1.43,2.04ZM21.73,215.12l.16,1.88-2.24-.86,2.08-1.02ZM28.52,206.47l-2.4,3.46.87.91-3.76,3.44-.67-.81,1.52-1.39-1.54-3.29,2.42-.09-.34-1.44,3.64-2.58.26,1.79ZM49.33,204.51l-.7,2.5-5.23,4.62-.31,1.36,1.2.42-1.87,3.03,4.93-1.64,8.95.08,1.98,2.86-5.47,11.29-4.83,2.41,2.86-.41,1.42,1.75-4.64,2.72-2.87-.81,4.99,1.73,3.02-.91,6.36,5.47,2.85,10.2,7.74,6.78-.82,1.1,2.16,4.7-2.6-1.45-2.62.15,2.47.36,3.8,4.04.57,1.97-2.09,2.85,1.58,1.08,1.88-1.77,3.34.09,4.05,2.29-.44,7-2.41,1.62-.3,1.83-2.94.8.94,1.72-3.15,1.14,6.68,1.1-.12,1.95-2.95,2.75-5.09,1.73-10.92-1.43-.67,1.6-3.47.23-.03,1.3-6.49-1.21-2.73.9-1.85,4.14-3.47-1.61-3.6,1.08-2.63,2.65-3.39-.13,9.66-12.07,7.09-.17,4.73-5.79-5.79,3.79-4.03-2.17-2.31.24.96-.98-1.98-.88-3.48,1.24-1.79-1.24-.1-2.27,6.5-3.6,1.6-2.91-.82-4.15-3.91,1.22,2.79-3.79,3.4-1.83,4.49.53.22-1.87,2.12,1.31-2.12-2.29,1.47-7.09-2.15.08-2.87-5.04.85-2.42,2.89-2.11-3.46.07-2.75,1.94-5.79-.76-.63,1.83-.82-.84-.94-2.62,3.34-6.09-1.31-4.45,1.93-.77-1.75-1.36.3-1.29-2.89,3.27.97-3.75-2.19,2.44-.99,7.24-1.17.65,1.32-10.97,2.33-6.19-3.12,2.77-3.25-2.14,2.72-1.8-.88-.61,2.02-4.08-1.57-1.82,1.43-2.11-1.1-1.52.92-2.69h3.04l-1.72-2.4.5-2.15,2.23-.31-.06-3.9,3.93.69,7.91-1.77,1.39.19-.38,1.54ZM49.68,197.84l1.78,1.6-3.63-.84.14-2.09,1.71,1.33ZM61.46,177.66l1.72,1.27-1.66,7.66.06-3.7-2.51-1.75,1.94-.69-1.32-2.2,1.4-1.58.37.99Z"/>
+        <path id="DK" class="cls-1" d="M135.89,249.76l-7.21-.91-.37-6.06-3.25-2.16.21-12.16,3.42,1.36,1.46-2.93,1.2-.72.87,1.13.4-3.81-1.77-.06-1.48,3.58-2.05.75-1.24-2.31,2.25-3.2,5.49-.78,3.56-5.08,4.37-1.96-2.21,13.87,4.34,2.16-1.17,2.45-2.92.35-.92,4.19-1.88.27.81.97-2.92,3.2.54,2.68-1.46,2.68,1.88.84.05,1.66ZM144.73,220.51l-1.21-.12,2.03-.76-.82.88ZM154.94,238.42l-2.38,3.79,1.33,2.14-2.17,1.16-.27,4.37-1.27.5-1.4-4.86-2.48-.2-2.07-6.14,2.32-.37,2.05-2.45,1.3,3.1.31-3.24,2.38-1.81,2.42.66-.07,3.35ZM155.6,240.67l-.78.48.33-1.48.45,1ZM141.98,240.51l.95,5.64-2.31,1-3.06-1.35-.86-4.17,5.28-1.12ZM171.91,247.46l-2.72-.94.55-2.29,2.47,1.79-.3,1.44ZM142.58,250.64l-.75-1.18,2.22-3.58-1.47,4.76ZM138.05,249.05l-1.72-.23-.17-1.91,1.89,2.14ZM154.81,248.12l-2.9.6,1.05-1.75,1.85,1.15ZM146.81,248.99l2.55.99.17,1.49-2.07.59-2.85-1.68.16-1.96,2.04.57Z"/>
+        <path id="DE" class="cls-1 dinosaurs" d="M166.01,259.88v.91s-1.93-.1-1.93-.1l-.66-2.84,2.59,2.03ZM162.63,254.91l-.02,1.17-3.48-.51.99-4.31,2.16,1.6.35,2.05ZM135.89,249.76l1.9,1.77-1.04,2.34,5.82,1.81,1.9-.73-1.07,4.25,3.67.74,7.92-6.01,3.05.65,2.83,3.11,1.86-.14.95,3.44,2.65,1.41,1.02,5.99-1.9,3.59,3.31,3.89.89,4.91-1.01,2.72,2.79,6.27-1.39,4.22-3.3-1.91.33,1.48-9.61,5.23-2.65.14-1.83,2.23-1.26-1.27,2.84,4.26-.82,1.63,1.96,3.37,7.64,6.65-.62,2.29-1.59-.41-.76,2.24-4.14,2.57,1.71,6.29-5.42-2.39-7.86,3.24-4.06-1.58-1.73,2.71-1.42-2.25-3.02-.19-6.41-2.5-.8,1.79-5.64.03v-5.65l3.49-8.3-4.61-1.84-4.81-.08-2.64-3.02.97-3.59-2.5-3.37,1.67-2.05-2.4-4.6.45-1.76-1.47-1.22,1.84-1.25.46-3.26-1.68-3.82,5.73-1.8-.51-1.23,2.1-3.3-1.93-.92-.26-1.7,2.17-.37,1.11-7.05-.83-2.21,1.42-2.31,4.88-.1.67,2.53,1.51-1.57,1.09,2.4.83-5.46,3.96.22,3.89,3.42-3.18-3.46-2.25-.77-.48-3.83-1.76-.4,2.1-2.81-1.93-4.26,7.21.91ZM126.24,250.21l.97-3.14-.49,1.82,1.69.1-2.17,1.22Z"/>
+        <path id="CZ" class="cls-1" d="M170.04,293.99l1.24-1.66,6.83,4.29,1.85-.47.92.87-1.41,1.59,2.89,3.38,2.36-1.42-.73-2,3.6,1.82,1.93-.56-.5,2.01,6.3,2.48,1.82,3.83-4.53,2.62-2.71,3.79-4.19.48-1.23,2.47-2.76-2.01-3.28.42-6.67-2.48-2.53,4.07-4.32-.04-1.58-1.67-7.64-6.65-1.96-3.37.82-1.63-2.84-4.26,1.26,1.27,1.83-2.23,2.65-.14,9.61-5.23-.33-1.48,3.3,1.91Z"/>
+        <path id="CH" class="cls-1 dinosaurs" d="M134.43,328.32l.03,2.52-.27,2.06.62.05,3.73,2.03,1.45-1.31.7,1.18-.15,3.09-2.32-.49.28,3.53-5.85-2.31-1.59,6.11-3.81-3.88-.24-1.95-4.27,5.09-5.17-.04-1.77-4.75-2.22-.15-1.54,2.32-1.55-.22,6.73-12.17,2.52.2,1.84-1.59,5.64-.03.8-1.79,6.41,2.5Z"/>
+        <path id="BY" class="cls-1" d="M259.89,234.11l4.36,2.42,1.59-.95,2.32.99.72,3.05,5.06-1.92,4.54,3.28-.65,3.13,1.12,3.02-1.2,3.13,2.38,1.84-.52,1.55,2.21,3.41,2.85,1.9-.49,2.52,4.69,1.33,1.73,3.1-3.8,3.69-4.88-1.17-1.07,2.01,2.06,2.87,1.34,7.26-4.61.26-2.18,1.99-1.42,6.82-2.59-2.29-5.48,1.03-1.65-2.65-2.5,2.1-3.69-1.89-3.26,1.41-3.76-2.97-9.13-1.87-9.59.61-5.1,3.8.32-5.69-3.22-2.7,4.99-5.34-2.9-13.19,5.61.53,7.7-4.58.33,1.53,1.61.03-1.36-2.01,2.1-6.83,6.18-4.16-2.15-.82.92-3.87,3.09-1.95,3.53.38,3.85-4.14Z"/>
+        <path id="BA" class="cls-1" d="M198.31,354.12l2.3-.11-1.55,4.92,3.14,2.97-2.28.73,1.68,3-2.02,1.02-1.65.06.52,2.17-1.19-.5-2.63,3.21-.16,4.01-5.18-3.09-.55-.37-2.1-4.68-6.56-6.32-3.8-6.04.35-3.93,3.4,1.62,4.22-2.55,6.02,1.9h5.72l2.32,1.98Z"/>
+        <path id="BG" class="cls-1" d="M262.84,364.71l-.81,3.27-2.23-.06-1.65,5.94-2.73,2.58,3.57,4.52-3.23.44-1.96-1.56-4.23,1.16-1.99,2.24-1.58.11.33,2.86-1.42.66-4.53.61-5.14-2.79-10.59,1.96.59-3.62-4.45-5.19,1.21-1.16-.38-3.67,3.44-3.17-4.09-5.5,2.26-4.26,2.17,1.49-.73,2.28,17.36,1.52,4.85-3.14,5.86-1.49,10.1,3.97Z"/>
+        <path id="BE" class="cls-1 dinosaurs" d="M109.15,291.84l-.16.73-.84,1.67,2.6.9,2.4,4.6-1.67,2.05-.95-.48-1.56,2.59.31,3.97-1.91.28-4.31-2.88-.33-3.82-1.84,2.02-2.66-.11.17-2.89-3.27-.62-3.93-5.01-2.34.3-1.58-3.7,5.56-3.01.56,1.42,3.16.41,2.18-1.93,5.42-.89,4.99,4.4Z"/>
+        <path id="AZ" class="cls-1" d="M397.48,382.09l5.96,9.35,6.12,4.6-5.49.75-1.27,7.38-3.42,4.42.1,3.47-1.86.21-4.02-3.81,1.99-1.42-1.26-1.93,1.46-1.37-2.19-2.48-10.14,6.75-.07-5.63-6.07-3.69,2.06-.42.53-1.83-3.43-2.63.89-2.79-3.48-2.04.79-.64-1.26-1.28,1.89-1.42,8.44,3.23.93-1.77-3.3-3.32,1.67-2.11,2.15.71,3.83,4.76,3.19.74,5.26-5.79ZM380.92,408.25l-4.27-1.11-4.47-5.59-.33-.46,1.78-.54.95,1.7,4.12.21,2.22,5.79Z"/>
+        <path id="AT" class="cls-1 dinosaurs" d="M184.48,317.51l-.59,2.15,1.9,3.85-.55,2.99-4.34.33,1.71,1.38-1.57,1.35.07,3.89-2.43,1.41-.91,1.83-7.17.7-2.31,2.01-5.73-1.17-8.83-1.79-1.48-3.73-6.23,1.05-1.69,1.95-3.64-.86-.7-1.18-1.45,1.31-3.73-2.03-.35-2.11-.03-2.52,3.02.19,1.42,2.25,1.73-2.71,4.06,1.58,7.86-3.24,5.42,2.39-1.71-6.29,4.14-2.57.76-2.24,1.59.41.62-2.29,1.58,1.67,4.32.04,2.53-4.07,6.67,2.48,3.28-.42,2.76,2.01Z"/>
+        <path id="AM" class="cls-1" d="M383.46,408.01l-2.54.24-2.22-5.79-4.12-.21-.95-1.7-1.78.54-3.22-2.95-4.2-.76-.65-3.13,1.03-2.1-1.91-3.45,10.52-1.64,1.26,1.28-.79.64,3.48,2.04-.89,2.79,3.43,2.63-.53,1.83-2.06.42,6.07,3.69.07,5.63Z"/>
+        <path id="AX" class="cls-1" d="M204.93,180.2l1.82,1.22-3.1,2.44-.76-2.52,1.09.64-.17-2.31,1.12.53ZM202.73,182.43l-.85.57.1-1.34.75.77Z"/>
+        <path id="AL" class="cls-1" d="M205.43,375.71l2.84,2.95.55,3.17-.52,5.39,3.2,3.77.45,2.02-2.52,4.45-2.33,1.21-.43,2.64-1.66-.27-1.01-2.93-3.56-3.2.92.02-.82-2.29,1.6-8.71-1.57-2.06-.41-2.75,2.52-4.15.9,1.39,1.85-.65Z"/>
+        <path id="IT" class="cls-1" d="M175.2,413.9l-3.21,6.48,1.32,3.4-1.23,3.09-4.11-.93-1.64-2.08-10.91-4.63-1.37-1.91,2.01-3.1,1.13,1.27,2.62-1.34,3.35,1.79,12.04-2.04ZM135.16,390.71l1.17,3.38-1.64,11.67-3.41-.63-.6,2.38-2.15.32-1.55-2.41.88-5.52-2.18-9.47,2.34.56,4.42-3.63,2.72,3.35ZM140.3,372.87l.16,1.33-1.94-.27,1.78-1.06ZM162.56,338.23l-2.16,2.52,1.71.82-1.03,1.63,2.66,3.8-1.04.26-.62-1.76h-2.84s-6.28,3.12-6.28,3.12l-.33,1.96,2.01,2.6-1.86,2.33,1,4.7,7.87,6.08,3,8.12,3.58,4.06,4.23,2.81,6.71.34-1.78,3.46,13.84,7.64,3.4,3.83-.78,3.51-3.23-4.03-6.31-1.57-2.74,6.24,4,3.2.4,3.32-3.76,1.7-.48,3.37-3.29,3.99-2.77-.79,3.8-7.78-3.48-9.13-5-2.19-1.25-3.79-2.87.62.82-1.15-2.81-.74-2.09-3.78-4.35-.07-9.77-9.4-3.6-1.18.44-1.08-4.4-3.95-2.2-9.08-9.59-4.45-5.12,5.11-3.45,1.02.97-3.71-2.15.25-2.81-1.85.62-4.64-2.41-2.29,3.44-2.99-2.39-3.45,1.55-1.79,5.17.04,4.27-5.09.24,1.95,3.81,3.88,1.59-6.11,5.85,2.31-.28-3.53,2.32.49.15-3.09,3.64.86,1.69-1.95,6.23-1.05,1.48,3.73,8.83,1.79Z"/>
+        <path id="KZ" class="cls-1 dinosaurs" d="M408.34,354.26l-1.28-.78.78-1.38.5,2.16ZM401.92,340.01l-4.65-2.62.12-1.48,2.7-.17-5.34-9.26-5.89-.32-1.54-4.94-2.72-1.81.28-4,2.22-3.54-1.55-2.23.59-3.42,4.14-7.39,5.6,5.71,2.86-.73-.91-7.2,4.71-2.54,1.17-2.48,5.05-2.21,3.68-4.75,3.71,2.75,5.9-2.54,2.37,2.47h5.16s5.4,4.74,5.4,4.74l2.8,5.37.58-5.06,7.04,4.56,5.43-4.65,3.5-.48,2.9,1.88,2.67-2.17,3.51.3,3.53,3.94,3.83.78.48,1.35,3.61-3.79,3.03,1.93,4.84-1.13,2.2-5.88-7.34-3.53-2.69-3.54,6.26-3.95-1.48-3.75,1.84-3.3,6.97-.37-2.85-2.5-3.1-.66.2-1.79,2.06-.87-3.74-1.12,1.69-4.53,4.7.83,17.06-5.03,4.23.51,2.62-3.28,18.04-4.13.34-2.16,5.2-2.7,8.11,2.68,3.75-1.68,3.01,8.27-.62,4.47,6.14.08,1.22-1.38,1.76,4.41,1.18-2.21,4.09,2.02,2.94-1.21-2.43,4.07.69,2.94,3.04-1.95,3.32,1.5,7.31-6.86,9.43-4.06-1.23,3.42-1.56-.07.42,1.48,9.26,8.55,14.33,27.29,2.94-1.83.17-2.54,1.93-1.18,2.64,1.1-.37,2.38,2.13.13.53,2.31,6.92.13,5.82-2.85,3.96,2.35,2.54,5.66,4.49,1.87,1.64,4.64,6.39,1.21,3.33-2.89-.33,2.23,4.69,4.93-3.47.37-1.74,5.32-5.39,1.45-1.5,4.73.88,6.59-1.16,1.89-4.71,2.3-.8-1.4-11.03-2.11-4.81,15.51,1.99,1.64-.6,2.85-3.89-.33-1.7-1.8-12.26,4.43,4.1,1.6-.85,5.81,2.9,8.69-2.66,1.09,1,1.56-2.27,1.27.05,4.96-1.94-2.02-3.32-.64-2.03-2.67-14.41-1.82-10.93,1.26-7.79-3.71-4.4,2.18-.42,5.43-8.2-3.18-6.87.22-2.09,4.42-12.67,7.95-3.84,6.75-3.02-1.78-.75-3.45-8.26.16-1.42-7.33-3.3-.09.62-8.99-2.01,1.05-6.04-7.74-3.12,1.52-8.32-.71-8.2,1.26-6.62-8.42-16.52-11.04-17.38,5.36v33.85s-3.65.23-3.65.23l-3.91-6.01-4.93-3.35-7.18,1.7-3.78,3.32.69-8.9-4.71-1-1.9-2.66-2.16-.14.05-2.85-3.18-6.62-3.89-2.53,1.05-1.53,7.64.88-3.59-3.7,2.73-4.15,12.03.24-2.88-2.3,2.44-5.98-.45-6.49-3.95-1.33-2.32,1.6-6.47-2.78-12.33,5.81-.78,1.77Z"/>
+        <path id="AD" class="cls-1" d="M82.48,375.28l-.19.06-.62.37-.35.13-.32.06-.25-.03-.14-.21v-.34s-.05-.3-.05-.3l-.04-.16.09-.43.2-.23.29-.2.45.08.95.27.2.26v.18s-.17.28-.17.28l-.05.21Z"/>
+        <path id="MT" class="cls-1" d="M167.3,431.46l-.4.13-.4-.25-.1-.15.56-.13.28.11.11.22-.05.07ZM169,432.91l-.22.27h-.65s-.57-.43-.57-.43v-.88l.65.17.59.59.2.28Z"/>
+        <path id="MC" class="cls-1" d="M121.05,363.81l-.41.17v-.2s.11-.11.11-.11l.13-.06.15.09.02.11Z"/>
+        <path id="SM" class="cls-1" d="M154.41,362.09l.29-.44.42-.07.08.34-.2.48-.4.07-.19-.38Z"/>
+        <path id="GI" class="cls-1" d="M34.33,431.25l-.02.42-.12-.28.14-.14h0Z"/>
+        <path id="VA" class="cls-1" d="M154.69,380.76h-.06l-.02-.03.02-.04h.05v.07Z"/>
+      </g>
+      <g id="RU">
+        <polygon id="RU" class="cls-1 dinosaurs" points="213.58 244.7 222.55 247.12 224.28 249.32 223.89 255.3 202.59 254.11 205.08 248.73 208.76 247.87 211.31 244.44 211.7 244.53 209.26 248.02 211.96 248.95 213.26 248.57 213.58 244.7"/>
+        <polygon id="RU" class="cls-1 dinosaurs" points="475.15 282.02 477.84 285.56 485.18 289.09 482.98 294.97 478.14 296.1 475.11 294.17 471.5 297.96 471.02 296.61 467.19 295.83 463.66 291.89 460.15 291.59 457.48 293.76 454.58 291.88 451.08 292.36 445.65 297.01 438.61 292.45 438.03 297.51 435.23 292.14 429.83 287.39 424.67 287.4 422.3 284.93 416.4 287.47 412.69 284.72 409.01 289.47 403.96 291.68 402.79 294.16 398.08 296.7 398.99 303.9 396.13 304.63 390.53 298.92 386.39 306.31 385.8 309.73 387.35 311.96 385.13 315.5 384.85 319.5 387.57 321.31 389.11 326.25 395 326.57 400.34 335.83 397.64 336 397.52 337.48 402.17 340.1 398.48 342.54 398.79 344.37 397.15 344.01 391.4 347.39 390.26 346.47 389.5 350.89 385.17 357.67 389.21 361.43 390.25 366.54 391.49 363.47 390.26 371.34 397.73 382.18 392.47 387.97 389.28 387.23 385.45 382.47 383.3 381.76 377.97 378.92 378.42 376.25 372.8 373.89 372.12 375.17 370.34 373.96 365.76 375.58 365.46 373.98 358.58 370.1 350.63 369.64 344.35 366.73 340.99 366.4 339.84 367.79 331.34 359.69 317.26 351.53 319.38 350.21 317.89 349.42 318.87 348.89 321.21 350.37 324.13 349.37 326.06 343.36 329.83 342.5 324.94 337.18 329.88 336.91 329.46 335.45 335.22 332.57 333.41 330.92 327.95 332.71 328.99 327.56 332.57 325.11 338.49 324.78 339.7 320.95 337.59 317.68 338.59 315.47 340.01 315.33 337.87 313.42 340.71 310.92 340.53 307.56 328.25 302.6 326.83 303.99 322.62 298.82 317.21 300.95 313.82 298.86 310.28 299.28 308.55 292.35 306.73 290.39 301 289.84 300.38 285.26 302.24 284.16 297.78 277.98 289.02 278.4 286.91 281.21 284.5 280.66 283.16 273.4 281.1 270.53 282.17 268.52 287.05 269.69 290.85 266 289.12 262.9 284.43 261.57 284.92 259.05 282.07 257.15 279.86 253.74 280.38 252.19 278 250.35 279.2 247.22 278.08 244.2 278.73 241.07 274.19 237.79 269.13 239.71 268.41 236.66 266.09 235.67 264.5 236.62 260.14 234.2 259.84 229.31 258.13 225.54 256.72 225.63 257.99 220.09 254.78 217.16 257.65 212.84 255.79 208.38 255.33 201.09 260.16 193.39 259.23 191.94 259.54 187.98 261.4 189.17 262.64 187.07 265.53 187.35 266.87 185.05 273.44 186.75 273.78 185.62 270.74 182.41 266.35 182.46 263.48 179.96 262.6 175.83 257.78 177.77 278.92 152.59 282.95 144.36 280.57 139.57 272.56 131.62 276.16 126.38 275.9 123.92 272.53 119.44 273.36 116.16 269.95 112.42 271.45 109.6 269.98 107.94 270.7 101.85 273.26 100.91 273.21 99.19 266.33 80.64 272.54 67.09 268.2 60.05 263.76 57.73 262.92 54.79 262.37 51.38 264.35 45.77 261.93 44.57 265.65 42.36 268.5 37.13 273.21 34.55 273.83 30.64 278.42 32.52 278.48 27.77 283.04 29.45 284.67 27.14 286.08 27.25 285.99 24.44 292.43 28.38 292.25 31.3 286.71 30.7 288.63 33.66 292.83 33.84 292.68 35.81 295.89 34.64 293.78 41.48 297.44 36.9 306.36 38.59 308.25 37.56 312.08 39.15 324.69 48.52 329.41 54.72 332.11 55.28 337.08 59.87 338.79 60.11 338.7 58.44 342.55 64.18 346.49 66.28 347.62 74.14 349.14 75.14 347.99 81.76 340.68 90.67 330.91 94.53 309.76 89.06 305.12 85.43 302.81 86.46 302.61 84.74 293.84 81.46 291.8 78.4 292.36 77.27 285.39 75.98 289.47 78.71 289.21 80.21 291.9 83.55 294.04 84.26 294.29 86.77 297.25 88.28 295.26 90.18 300.32 91.92 304.22 96.47 304.8 99.49 302.3 105.56 306.54 120.75 310.66 121.72 315.49 127.53 322.75 130.43 326.29 128.35 326.93 126.17 326.2 122.62 321.01 121.25 316.93 115.24 316.64 112.9 318.33 112.13 318.98 109.16 320.72 108.81 326.57 113.85 338.35 118.61 340.37 115.56 342.98 115.43 338.62 106.31 338.75 102.28 344.64 96.28 349.92 93.62 354.87 86.98 357.51 88.61 361.76 88.73 364.59 91.49 363.84 93.62 365.88 93.31 367.63 95.53 367.58 91.76 370.22 84.39 369.82 79.85 368.89 77.05 365.46 74.36 368.45 61.24 368.31 56.57 362.44 48.87 367.25 51.17 374.19 50.63 379.67 52.44 385.01 61.69 385.05 63.87 377.23 65.5 373.01 71.59 377.45 75.56 379.63 80.65 383.72 82.2 391.56 79.19 393.03 68.58 399.49 66.85 398.95 63.03 401.66 63.49 412.06 55.41 420.78 51.36 422.74 52.81 422.06 54.37 423.49 54.79 425.69 52.36 423.14 50.08 432.96 42.85 437.6 42.92 432.93 44.51 434.09 45.69 432.67 49.6 433.83 53.25 429.31 56.32 434.08 57.05 437.5 55.84 440.1 57.52 440.51 54.38 443.84 50.82 448.06 49.32 455.36 51.07 462.4 44.84 462.83 45.88 463.62 44.34 468.36 42.66 470.47 47.66 468.73 49.92 468.64 53.09 472.86 54.79 474.19 52.6 474.01 48.26 478.01 47.84 481 43.02 475.86 31.51 480.84 26.53 497.36 29.87 502.94 32.59 507.91 36.61 507.69 38.09 521.88 45.14 527.76 52.5 530.42 73.81 525.53 83.53 507.05 100.34 495.47 113.42 482.91 123.21 475.14 192.52 466.62 213.07 461.42 219.64 460.08 225.47 465.14 229.69 465.89 234.87 461.9 241.31 448.33 243.72 458.68 259.59 467.67 253.28 472.21 255.85 473.92 260.07 468.01 268.7 468.23 275.09 472.74 276.69 475.15 282.02"/>
+      </g>
+    </g>
+  </svg>`,
+  northAmerica: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="continent" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 328.14 433.79">
+    <defs>
+      <style>
+        .cls-1 {
+          fill: #ccc;
+          stroke: #fff;
+          stroke-width: .5px;
+        }
+        .dinosaurs {
+          fill: #2a7035;
+          cursor: pointer;
+        }
+        .dinosaurs:hover {
+          fill: #194520;
+        }
+      </style>
+    </defs>
+    <g id="continent" data-name="Layer 1">
+      <g>
+        <path id="BS" class="cls-1" d="M258.11,385.32l-.69.15-.71-1.76-1.05-.89.61-1.95.84.12.98,2.55.02,1.78ZM257.31,376.63l-3.06.5-.2-1.15,1.32-.25,1.85.09.09.81ZM259.61,376.6l-.48,2.21-.52-.4.05-1.62-1.26-1.23v-.36s2.21,1.4,2.21,1.4Z"/>
+        <path id="BZ" class="cls-1" d="M225.56,403.08l-.02-.43.34-.14.51.35,1-1.77.53-.04v.43s.54.01.54.01l-.04.8-.46,1.27.25.45-.29,1.05.17.27-.32,1.48-.55.77-.51.09-.55,1h-.84l.22-3.28.02-2.31Z"/>
+        <path id="CA" class="cls-1 dinosaurs" d="M199.18,86.35l-.22-5.9,3.63.58,1.63.96,3.35,4.93-.76,4.97-4.15,2.77-2.28-3.12-1.2-5.19ZM212.39,99l.33-1.49-1.97-2.45-5.65-.19.75,3.68,5.25.83,1.29-.38ZM248.74,145.95l3.09,5.1.81.57,3.07-1.27,3.02.2,2.98.28-.25-2.64-4.83-5.38-6.42-1.08-1.35.67-.12,3.55ZM183.31,83.25l-2.71,4.19,6.24.52,4.62,4.44,4.58,1.5-1.09-5.68-2.14-6.73-7.58-5.35-5.5-2.04.2,5.69,3.38,3.46ZM209.21,73.02l5.13-.12-2.22,4-.04,5.3,3.01,5.76,5.81,1.77,4.96-.99,5.18-10.73,3.85-4.45-3.38-4.97-2.21-10.65-4.6-3.19-4.72-3.68-3.58-9.56-6.52.94,1.23,4.15-2.87,1.25-1.94,5.32-1.94,7.46,1.78,7.26,3.07,5.13ZM145.46,126.39l3.92,1.95,12.67-1.3-5.82,4.77.36,3.43,4.26-.24,7.07-4.58,9.49-1.67,1.71-5.22-.49-5.57-2.94-.5-2.5,1.93-1.1-4.13-.94-5.7-2.89-1.41-2.57,4.41,4.01,11.05-4.9-.85-4.98-6.79-7.89-4-2.64,3.32-3.83,11.1ZM168.02,84.33l-3.65-2.9-1.5-.66-2.88,4.28-.04,2h4.66s3.41-2.72,3.41-2.72ZM166.56,96.68l.93-3.98-3.95-2.12-4.09,1.39-2.27,4.26,4.16,4.21,5.22-3.76ZM195.65,129.92l4.62-1.11,1.28-8.25-.09-5.94-2.14-5.56-.22,1.6-3.94-.7-4.22,4.09-3.02-.37.18,8.92,4.6-.87-.06,6.47,3.01,1.72ZM192.37,175.53l-5.06-3.93-4.71-4.21-.87-6.18-1.76-8.93-3.14-3.84-2.79-1.55-2.47,1.42,1.99,9.59-1.41,3.73-2.29-8.98-2.56-3.1-3.17,4.81-3.9-4.76-6.24,2.87,1.4-4.46-2.86-1.87-7.51,5.84-1.95,3.71-2.35,6.77,4.9,2.32,4.33-.12-6.5,3.46,1.48,3.13,3.98.17,5.99-.67,5.42,1.96-3.66,1.45-3.95-.37-4.33,1.41-1.86.87,3.46,6.35,2.49-.88,3.83,2.15,1.52,3.65,4.99-.72,7.11-1.16,5.26-2.65,3.26-.48,4.82,2.12,5.07,1.22.95-2.86-1.79-3.05,4.6-.64.28-3.59ZM200.11,174.55l-1.96,3.54-2.47,2.49,3.83,3.54,2.28-.85,3.78,2.36,1.74-2.73-1.71-3.03-.84-1.53-1.68-1.46-2.97-2.33ZM182.5,145.1l-2.13-2.17-3.76.4-.95,1.38,4.37,6.75,2.47-6.36ZM211.19,158.27l3.01-6.93,3.34-1.85,4.19-8.74-5.36-2.47-5.84-.36-2.78,2.77-1.46,4.23-.04,4.82,1.75,8.19,3.19.34ZM228.34,135.27l5.76-.18,8.04-1.61,3.59,1.27,4.18-2.26,1.75-2.84-.63-4.52-3-4.23-4.56-.8-5.71.97-4.46,2.44-4.09-.94-3.78-.49-1.78-2.7-3.22-2.61.64-4.43-2.42-3.98-5.52.03-3.11-3.99-5.78-.8-1.05,5.1,3.25,3.75,5.8,1.45,2.82,5.1.34,5.6.97,5.99,7.45,3.42,4.52,1.26ZM139.32,117l5.21-5.05,2.62-.59,2.16-4.23.39-9.77-3.85,1.91-4.3-.18-5.76,8.19-4.76,8.98,3.8,2.51,4.49-1.77ZM211.5,133.17l1.53-4.14-1.02-3.46-2.45-3.92-4.03,3.02-1.49,4.92,3.4,2.79,4.06.79ZM203.19,144.61l-.73-2.88-5,1.26-3.34-2.11-3.32,4.8,3.09,6.28-5.72-1.17-.06,3.01,6.97,7.05,1.94,3.38,2.7.73,4.6-3.41.5-8.21-4.24-4.07,2.61-4.66ZM129.2,298.35l-1.16-2.34-2.8-1.77-1.39-2.05-.95-1.5-2.63-.46-1.72-.67-2.94-.96-.24,1.02,1.08,2.38,2.89.78.51,1.23,2.51,1.5.84,1.51,4.6,1.92,1.4-.59ZM250.91,220.72l-2-2.11-2.06.5-.25-3.06-3.21-2.03-3.07-2.27-1.63-1.75-1.43,1.03-.52-2.96-2.03-.56-.96,6.13-.36,5.11-2.44,3.14,3.8-.6.96,3.65,3.99-3.22,2.78-3.38,1.58,2.86,4.36,1.51,2.49-1.99ZM130.37,168.18l7.38-4.18v-3.87l3.48-6.41,6.88-6.69,3.53-2.47-3.01-4.2-2.72-2.95-7.16-.57-4-2.16-9.48,1.63,2.74,6.23-2.43,6.43-1.94,6.87-1.2,3.86,6.47,4.69,1.46,3.79ZM264.61,195.48l.32-1.01-.03-3.17-2.19-2.08-2.57,1.05-1.19,4.17.7,3.56,3.14-.36,1.82-2.16ZM288.43,203.02l4.41,6.6,3.45,2.86,4.92-7.87.87-4.93-4.41-.47-4.03-6.7-4.45-1.64-6.6-4.97,5.15-3.63-2.65-7.54-2.44-3.35-6.77-3.35-2.92-5.55-5.21,1.99-.36-3.86-3.86-4.32-6.22-4.71-2.65,3.71-5.55,2.66.42-6.06-4.81-10.05-7.11,4.06-2.59,7.7-2.21-5.92,2.06-6.37-7.24,2.65-2.88,3.99-2.15,8.42.89,9.05,3.98.04-2.93,3.92,2.33,2.96,4.55,1.25,5.93,2.42,10.2,1.82,5.08-1.04,1.5-2.42,2.21,2.79,2.47.46,2.97,4.97-1.8,1.98,5.68,2.63,4.3,3.68,1.08,2.55.77,3.24-3.63,6.93-.98,3.44.94,2.42-5.77.86-5.27.12-1.85,4.87,2.37,2.23,8.11-1.03-.04-1.89,4.08,3.15,4.18,3.28-.98,1.77,3.4,3.02,6.02,3.54,7.6,2.39-.46-2.09-2.92-3.67-3.96-5.37,7.03,5,3.54,1.66.97-4.44-1.82-6.3-1.15-1.73-3.81-3.03-2.95-3.91.35-3.94,3.62-.93ZM222.6,41.46l2.34,7.29,4.96,5.88,9.81-1.09,6.31,1.97-4.37,6.05-2.21-1.78-7.66-.71,1.19,8.31,3.96,6.04-.79,5.2-4.97,3.46-2.27,5.47,4.55,2.65,3.82,8.55-7.5-5.7-1.71.94,1.38,9.38-5.18,2.83.35,5.85,5.3.63,4.17,1.44,8.24-1.84,7.33,3.27,7.49-7.19-.06-3.02-4.79.48-.39-2.84,3.92-3.83,1.33-5.15,4.33-3.83,2.66-4.76-2.32-7.1,1.94-2.65-3.86-1.89,8.49-1.63,1.79-3.15,5.78-2.6,4.8-13.47,4.57-4.94,6.62-11.12-6.1.1,2.54-4.3,6.78-3.99,6.84-8.9.12-5.73-5.13-6.04-6.02-2.93-7.49-1.82-6.07-1.49-6.07-1.5-8.1,3.98-1.49-2.53-8.57.98-5.03,2.57-3.7,3.65-2.18,11.71-3.03-5.98-3.48-1.14-4.12,7.97-5.5,3.35-3.27.66-4.17,3.84.61,6.65,3.28,5.49ZM297,306.46l-.98-1.98-1.06,1.26.7,1.36,3.56,1.71,1.04-.26,1.38-1.66-2.6.11-2.04-.54ZM240,228.6l.61,1.63,1.98.14,3.28-3.34.06-1.19-3.85-.06-2.08,2.82ZM302.13,295.04l-2.87-1.8-3.69-1.09-.97.37,2.61,2.04,3.63,1.34,1.37-.08-.08-.78ZM327.02,299.83l-.36-2.23-1.96.72.87-3.11-2.8-1.32-1.29,1.05-2.49-1.18.98-1.51-1.88-.93-1.83,1.47,1.86-3.82,1.5-2.8.54-1.22-1.3-.2-2.43,1.55-1.74,2.53-2.9,6.92-2.35,2.56,1.22,1.14-1.75,1.47.43,1.23,5.44.13,3.01-.25,2.69,1-1.98,1.93,1.67.14,3.25-3.58.78.53-.61,3.37,1.84.77,1.27-.15,1.18-3.61-.86-2.6ZM305.82,304.6l-2.81,4.56-4.63.58-3.64-2.01-.91-3.07-.89-4.46,2.65-2.83-2.48-2.09-4.19.43-5.88,3.53-4.5,5.45-2.38.67,3.23-3.8,4.04-5.57,3.58-1.9,2.35-3.11,2.9-.3,4.21.03,6,.92,4.74-.71,3.53-3.62,4.62-1.59,2.01-1.58,2.04-1.71-.2-5.19-1.13-1.77-2.18-.63-1.11-4.05-1.8-1.55-4.47-1.26-2.52-2.82-3.73-2.83,1.13-3.2-3.1-6.26-3.65-6.89-2.18-4.98-1.85,2.61-2.68,6.05-4.06,2.97-2.03-3.15-2.56-.85-.93-6.99.08-4.8-5-.44-.85-2.27-3.45-3.44-2.61-2.04-2.32,1.58-2.88-.58-4.81-1.65-1.95,1.4.94,9.18,1.22,5.12-3.31,5.75,3.41,4.02,1.9,4.44.23,3.42-1.55,3.5-3.18,3.46-4.49,2.28,1.98,2.53,1.46,7.4-1.52,4.68-2.16,1.46-4.17-4.28-2.03-5.17-.87-4.76.46-4.19-3.05-.47-4.63-.28-2.97-2.08-3.51-1.37-2.01-2.38-2.8-1.93-5.21-2.23-3.92,1.02-1.31-3.95-1.26-4.99-4.12-.9.16-6.41,1.09-4.48,3.04-6.6,3.43-4.9,3.26-.77.19-4.05,2.21-2.68,4.01-.42,3.25-4.39.82-2.9,2.7-5.72.84-3.5,2.9,2.11,3.9-1.08,5.49-4.96.36-3.54-1.98-3.98,2.09-4.06-.17-3.86-3.76-3.95-4.14-1.19-3.98-.62-.15,8.71-2.04,6.56-2.93,5.3-2.71-4.95.84-5.61-3.35-5.02-3.75,6.09v-7.99s-5.2-1.63-5.2-1.63l2.49-4.01-3.81-9.59-2.95-3.92-3.7-1.44-3.31,6.43-.22,9.34,3.27,3.29,3,4.91-1.27,7.71-2.25-.2-1.78,5.88.02-7-4.34-2.58-2.49,1.34.32,4.67-4.09-.18-4.35,1.17-4.95-3.35-3.13.6-2.82-4.11-2.26-1.84-2.24.77-3.41.36-1.81,2.61,2.86,3.19-3.05,3.73-2.99-4.42-2.39,1.3-7.57.87-5.07-1.59,3.95-3.74-3.78-3.9-2.75.5-3.86-1.32-6.56-2.89-4.29-3.37-3.4-.47-1.06,2.36-3.44,1.31-.38-6.15-3.73,5.51-4.74-7.32-1.94-.89-.63,3.91-2.09,1.9-1.93-3.39-4.59,2.05-4.2,3.55-4.16-.98-3.4,2.5-2.46,3.28-2.92-.72-4.41-3.8-5.23-1.94-.02,27.65v35.43s2.75.17,2.75.17l2.73,1.56,1.96,2.44,2.49,3.6,2.73-3.05,2.82-1.79,1.49,2.86,1.89,2.23,2.57,2.42,1.75,3.79,2.87,5.88,4.77,3.2.08,3.12-1.56,2.36.06,2.48,3.39,3.45.49,3.76,3.59,1.96-.4,2.79,1.56,3.96,5.08,1.83,2,1.89,5.43,4.23h.38s7.96.01,7.96.01h69.65v-1.64s.96-.02.96-.02l.5,2.35.87.72,1.96.26,2.86.67,2.72,1.3,2.27-.55,3.45,1.09,1.14-1.66,1.59-.66.62-1.03.63-.55,2.61.86,1.93.1.67.57.94,2.38,3.15.63-.49,1.18,1.11,1.21-.48,1.56,1.18.51-.59,1.37.75.13.53-.6.55.9,2.1.5,2.13.04,2.27.41,2.51.78.92,1.26,1.82,3.04-.9,1.3-2.28-.54-1.41-2.44.36,2.49-1.34,2.17.15,1.84-.23,1.07-1.81,1.27-1.32,2.09-.62,1.32,1.54.24,2.08-1.2,1.23-1.06.83-.17,1.54.38.75-.59,1.37-.48,2.44-.47h0l-.25-1.15-.13.04-.86.2-1.12-.36.84-1.32.86-.46,1.98-.56,2.37-.53,1.24.73.78-.85.89-.54.6.29.03.06,2.87-2.73,1.26-.73,4.26-.03h5.17l.28-.98.9-.2,1.19-.62,1-1.82.86-3.15,2.14-3.1.93,1.08,1.88-.7,1.25,1.19v5.53l1.83,2.25,3.12-.48,4.49-.13-4.87,3.26.11,3.29,2.13.28,3.13-2.79,2.78-1.58,6.21-2.35,3.47-2.62-1.81-1.46-.29-3.03ZM252.16,233.49l1.1-3.12-.71-1.23-1.15-.13-1.08,1.8-.13.41.74,1.77,1.23.5ZM109.5,269.92h0l1.56-2.35-1.56,2.35Z"/>
+        <path id="CR" class="cls-1" d="M242.88,430.52l-1.52-.63-.57-.59.32-.49-.1-.62-.78-.68-1.1-.55-.97-.36-.18-.83-.74-.5.18.82-.56.67-.64-.78-.9-.28-.38-.57.02-.86.37-.9-.79-.4.64-.54.42-.37,1.85.75.64-.37.89.24.47.58.82.19.68-.6.71,1.54,1.08,1.14,1.32,1.21-1.09.25.02,1.14.58.41-.42.34.11.51-.23.57-.15.56Z"/>
+        <path id="CU" class="cls-1" d="M244.84,387.06l2.42.22,2.2.03,2.63,1.03,1.12,1.11,2.62-.34.99.71,2.38,1.86,1.74,1.35.92-.04,1.68.61-.21.84,2.07.12,2.12,1.22-.33.69-1.87.38-1.88.15-1.94-.24-4.01.29,1.88-1.66-1.14-.77-1.81-.2-.97-.86-.67-1.7-1.58.11-2.61-.8-.85-.63-3.65-.47-.98-.59,1.05-.75-2.75-.15-2.01,1.56-1.16.04-.41.74-1.38.33-1.2-.29,1.48-.93.6-1.08,1.27-.68,1.43-.59,2.13-.29.68-.33Z"/>
+        <path id="DO" class="cls-1" d="M274.43,397.47l.35-.51,2.19.02,1.66.76.74-.08.51,1.05,1.54-.06-.1.88,1.25.11,1.38,1.08-1.04,1.2-1.33-.64-1.29.12-.92-.14-.51.54-1.08.18-.42-.71-.93.42-1.12,2-.73-.46-.14-.84.06-.8-.72-.88.68-.5.22-1.13-.25-1.61Z"/>
+        <path id="GT" class="cls-1" d="M222.89,414.87l-1.44-.5-1.75-.05-1.28-.57-1.51-1.18.07-.84.32-.68-.39-.54,1.35-2.36h3.59s.08-.99.08-.99l-.46-.18-.31-.63-1.04-.67-1.04-.98h1.27v-1.65h2.62l2.59.03-.02,2.31-.22,3.28h.84l.91.53.24-.43.82.36-1.27,1.11-1.33.81-.2.55.22.57-.58.73-.66.17.15.34-.52.32-.96.72-.09.42Z"/>
+        <path id="HN" class="cls-1" d="M230.68,417.02l-.48-.89-.86-.25.2-1.15-.38-.31-.58-.2-1.23.34-.1-.39-.85-.46-.6-.57-.83-.24.58-.73-.22-.57.2-.55,1.33-.81,1.27-1.11.29.12.62-.51.8-.04.26.23.44-.14,1.3.26,1.3-.08.9-.32.33-.32.89.15.67.2.73-.07.56-.25,1.28.4.45.06.85.54.81.65,1.02.45.74.79-.96-.06-.39.39-.97.38h-.71l-.62.37-.56-.13-.48-.44-.29.08-.36.69-.27-.03-.05.6-.98.79-.51.34-.29.36-.83-.58-.6.76-.59-.02-.66.07.06,1.41-.41.02-.35.66-.87.11Z"/>
+        <path id="HT" class="cls-1" d="M270.29,396.87l1.71.13,2.43.47.25,1.61-.22,1.13-.68.5.72.88-.06.8-1.86-.5-1.32.2-1.71-.21-1.31.55-1.51-.92.25-.94,2.58.4,2.12.24,1.01-.65-1.28-1.28.02-1.13-1.77-.46.63-.82Z"/>
+        <path id="JM" class="cls-1" d="M258.01,401.08l1.89.26,1.49.71.46.8-1.97.05-.85.49-1.57-.47-1.6-1.07.33-.67,1.18-.2.64.1Z"/>
+        <path id="MX" class="cls-1" d="M203.14,378.84l-1.09,2.71-.49,2.21-.21,4.08-.27,1.47.49,1.64.87,1.47.56,2.31,1.86,2.22.65,1.68,1.1,1.46,2.98.78,1.16,1.22,2.46-.82,2.13-.29,2.1-.53,1.77-.51,1.78-1.2.67-1.72.23-2.5.49-.87,1.89-.79,2.97-.69,2.48.1,1.7-.25.68.63-.1,1.44-1.51,1.77-.66,1.81.51.51-.42,1.28-.7,2.3-.71-.76-.59.05-.53.04-1,1.77-.51-.35-.34.14.02.43-2.59-.03h-2.62v1.65h-1.27l1.04.98,1.04.67.31.63.46.18-.08.98h-3.59s-1.35,2.37-1.35,2.37l.39.54-.32.68-.07.84-3.17-3.11-1.45-.94-2.29-.76-1.56.21-2.26,1.09-1.41.29-1.98-.76-2.1-.56-2.62-1.33-2.1-.4-3.18-1.36-2.34-1.4-.71-.78-1.57-.17-2.87-.93-1.17-1.34-3.01-1.67-1.4-1.87-.67-1.45.93-.29-.28-.85.64-.77v-1.04s-.93-1.34-.93-1.34l-.26-1.2-.94-1.52-2.47-3.02-2.82-2.39-1.36-1.91-2.41-1.26-.51-.76.43-1.92-1.43-.73-1.66-1.52-.7-2.19-1.51-.26-1.62-1.66-1.32-1.55-.12-1-1.51-2.42-.99-2.48.04-1.25-2.03-1.29-.93.14-1.6-.9-.45,1.33.46,1.56.27,2.43.97,1.33,2.08,2.21.46.75.43.22.37,1.1.49-.05.57,2.04.85.8.6,1.11,1.76,1.6.93,2.89.83,1.35.78,1.44.15,1.62,1.36.1,1.12,1.39,1.02,1.36-.07.54-1.18,1.11h-.5s-.74-1.86-.74-1.86l-1.83-1.73-2.02-1.48-1.44-.78.09-2.25-.42-1.68-1.34-.96-1.92-1.39-.37.4-.71-.82-1.73-.75-1.65-1.83.2-.24,1.15.18,1.05-1.18.1-1.43-2.16-2.27-1.65-.89-1.03-2-1.04-2.13-1.3-2.61-1.14-2.96,3.19-.25,3.56-.36-.26.64,4.23,1.61,6.4,2.31,5.58-.02h2.22v-1.36h4.86l1.02,1.17,1.44,1.03,1.66,1.43.93,1.69.7,1.76,1.45.97,2.33.96,1.76-2.53,2.3-.06,1.97,1.28,1.41,2.18.97,1.86,1.65,1.8.62,2.19.79,1.47,2.18.96,1.99.68,1.1-.09Z"/>
+        <path id="NI" class="cls-1" d="M235.18,422.43l-.97-.9-1.31-1.15-.62-.96-1.18-.89-1.41-1.29.32-.44.46.43.21-.21.87-.11.35-.66.41-.02-.06-1.41.66-.07.59.02.6-.76.83.58.29-.36.51-.34.98-.79.05-.6.27.03.36-.69.29-.08.48.44.56.13.62-.37h.71l.97-.38.39-.39.96.06-.24.28-.14.64.28,1.05-.64.98-.3,1.16-.09,1.26.15.73.07,1.29-.43.28-.26,1.22.19.75-.57.73.13.76.43.47-.68.6-.82-.19-.47-.58-.89-.24-.64.37-1.85-.75-.42.37Z"/>
+        <path id="PA" class="cls-1" d="M257.14,433.33l-.94-.81-.6-1.52.69-.75-.71-.19-.52-.93-1.4-.78-1.22.18-.57.98-1.14.7-.61.1-.27.59,1.34,1.52-.77.36-.41.42-1.3.14-.49-1.68-.36.48-.93-.16-.56-1.14-1.15-.18-.73-.33h-1.2l-.09.61-.32-.42.15-.56.23-.57-.11-.51.42-.34-.58-.41-.02-1.14,1.09-.25,1,1.01-.05.6,1.11.12.27-.23.77.7,1.38-.21,1.19-.71,1.7-.57.96-.84,1.55.16-.1.28,1.56.1,1.25.49.91.84,1.06.78-.35.42.66,1.65-.53.84-.91-.2-.35,1.36Z"/>
+        <path id="PR" class="cls-1" d="M289.66,401.01l1.43.26.51.58-.72.74-2.11-.02-1.64.1-.17-1.25.4-.43,2.3.02Z"/>
+        <path id="SV" class="cls-1" d="M229.34,415.88l-.31.67-1.62-.04-1.01-.27-1.16-.57-1.56-.18-.79-.62.09-.42.96-.72.52-.32-.15-.34.66-.17.83.24.6.57.85.46.1.39,1.23-.34.58.2.38.31-.2,1.15Z"/>
+        <path id="US" class="cls-1 dinosaurs" d="M109.5,269.92h0l-1.54-1.83-2.47-1.57-.79-4.36-3.61-4.13-1.51-4.94-2.69-.34-4.46-.13-3.29-1.53-5.8-5.64-2.68-1.05-4.9-1.99-3.88.48-5.51-2.59-3.33-2.43-3.11,1.21.58,3.93-1.55.36-3.24,1.16-2.47,1.86-3.11,1.16-.4-3.24,1.26-5.53,2.98-1.77-.77-1.46-3.57,3.22-1.91,3.77-4.04,3.95,2.05,2.65-2.65,3.85-3.01,2.21-2.81,1.59-.69,2.29-4.38,2.63-.89,2.36-3.28,2.13-1.92-.38-2.62,1.38-2.84,1.67-2.33,1.63-4.81,1.38-.44-.81,3.07-2.27,2.74-1.51,2.99-2.71,3.48-.56,1.38-2.06,3.89-3.05.63-1.03,2.07-1.83.48-4,1.43-3.17-3.23,1.64-.9-.93-1.52,1.95-1.83-2.73-.76,1.94-1.05-2.69-2.8,2.17h-1.72l-.24-3.23.51-2.02-1.81-1.98-3.65,1.07-2.37-2.63-1.92-1.36v-3.25s-2.17-2.48-2.17-2.48l1.09-3.4,2.29-3.37,1-3.15,2.27-.45,1.92.99,2.26-3.01,2.04.54,2.14-1.95-.52-2.92-1.57-1.16,2.08-2.52-1.72.08-2.97,1.43-.85,1.43-2.21-1.43-3.97.73-4.11-1.56-1.18-2.65-3.55-3.91,3.94-2.87,6.25-3.41h2.31l-.38,3.48,5.92-.27-2.28-4.34-3.45-2.72-1.99-3.64-2.69-3.17-3.85-2.38,1.57-4.03,4.97-.25,3.54-3.58.67-3.92,2.86-3.91,2.73-.95,5.32-3.76,2.58.57,4.32-4.61,4.24,1.83,2.03,3.87,1.25-1.65,4.74.51-.17,1.95,4.29,1.43,2.86-.84,5.91,2.64,5.39.78,2.16,1.07,3.73-1.33,4.25,2.46,3.05,1.14-.02,27.65v35.43s2.75.17,2.75.17l2.73,1.56,1.96,2.44,2.49,3.6,2.73-3.05,2.82-1.79,1.49,2.86,1.89,2.23,2.57,2.42,1.75,3.79,2.87,5.88,4.77,3.2.08,3.12-1.66,2.24ZM285.43,304.36l-1.25-1.19-1.88.7-.93-1.08-2.14,3.1-.85,3.15-.99,1.82-1.19.62-.9.2-.28.98h-5.17l-4.26.03-1.26.73-2.87,2.73.29.54.17,1.51-2.1,1.27-2.3-.32-2.2-.14-1.33.44.25,1.15h0l.06.37-2.42,2.27-2.12,1.09-1.44.51-1.66,1.03-2.03.5-1.4-.19-1.73-.77.96-1.45.62-1.32,1.32-2.09-.14-1.57-.5-2.24-1.04-.39-1.74,1.71-.56-.03-.14-.97,1.54-1.56.26-1.79-.23-1.79-2.08-1.55-2.38-.8-.39,1.52-.62.41-.5,1.95-.26-1.32-1.12.95-.7,1.32-.73,1.92-.14,1.65.93,2.38-.08,2.51-1.14,1.84-.57.52-.76.41-.95.02-.26-.25-.75-1.97-.02-.98.08-.94-.35-1.87.53-2.18.63-2.71,1.46-3.03h-.42s-2.06,2.55-2.06,2.55l-.38-.46,1.1-1.42,1.67-2.57,1.91-.36,2.19-.8,2.21.42.09.02,2.47-.36-1.39-1.61-.75-.12-.85-.16-.59-1.14-2.75.36-2.49.91-1.97-1.55-1.59-.52.9-2.17-2.47,1.37-2.25,1.33-2.16,1.04-1.72-1.4-2.81.85v-.6s1.91-1.73,1.91-1.73l1.99-1.65,2.86-1.37-3.45-1.09-2.27.55-2.72-1.3-2.86-.67-1.96-.26-.87-.72-.5-2.35-.95.02v1.64h-77.62l.95,3.47.45,3.41-.69,1.09-1.49-3.91-4.05-1.42-.34.82.82,1.94.89,3.53.51,5.42-.34,3.59-.34,3.54-1.1,3.61.9,2.9.1,3.2-.61,3.06,1.49,1.99.39,2.95,2.17,2.99,1.24,1.17-.1.82,2.34,4.85,2.72,3.45.34,1.87.72.55,2.61.33,1,.91,1.57.17.31.97,1.31.4,1.82,1.92.47,1.7,3.19-.25,3.56-.36-.26.65,4.23,1.6,6.4,2.31,5.58-.02h2.22v-1.35h4.87l1.02,1.16,1.43,1.03,1.67,1.43.93,1.69.7,1.77,1.45.97,2.33.96,1.77-2.53,2.29-.06,1.98,1.28,1.41,2.18.97,1.86,1.65,1.8.62,2.19.79,1.47,2.19.96,1.99.68,1.09-.09-.53-1.05-.14-1.49.03-2.16.65-1.42,1.53-1.51,2.79-1.37,2.55-2.37,2.36-.75,1.74-.22,2.04.74,2.44-.4,2.09,1.69,2.04.1,1.05-.61,1.04.47.53-.42-.59-.63.05-1.3-.5-.86,1.16-.5,2.14-.22,2.49.36,3.17-.41,1.76.8,1.36,1.5.5.16,2.83-1.46,1.09.49,2.19,2.68.79,1.75-.58,2.1.42,1.23,1.3,2.4,1.49,2.68,1.07.71.44,1.35,1.38.37.84-.39.7-1.89.12-1.21.09-2.1-1.33-3.65-.02-1.37-1.25-2.25-.94-2.74-.5-2.25.43-2.31,1.32-1.94,1.58-1.57,3.08-2.16.4-1.12,1.42-1.23,1.4-.21,1.84-1.98,2.9-1.01,1.78-2.53-.39-3.45-.29-1.2-.8-.24-.12-3.35-1.93-1.14,1.85.56-.6-2.26.54-1.55.33,2.97,1.43,1.36-.87,2.4.26.14,1.58-2.81.9-1.38-.04-1.35-.7-.64-.58-1.94.92.9.62.19.21.92,2.04-2.78.61-2.62-.82-.17.85-1.02-.08.45h1.79s3.93-1.12,3.93-1.12l-.83-.7-4.12.7,2.34-1.07,1.63-.18,1.22-.19,2.07-.65,1.35.07,1.89-.6.22-1.07-.84-.83.29,1.37-1.16-.09-.92-1.99.03-2.01.48-.86,1.48-2.28,2.96-1.15,2.88-1.34,2.99-1.9-.48-1.29-1.83-2.25-.15-5.74ZM45.87,253.91l-1.5.8-2.55,1.86.43,2.42,1.43,1.32,2.8-1.95,2.43-2.46-1.19-1.63-1.85-.36ZM.25,225.35l2.04-1.26.23-.68-2.27-.68v2.62ZM8.75,240.71l-2.77.97,1.7,1.52,1.84,1.04,1.72-.87-.27-2.15-2.22-.51ZM106.1,273.21l-2.69.38-1.32-.62-.17,1.52.52,2.07,1.42,1.46,1.04,2.13,1.69,2.1h1.12s-2.44-3.69-2.44-3.69l.83-5.35ZM37.38,393.89l-1-.28-.27.26.03.19.32.24.48.63.94-.21.23-.36-.73-.47ZM34.39,393.35l1.5.09.09-.32-1.38-.13-.21.36ZM40.28,396.64l-.5-.26-1.07-.5-.21-.05-.16.28.19.58-.49.48-.14.33.46,1.08-.08.83.7.42.41-.49.9-.46,1.1-.63.07-.16-.71-1.04-.47-.41ZM32.42,391.5l-.75.41.11.12.36.68.98.11.2.04.15-.17-.81-.99-.24-.2ZM28.02,389.94l-.43.3-.14.22.94.55.33-.3-.06-.7-.64-.07Z"/>
+      </g>
+    </g>
+  </svg>`,
+  southAmerica: `<?xml version="1.0" encoding="UTF-8"?>
+  <svg id="continent" data-name="Layer 2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 131.43 223">
+    <defs>
+      <style>
+        .cls-1 {
+          fill: #ccc;
+          stroke: #fff;
+          stroke-width: .5px;
+        }
+        .dinosaurs {
+          fill: #2a7035;
+          cursor: pointer;
+        }
+        .dinosaurs:hover {
+          fill: #194520;
+        }
+      </style>
+    </defs>
+    <g id="continent" data-name="Layer 1">
+      <g>
+        <path id="AR" class="cls-1 dinosaurs" d="M44.89,220.71l-2.66.25-1.43-1.73-1.69-.13h-3v-10.57l1.08,2.15,1.4,3.53,3.65,2.87,3.93,1.21-1.28,2.42ZM46.39,98.27l1.65,2.18,1.09-2.43,3.2.12.45.64,5.15,4.94,2.29.46,3.43,2.26,2.89,1.2.4,1.36-2.76,4.73,2.83.85,3.15.48,2.22-.5,2.54-2.4.46-2.74,1.39-.59,1.41,1.79-.06,2.49-2.36,1.73-1.88,1.28-3.16,3.08-3.74,4.37-.7,2.59-.75,3.37.03,3.3-.61.74-.22,2.17-.19,1.76,3.56,2.91-.38,2.37,1.75,1.51-.14,1.7-2.69,4.52-4.16,1.91-5.62.75-3.08-.36.59,2.15-.57,2.72.52,1.85-1.68,1.3-2.87.51-2.7-1.35-1.08.97.39,3.71,1.89,1.14,1.54-1.19.84,1.96-2.58,1.18-2.25,2.38-.41,3.91-.66,2.11h-2.65s-2.2,2.05-2.2,2.05l-.8,3.01,2.76,2.98,2.68.83-.96,3.73-3.31,2.38-1.82,5.03-2.56,1.72-1.15,2.06.91,4.64,1.87,2.63-1.18-.23-2.6-.71-6.78-.61-1.16-2.63.05-3.33-1.87.28-.99-1.6-.25-4.6,2.15-1.88.89-2.68-.33-2.11,1.49-3.52,1.02-5.35-.3-2.33,1.22-.75-.3-1.48-1.3-.78.92-1.63-1.27-1.46-.65-4.4,1.13-.77-.47-4.54.66-3.75.75-3.22,1.68-1.3-.85-3.46v-3.22s2.11-2.26,2.11-2.26l-.06-2.87,1.6-3.31v-3.09s-.72-.61-.72-.61l-1.29-5.69,1.73-3.34-.27-3.11,1-2.9,1.84-2.96,1.98-1.95-.84-1.23.59-1-.09-5.14,3.05-1.51.96-3.16-.34-.76,2.34-2.72,3.62.72Z"/>
+        <path id="BO" class="cls-1" d="M52.33,98.15l-3.2-.13-1.09,2.43-1.65-2.18-3.67-.73-2.33,2.72-2.03.41-1.1-4.15-1.5-3.34.88-2.87-1.47-1.25-.37-2.12-1.38-2,1.77-3.14-1.21-2.44.65-.97-.51-1.07,1.1-1.44.06-2.44.13-2.02.61-.96-2.43-4.58,2.09.24,1.44-.07.63-.85,2.45-1.15,1.47-1.06,3.67-.48-.29,2.12.34,1.09-.23,1.9,3.05,2.55,3.14.47,1.1,1.07,1.9.57,1.16.83,1.76-.03,1.63.85.12,1.66.55.84.04,1.25-.82.04,1.08,3.37,5.37.12-.41,1.68.3,1.15,1.53.82.67,1.82-.5,2.32-.77,1.29.27,1.69-.88.61-.04-.91-2.62-1.51-2.6-.05-4.89.86-1.34,2.62-.07,1.6-1.11,3.59-.45-.64Z"/>
+        <path id="BR" class="cls-1 dinosaurs" d="M66.97,123.59l3.74-4.37,3.17-3.08,1.88-1.28,2.36-1.73.06-2.49-1.41-1.79-1.39.59.55-1.78.38-1.82v-1.68l-1.01-.55-1.05.49-1.04-.13-.33-1.18-.26-2.77-.53-.9-1.89-.82-1.14.59-2.96-.58.18-4.06-.83-1.66.88-.61-.27-1.69.77-1.29.5-2.32-.67-1.82-1.53-.82-.3-1.15.41-1.68-5.37-.12-1.08-3.37.82-.04-.04-1.25-.55-.84-.12-1.66-1.63-.85-1.76.03-1.16-.83-1.9-.57-1.1-1.07-3.14-.47-3.05-2.55.23-1.9-.34-1.09.29-2.12-3.67.48-1.47,1.06-2.45,1.15-.63.85-1.44.07-2.09-.24-1.58.49-1.28-.33.19-4.3-2.3,1.66-2.47-.07-1.06-1.51-1.86-.16.59-1.21-1.56-1.72-1.17-2.53.74-.51v-1.19l1.7-.81-.28-1.51.71-.98.21-1.3,3.2-1.91,2.3-.53.37-.42,2.53.13,1.26-7.65.07-1.21-.44-1.59-1.24-1.02v-2.02s1.59-.46,1.59-.46l.56.29.09-1.07-1.64-.29-.03-1.74,5.46.06.93-.96.78.88.54,1.65.53-.35,1.55,1.48,2.18-.18.54-.86,2.08-.65,1.16-.45.32-1.18,2.01-.8-.16-.58-2.37-.24-.39-1.76.11-1.87-1.25-.72.52-.26,2.08.36,2.23.7.81-.66,2.01-.44,3.14-1.04,1.03-1.07-.38-.79,1.46-.12.66.64-.37,1.23.96.42.65,1.3-.78.98-.45,2.38.72,1.41.2,1.29,1.73,1.3,1.38.14.31-.54.88-.12,1.27-.49.91-.74,1.55.23.68-.1,1.53.23.25-.57-.47-.55.28-.81,1.13.25,1.33-.29,1.6.59,1.23.58.87-.76.62.12.39.79,1.34-.2,1.07-1.06.86-2.06,1.66-2.55.96-.13.69,1.54,1.57,4.88,1.5.46.08,1.92-2.11,2.29.87.84,4.96.44.1,2.79,2.13-1.83,3.53,1.01,4.65,1.7,1.37,1.63-.46,1.54,3.26-.86,5.46,1.48,4.19-.11,4.14,2.31,3.58,3.13,2.16.8,2.4.12,1.02.88.95,3.57.47,1.69-1.12,4.66-1.43,1.84-3.95,3.94-1.79,3.21-2.07,2.48-.7.06-.79,2.1.2,5.4-.78,4.48-.3,1.93-.88,1.15-.5,3.94-2.84,3.88-.48,3.09-2.27,1.31-.66,1.81h-3.04s-4.41,1.16-4.41,1.16l-1.98,1.35-3.14.89-3.3,2.44-2.37,3.06-.41,2.32.47,1.73-.53,3.18-.63,1.55-1.96,1.75-3.11,5.68-2.47,2.59-1.91,1.54-1.27,3.16-1.86,1.91-.78-1.9,1.24-1.57-1.62-2.25-2.2-1.82-2.89-2.08-1.04.09-2.81-2.5-1.82.34Z"/>
+        <path id="CL" class="cls-1 dinosaurs" d="M36.1,208.53v10.57h3l1.69.13-.93,1.98-2.4,1.53-1.38-.16-1.66-.4-2.04-1.48-2.94-.71-3.53-2.71-2.86-2.57-3.86-5.25,2.31.97,3.94,3.13,3.72,1.7,1.45-2.17.91-3.2,2.58-1.91,2,.55ZM37.26,96.52l1.1,4.15,2.02-.41.34.76-.96,3.16-3.05,1.51.09,5.14-.59,1,.84,1.23-1.98,1.95-1.84,2.96-1,2.9.27,3.11-1.73,3.34,1.29,5.69.73.61v3.09s-1.61,3.31-1.61,3.31l.06,2.87-2.12,2.26v3.22s.86,3.46.86,3.46l-1.68,1.3-.75,3.22-.66,3.75.47,4.54-1.13.77.65,4.4,1.27,1.46-.92,1.63,1.3.78.3,1.48-1.22.75.3,2.33-1.02,5.35-1.49,3.52.33,2.11-.89,2.68-2.15,1.88.25,4.6.99,1.6,1.87-.28-.05,3.33,1.16,2.63,6.78.61,2.6.71-2.49-.03-1.35,1.13-2.53,1.67-.45,4.38-1.19.11-3.16-1.54-3.21-3.25h0l-3.49-2.63-.88-2.87.79-2.62-1.41-2.94-.36-7.34,1.19-4.03,2.96-3.19-4.26-1.19,2.67-3.57.95-6.56,3.12,1.37,1.46-7.97-1.88-1-.88,4.75-1.77-.54.88-5.42.96-6.84,1.29-2.48-.81-3.5-.23-3.98,1.18-.11,1.72-5.6,1.94-5.43,1.19-4.97-.65-4.91.84-2.67-.34-3.96,1.64-3.87.51-6.04.9-6.37.88-6.75-.21-4.87-.58-4.15,1.44-.75.75-1.5,1.37,1.99.37,2.12,1.47,1.25-.88,2.87,1.51,3.34Z"/>
+        <path id="CO" class="cls-1" d="M17.21,35.61l-1.2-.66-1.38-.92-.8.44-2.38-.39-.68-1.2-.52.05-2.81-1.59-.38-.87,1.05-.21-.12-1.39.65-1.01,1.39-.19,1.19-1.75,1.07-1.46-1.04-.67.53-1.62-.63-2.56.6-.73-.44-2.37-1.14-1.5.36-1.36.91.2.53-.84-.65-1.65.34-.42,1.44.09,2.11-1.97,1.15-.3.03-.93.52-2.39,1.61-1.32,1.76-.05.22-.59,2.2.23,2.21-1.43,1.09-.64,1.35-1.37,1,.17.73.75-.54.96-1.8.48-.71,1.42-1.09.81-.81,1.06-.35,2.01-.77,1.66,1.44.18.36,1.3.62.62.22,1.13-.33,1.04.1.59.69.23.67.98,3.6-.27,1.63.36,1.98,2.41,1.13-.3,2.02.15,1.6-.32.99.49-.51,1.5-.62.94-.22,2.01.56,1.85.8.83.09.63-1.42,1.39,1.02.61.75.98.85,2.77-.53.35-.54-1.65-.78-.88-.93.96-5.46-.06.03,1.74,1.64.29-.09,1.07-.56-.29-1.58.46v2.02s1.23,1.02,1.23,1.02l.44,1.59-.07,1.21-1.26,7.65-1.4-1.49-.84-.06,1.81-2.84-2.15-1.31-1.68.24-1.01-.48-1.55.74-2.09-.35-1.65-2.92-1.3-.72-.89-1.32-1.86-1.32-.75.27Z"/>
+        <path id="EC" class="cls-1" d="M3.39,44.67l1.49-2.08-.61-1.22-1.07,1.3-1.68-1.23.57-.78-.47-2.53.98-.42.52-1.73,1.06-1.8-.2-1.13,1.54-.6,1.92-1.11,2.81,1.59.52-.05.68,1.2,2.38.39.8-.44,1.38.92,1.2.66.39,2.11-.87,1.81-3.06,2.92-3.37,1.1-1.72,2.43-.53,1.88-1.59,1.15-1.17-1.41-1.14-.3-1.16.22-.07-1.02.8-.66-.33-1.17Z"/>
+        <path id="FK" class="cls-1" d="M56.95,204.93l3.36-2.69,2.39,1.12,1.68-1.79,2.24,2.01-.84,1.58-3.79,1.36-1.26-1.59-2.38,2.05-1.4-2.05Z"/>
+        <path id="GF" class="cls-1" d="M81.18,28.21l-1.07,1.06-1.34.2-.38-.78-.63-.12-.87.76-1.22-.57.71-1.19.24-1.27.48-1.2-1.09-1.65-.22-1.91,1.46-2.41.95.31,2.06.66,2.97,2.36.46,1.14-1.66,2.55-.85,2.06Z"/>
+        <path id="GY" class="cls-1" d="M60.99,11.8l1.84,1.03,1.74,1.83.07,1.45,1.06.07,1.5,1.36,1.11.98-.45,2.52-1.7.73.15.65-.52,1.45,1.25,2.02h.89s.37,1.58.37,1.58l1.71,2.42-.68.1-1.55-.23-.91.74-1.27.49-.88.12-.31.54-1.38-.14-1.73-1.3-.2-1.29-.72-1.41.45-2.38.78-.98-.65-1.3-.96-.42.37-1.23-.66-.64-1.46.12-1.89-2.12.76-.77-.06-1.3,1.73-.45.69-.52-.96-1.04.25-1.03,2.22-1.65Z"/>
+        <path id="PE" class="cls-1" d="M33.42,84.94l-.75,1.51-1.44.74-2.81-1.68-.25-1.2-5.55-2.92-5.03-3.17-2.17-1.78-1.16-2.37.46-.83-2.37-3.75-2.77-5.24-2.64-5.62-1.15-1.29-.88-2.06-2.18-1.84-2-1.13.91-1.25-1.36-2.67.87-1.95,2.24-1.77.33,1.17-.8.66.07,1.02,1.16-.22,1.14.3,1.17,1.41,1.59-1.15.53-1.88,1.72-2.43,3.37-1.1,3.06-2.92.87-1.81-.39-2.11.75-.27,1.86,1.32.89,1.32,1.3.72,1.65,2.92,2.09.35,1.55-.74,1.01.48,1.68-.24,2.15,1.31-1.81,2.84.84.06,1.4,1.49-2.53-.13-.37.42-2.3.53-3.2,1.91-.21,1.3-.71.98.28,1.51-1.7.81v1.19l-.74.51,1.17,2.53,1.56,1.72-.59,1.21,1.86.16,1.06,1.51,2.47.07,2.3-1.66-.19,4.3,1.28.33,1.58-.49,2.43,4.58-.61.96-.13,2.02-.06,2.44-1.1,1.44.51,1.07-.65.97,1.21,2.44-1.77,3.14Z"/>
+        <path id="PY" class="cls-1" d="M52.78,98.79l1.11-3.59.07-1.6,1.34-2.62,4.89-.86,2.6.05,2.62,1.51.04.91.83,1.66-.18,4.06,2.96.58,1.14-.59,1.89.82.53.9.26,2.77.33,1.18,1.04.13,1.05-.49,1.01.55v1.68l-.38,1.82-.55,1.78-.46,2.75-2.54,2.4-2.22.5-3.15-.48-2.82-.85,2.76-4.73-.41-1.37-2.88-1.2-3.43-2.26-2.29-.46-5.16-4.95Z"/>
+        <path id="SR" class="cls-1" d="M68.31,18.52l3.36.56.3-.51,2.27-.2,3.01.76-1.46,2.4.22,1.91,1.11,1.66-.49,1.2-.25,1.27-.72,1.17-1.6-.59-1.33.29-1.13-.25-.28.81.47.55-.25.57-1.53-.23-1.71-2.42-.37-1.57h-.89s-1.25-2.03-1.25-2.03l.52-1.45-.15-.65,1.7-.73.45-2.52Z"/>
+        <path id="UY" class="cls-1 dinosaurs" d="M66.97,123.59l1.82-.34,2.81,2.5,1.04-.09,2.89,2.08,2.2,1.82,1.62,2.25-1.24,1.57.78,1.9-1.21,2.12-3.17,1.88-2.07-.68-1.52.37-2.59-1.46-1.9.11-1.71-1.87.22-2.16.61-.74-.03-3.3.75-3.37.7-2.59Z"/>
+        <path id="VE" class="cls-1" d="M28.54,2.15l-.08.67-1.65.33.92,1.29-.04,1.49-1.23,1.64,1.06,2.24,1.21-.18.63-2.04-.87-1-.14-2.14,3.49-1.16-.39-1.34.98-.9,1.01,2,1.97.05,1.82,1.58.11.94,2.51.02,3-.29,1.61,1.27,2.14.35,1.57-.88.03-.72,3.48-.17,3.36-.04-2.38.84.95,1.34,2.25.21,2.12,1.39.45,2.26,1.46-.07,1.1.67-2.22,1.65-.25,1.03.96,1.04-.69.52-1.73.45.06,1.3-.76.77,1.89,2.12.38.79-1.03,1.07-3.14,1.04-2.01.44-.81.66-2.23-.7-2.08-.36-.52.26,1.25.72-.11,1.87.39,1.76,2.37.24.16.58-2.01.8-.32,1.18-1.16.45-2.08.65-.54.86-2.18.18-1.55-1.48-.85-2.77-.75-.98-1.02-.61,1.42-1.39-.09-.63-.8-.83-.56-1.85.22-2.01.62-.94.51-1.5-.99-.49-1.6.32-2.02-.15-1.13.3-1.98-2.41-1.63-.36-3.6.27-.67-.98-.69-.23-.1-.59.33-1.04-.22-1.13-.62-.62-.36-1.3-1.44-.18.77-1.66.35-2.01.81-1.06,1.09-.81.71-1.42,1.8-.48Z"/>
+      </g>
+    </g>
+  </svg>`
+}
+export { countryCodes, allMaps }
